### 2017-06-01 #118 @timtrice
rrricanes
Scrapes and parses tropical cyclone advisories and GIS data from the National Hurricane Center. Can access real-time or archived data for storms since 1998.

```
Package: rrricanes
Title: Web scraper for real-time and archived advisory products for Atlantic and
    east Pacific hurricanes and tropical storms
Version: 0.1.0.1
Authors@R: person("Tim", "Trice", email = "tim.trice@gmail.com",
                  role = c("aut", "cre"))
Description: Get archived data of past and current hurricanes and tropical
  storms for the Atlantic and eastern Pacific oceans. Data is available for
  storms since 1998.
  data.table (>= 1.9.6),
  lubridate (>= 1.6.0),
  magrittr (>= 1.5),
  purrr (>= 0.2.2.2),
  readr (>= 1.1.0),
  rvest (>= 0.3.2),
  stringr (>= 1.2.0),
  tibble (>= 1.3.1),
  tidyr (>= 0.6.2),
  xml2 (>= 1.1.1)
  devtools,
```

https://github.com/timtrice/rrricanes

- Please indicate which category or categories from our [package fit policies](https://github.com/ropensci/onboarding/blob/master/policies.md#package-fit) this package falls under and why? (e.g., data retrieval, reproducibility. If you are unsure, we suggest you make a pre-submission inquiry.):
data extraction, data visualization, data munging, data packages

Those with an interest in studying tropical meteorology, tracking storm developments and comparing the forecast abilities of the National Hurricane Center storm-to-storm, year-to-year.

I have searched (I believe) extensively and have not found one at all. To my knowledge the only other hurricane package in CRAN or GitHub is HURDAT - and I wrote that one, too :)

R 3.3.3
Most tidyverse packages.
rnaturalearthdata and rnaturalerathhires will be required for release 0.1.1.

In release 0.1.0.1, only 2 notes about global variables. This has largely been resolved in unreleased 0.1.1.
I used a CONTRIBUTORS.md file instead of CODE_OF_CONDUCT.md.
I use `@keywords internal` for internal functions
