### 2017-06-01 ea036b1
Drop tidyverse for dplyr in vignette
