+++ inst/doc/getting-started.R
@@ -1,5 +1,5 @@
 ## ------------------------------------------------------------------------
-library(tidyverse)
+library(dplyr)
 library(rrricanes)
 
 ## ----get.storms----------------------------------------------------------
