### 2017-06-01 477a198
Remove rnaturalearthhires script
