+++ appveyor.yml
@@ -38,11 +38,6 @@ environment:
 matrix:
   fast_finish: true
 
-build_script:
-  - travis-tool.sh install_github ropenscilabs/rnaturalearthhires
-  - travis-tool.sh r_binary_install proto rgdal
-  - travis-tool.sh install_deps
-
 test_script:
   - travis-tool.sh run_tests
 
