+++ appveyor.yml
@@ -38,6 +38,9 @@ environment:
 matrix:
   fast_finish: true
 
+build_script:
+  - travis-tool.sh install_deps
+
 test_script:
   - travis-tool.sh run_tests
 
