### 2017-06-01 439eb65
Removing osx r-devel test. See Issue #64
