+++ .travis.yml
@@ -21,8 +21,6 @@ matrix:
       r: 3.3.3
     - os: osx
       r: release
-    - os: osx
-      r: devel
 
 r_packages:
   - covr
