### 2017-06-01 950aaeb
Update R, requirements
