+++ README.Rmd
@@ -25,7 +25,7 @@ badgecreatr::badgeplacer(githubaccount = "timtrice", githubrepo = "rrricanes", b
 
 # rrricanes
 
-`rrricanes` is a R library that extracts information from available archives on past and current tropical cyclones. Currently, archives date back to 1998. This is considered "raw data" in the sense that information extracted is from then real-time advisories. 
+`rrricanes` is a R library that extracts information from available archives on past and current tropical cyclones. Currently, archives date back to 1998. 
 
 Data can be obtained for cyclones in the north Atlantic (considered the Atlantic Basin) and north-eastern Pacific (the East Pacific Basin from 140&deg;W and eastward, and Central Pacific Basin from 140&deg;W to 180&deg;W). 
 
@@ -39,14 +39,7 @@ vignette("getting-started", package = "rrricanes")
 
 ### Prerequisites
 
-`rrricanes` does require an active internet connection as data is extracted from online archives. 
-
-The following R packages are also currently used for data processing:
-
-* data.table (>= 1.9.6), 
-* tidyverse (>= 1.0.0)
-
-On future revisions I will work on minimizing package dependencies if it helps increase processing speed.
+`rrricanes` only require an active internet connection as data is extracted from online archives. 
 
 ### Installing
 
@@ -58,7 +51,7 @@ devtools::install_github("timtrice/rrricanes")
 
 ## Built With
 
-* [R 3.2.3](https://www.r-project.org/) - The R Project for Statistical Computing
+* [R 3.3.3](https://www.r-project.org/) - The R Project for Statistical Computing
 
 ## Contributing
 
