### 2017-06-02 6ed63d3
Add fun status_abbr_to_str
Take a vector of status abbreviations and return a vector of strings.
