+++ NAMESPACE
@@ -29,6 +29,7 @@ export(load_storm_data)
 export(mb_to_in)
 export(na.omit)
 export(saffir)
+export(status_abbr_to_str)
 export(tracking_chart)
 export(twoep)
 export(wndprb)
