### 2017-06-02 16acd1a
Remove storm report files
