+++ reports/storm_report.Rmd
@@ -1,110 +1,0 @@
----
-author: "Tim Trice"
-date: "`r strftime(Sys.time(), '%a %b %e, %Y %R')`"
-output: pdf_document
-params:
-    key: "EP022017"
-    set_title: "`r as.character()`"
-title: "`r params$set_title`"
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE, fig.width = 10, fig.height = 7, fig.align = "center")
-```
-
-```{r libraries}
-library(dplyr)
-library(ggplot2)
-library(purrr)
-library(rrricanes)
-```
-
-```{r}
-# Key is set in params
-basin <- stringr::str_sub(key, 0L, 2L)
-year_num <- as.numeric(stringr::str_sub(key, 3L, 4L))
-year <- as.numeric(stringr::str_sub(key, 5L, 8L))
-```
-
-```{r}
-df <- get_storms(year = year, basin = basin)
-```
-
-```{r}
-df.fstadv <- df %>% 
-    slice(year_num) %>% 
-    .$Link %>% 
-    get_fstadv()
-
-adv <- as.numeric(last(df.fstadv$Adv))
-date <- last(df.fstadv$Date)
-```
-
-## Current
-
-```{r, results = "hide"}
-gis <- gis_latest(basins = basin)
-```
-
-```{r}
-sprintf("%s_tracking_chart", stringr::str_to_lower(basin)) %>% 
-    purrr::invoke(.x = list(res = 50, color = "black", size = 0.1, fill = "white")) + 
-    geom_polygon(data = gis[[1]]$ep022017_002_5day_pgn, aes(x = long, y = lat), fill = "red", alpha = 0.2) + 
-    geom_path(data = gis[[1]]$ep022017_002_5day_lin, aes(x = long, y = lat)) + 
-    geom_point(data = gis[[1]]$EP022017_pts, aes(x = LON.x, y = LAT.x)) + 
-    geom_path(data = gis[[1]]$EP022017_pts, aes(x = LON.x, y = LAT.x)) + 
-    geom_point(data = gis[[1]]$ep022017_002_5day_pts, aes(x = LON.x, y = LAT.x, color = TCDVLP.x, size = MAXWIND.x)) + 
-    geom_point(data = gis[[1]]$ep022017_002_ww_wwlin, aes(x = long, y = lat, color = TCWW)) + 
-    ggplot2::coord_equal(xlim = c(-105, -90), ylim = c(10, 20)) + 
-    ggplot2::scale_x_continuous(expand = c(0, 0)) + 
-    ggplot2::scale_y_continuous(expand = c(0, 0)) + 
-    labs(title = storm_name, 
-         subtitle = sprintf("Adv %s (%s)", 
-                            adv, 
-                            strftime(date, "%b %e, %Y %R %Z")), 
-         x = "Lon", 
-         y = "Lat", 
-         caption = "rrricanes 0.1.1")
-```
-
-## Past Advisories
-
-```{r}
-knitr::kable(df.fstadv %>% select(Status, Adv, Date, Wind, Gust, Pressure, Eye))
-```
-
-## Forecast
-
-```{r EP022017_fcst_split}
-v <- c("FcstDate", "Lat", "Lon", "Wind", "Gust")
-df.fcst <- map_df(.x = c(12, 24, 36, 48, 72, 96, 120),
-                        .f = function(y) {
-                            select_(df.fstadv,
-                                    .dots = c("Key", "Adv", "Date", paste0("Hr", y, v))) %>%
-                                rename_("Key" = "Key", "Adv" = "Adv", "Date" = "Date",
-                                        "FcstDate" = paste0("Hr", y, "FcstDate"),
[…]
