### 2017-06-02 8987e5b
Correct examples
