### 2017-06-02 ca10612
Correct examples
