### 2017-06-02 6846255
Update fun gis_advisory to return full urls
