### 2017-06-02 1aac2c4
Update gis_prob_storm_surge to return full URLs
Return full URLs with domain

Clarify purpose of param nobs in documentation

Corrected examples
