### 2017-06-02 419a6a8
Update gis_prob_storm_surge to return full URLs
Return full URLs with domain

Clarify purpose of param nobs in documentation

Corrected examples
