+++ man/gis_prob_storm_surge.Rd
@@ -4,7 +4,7 @@
 \alias{gis_prob_storm_surge}
 \title{gis_prob_storm_surge}
 \usage{
-gis_prob_storm_surge(key, products, datetime = NULL, nobs = ~length(x))
+gis_prob_storm_surge(key, products, datetime = NULL, nobs = NULL)
 }
 \arguments{
 \item{key}{Key of storm (i.e., AL012008, EP092015)}
@@ -64,7 +64,7 @@ want.
         day then all data for August will be returned. You cannot drop values
         on the left side; only from the right. This is to help ensure you get
         the dataset you are requesting.}
-    \item{nobs}{A formula or numeric vector specifying the range of datasets
+    \item{nobs}{A formula specifying the range of datasets
         to retrieve.}
 }
 See examples for more guidance on using datetime and nobs together.
@@ -76,15 +76,15 @@ See examples for more guidance on using datetime and nobs together.
 gis_prob_storm_surge("AL092016", products = list("psurge" = 0))
 
 # Return the first psurge0 and esurge10 product for storm AL092016
-gis_prob_storm_surge("AL092016", products = list("psurge" = 0, "esurge" = 10), nobs = 1)
+gis_prob_storm_surge("AL092016", products = list("psurge" = 0, "esurge" = 10), nobs = ~1)
 
-# Return all psurge0 products for September 3, 2016, storm AL092016
+# Return the first and last psurge0 products for Sep 2, 2016, storm AL092016
 gis_prob_storm_surge("AL092016", products = list("psurge" = 0),
-                     datetime = "20160903", nobs = ~c(1:length(x)))
+                     datetime = "20160902", nobs = ~c(1, length(x)))
 
 # Last five esurge40 and esurge50 products for month of September, storm AL092016
 gis_prob_storm_surge("AL092016", products = list("esurge" = c(40, 50)),
-                     datetime = "201609", nobs = ~c((length(x)-10):length(x)))
+                     datetime = "201609", nobs = ~c((length(x)-5):length(x)))
 }
 }
 \seealso{
