### 2017-06-02 a80ba4d
Update doc for fun gis_advisory
