### 2017-06-02 b862594
Update doc for fun gis_advisory
