### 2017-06-02 d7a2001
Update fun gis_outlook to return URL
Updated documentation
