### 2017-06-02 b648108
Update fun gis_outlook to return URL
Updated documentation
