### 2017-06-03 7119767
Update gis functions to extract URLs, parse dfs
Major changes to the functions

* Only gis_download returns datasets. All other gis functions extract
the dataset URLs from their respective pages. This is to help users
refine what they are requesting before actually submitting it as some
of the datasets may take quite a bit of time.

* Updated documentation of GIS datasets but still needs work.

* Modified gis_download to handle datasets based on type of dataset.
Still needs refinement to ensure all data is being put together
correctly.

Additionally, once files are parsed they need to be remov
[…]
