+++ R/gis.R
@@ -3,7 +3,6 @@
 #' @param key Key of storm (i.e., AL012008, EP092015)
 #' @param advisory Advisory number. If NULL, all advisories are returned. Intermediate
 #' advisories are acceptable.
-#' @return Full urls of requested advisories.
 #' @seealso \code{\link{gis_download}}
 #' @export
 gis_advisory <- function(key, advisory = as.character()) {
@@ -52,16 +51,28 @@ gis_advisory <- function(key, advisory = as.character()) {
 #' @param destdir Directory to save shapefile data.
 #' @export
 gis_download <- function(url, destdir = tempdir()) {
+    message("Remove files when processing complete.")
     utils::download.file(file.path(url), zip_file <- tempfile())
     utils::unzip(zip_file, exdir = destdir)
     shp_files <- list.files(path = destdir, pattern = ".+shp$")
     ds <- purrr::map2(.x = shp_files, .y = destdir, .f = function(f, d) {
         shp_file <- stringr::str_match(f, "^(.+)\\.shp$")[,2]
         shp <- rgdal::readOGR(dsn = d, layer = shp_file)
-        shp@data$id <- rownames(shp@data)
-        shp.points <- broom::tidy(shp, region = "id")
-        df <- dplyr::left_join(shp.points, shp@data, by = "id")
-        return(df)
+        # For linear plots
+        if (any(grepl(".+lin\\.shp$", f), grepl(".+pgn\\.shp$", f))) {
+            shp@data$id <- rownames(shp@data)
+            shp.points <- broom::tidy(shp, region = "id")
+            df <- dplyr::left_join(shp.points, shp@data, by = "id")
+            return(df)
+        } else if (grepl(".+pts\\.shp$", f)) {
+            df <- tibble::as_data_frame(shp)
+            return(df)
+        } else {
+            shp@data$id <- rownames(shp@data)
+            shp.points <- broom::tidy(shp, region = "id")
+            df <- dplyr::left_join(shp.points, shp@data, by = "id")
+            return(df)
+        }
     })
     shp_file_names <- stringr::str_match(shp_files, "^(.+)\\.shp$")[,2] %>%
         stringr::str_replace_all("[[:punct:][:space:]]", "_")
@@ -86,15 +97,16 @@ gis_latest <- function(basins = c("AL", "EP"), destdir = tempdir()) {
     x.ziplinks <- purrr::map(x.links, ~ stringr::str_detect(.x, pattern = "^<link>.+zip</link>$"))
     x.matches <- purrr::map2(x.links, x.ziplinks, ~ .x[.y])
     zips <- purrr::map(x.matches, as.character) %>% purrr::flatten_chr() %>% stringr::str_match("^<link>(.+)</link>$")
-    if (any(purrr::is_empty(zips), purrr::is_empty(zips[,2])))
-        return(NULL)
-    ds <- purrr::map(zips[,2], gis_download, destdir = destdir)
-    return(ds)
+    tryCatch(links <- ds <- purrr::map(zips[,2], gis_download, destdir = destdir)[,2],
+             error = function(c) {
+                 c$message <- "No data available for current storms."
+                 stop(c$message, call. = FALSE)
+             })
+    return(links)
 }
 
 #' @title gis_outlook
 #' @description Tropical Weather Outlook
-#' @return URL to latest Tropical Weather Outlook GIS dataset.
 #' @seealso \code{\link{gis_download}}
 #' @export
 gis_outlook <- function() {
@@ -154,6 +166,7 @@ gis_outlook <- function() {
 #' }
 #' See examples for more guidance on using datetime and nobs together.
[…]
