### 2017-06-03 9203189
Re-add dynamic reporting drafts
