+++ NAMESPACE
@@ -21,7 +21,6 @@ export(gis_download)
 export(gis_latest)
 export(gis_outlook)
 export(gis_prob_storm_surge)
-export(gis_storm_surge_flood)
 export(gis_windfield)
 export(gis_wsp)
 export(knots_to_mph)
