### 2017-06-05 88f8208
Simplify fun gis_download
