### 2017-06-05 524259d
Removed nobs parameter. Can use dplyr helpers
