### 2017-06-05 3c348be
Removed nobs parameter. Can use dplyr helpers
