### 2017-06-05 dd45339
Updated vignette to reflect gis_download update
