+++ vignettes/gis_data.Rmd
@@ -15,74 +15,122 @@ library(ggplot2)
 library(rrricanes)
 ```
 
+Most storms will contain a variation of GIS datasets that can be plotted with `ggplot2`. The helper functions for this have the prefix 'gis'.
+
 **All products are experimental and there maybe fluctuations particularly in current datasets.**
 
-Functions that do not take key or advisory parameters but require datetime instead, datetime will *typically* be the datetime of the advisory minus three hours. Forecast/Advisory products, for example, would be issued at 3AM but there is a note notating the location three hours prior; 12AM. These hours, 00, 06, 12, and 18 seem to reflect the datetime parameters required in some GIS products. These products will most likely contain data for multiple storms, if they exist.
+In general, datasets are available for storms dated back to 1998. However, products such as Wind Speed Probabilities only go back to 1999. 
 
-## Obtaining GIS data
-
-Hurricane Mathew Adv 32
-
-```{r}
-gis_advisory(key = "AL142016", advisory = "32")
-```
-
-```{r}
-df <- gis_advisory(key = "AL142016", advisory = "32") %>% gis_download()
-```
-
-```{r}
-names(df)
-```
+Some datasets require the use of the storm key and an optional advisory. Other products require a datetime value and cannot be isolated by storm key or advisory. The datetime values are not based on the issue time of the advisory, but rather three hours prior. For example, if you are seeking a dataset where the forecast/advisory was issued at 9:00AM UTC, you will want the dataset for 6:00AM UTC. This will be explained a little further below.
 
 ## Build a Tracking Chart
 
+There are three functions available to help you plot GIS data; `tracking_chart`, `al_tracking_chart` and `ep_tracking_chart`. `al_tracking_chart` and `ep_tracking_chart` are just helpers centered on the Atlantic and northeast Pacific ocean, respectively. 
+
+Building tracking charts requires the `rnaturalearthdata` and, for high resolution charts, `rnaturalearthhires` packages.
+
 ```{r}
-(bp <- al_tracking_chart(res = 50, color = "black", size = 0.1, fill = "white"))
+args(rrricanes::tracking_chart)
 ```
 
-Refine chart
+The countries and states parameters are TRUE by default. This means a basic call to `tracking_chart` will return a map with country and state borders. The basic call does not show this.
 
 ```{r}
-(bp <- bp + 
-    coord_equal(xlim = c(-90, -60), ylim = c(20, 40)) +
-    scale_x_continuous(expand = c(0, 0)) +
-    scale_y_continuous(expand = c(0, 0)))
+tracking_chart()
 ```
 
-## Five Day Line Forecast
-
-x = long
-y = lat
+You can pass typical `aes` parameters to refine the color and fill of the plot; remember the tracking chart is a ggplot object.
 
 ```{r}
-bp + geom_path(data = df$al142016_032_5day_lin, aes(x = long, y = lat))
+tracking_chart(color = "black", fill = "white", size = 0.1)
 ```
 
-## Five Day Points
-
-x = LON
-y = LAT
+You may choose to only show coastline data instead. In this case, just set the countries parameter to FALSE.
 
 ```{r}
-bp + geom_point(data = df$al142016_032_5day_pts, aes(x = LON, y = LAT))
[…]
