+++ man/saffir.Rd
@@ -6,7 +6,11 @@
 \usage{
 saffir(x)
 }
+\arguments{
+\item{x}{Vector of wind speed values.}
+}
 \description{
 Return category of storm based on wind. Assumes storm is a
 cyclone. Saffir-Simpson Hurricane Scale does not apply to non-tropical storms.
 }
+\keyword{internal}
