+++ NAMESPACE
@@ -27,7 +27,6 @@ export(knots_to_mph)
 export(load_storm_data)
 export(mb_to_in)
 export(na.omit)
-export(saffir)
 export(status_abbr_to_str)
 export(tracking_chart)
 export(twoep)
