### 2017-06-05 bbec07a
Correct docs for gis_download
