### 2017-06-05 460d07f
Correct docs for gis_download
