### 2017-06-05 b217d33
Update docs for gis_latest
