### 2017-06-05 c2b7921
Update docs for gis_prob_storm_surge
