### 2017-06-05 21be918
Replace headers in vignette
