+++ vignettes/gis_data.Rmd
@@ -1,13 +1,13 @@
 ---
 title: "GIS data"
 author: "Tim Trice"
-date: "June 3, 2017"
-output: html_document
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{GIS data}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
 ---
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-```
 
 ```{r}
 library(dplyr)
