+++ R/products.R
@@ -5,7 +5,7 @@
 #' where `title`, `description` and `pubDate` reside. `link` is also
 #' available to point to the NHC website.
 #' @return Returns nested list of xml nodes.
-#' @seealso \link{http://www.nhc.noaa.gov/xml/TWOEP.xml}
+#' @seealso \url{http://www.nhc.noaa.gov/xml/TWOEP.xml}
 #' @export
 twoep <- function() {
     url <- "http://www.nhc.noaa.gov/xml/TWOEP.xml"
