### 2017-06-05 e49fc2e
Temporarily disable testing
The archive page for the 1998 season is currently down preventing
some tests from passing. Disable for time being.
