+++ appveyor.yml
@@ -13,6 +13,7 @@ install:
 # Adapt as necessary starting from here
 
 environment:
+  R_CHECK_ARGS: --no-tests # Disable tests due to issues w/ NHC archives.
   global:
     WARNINGS_ARE_ERRORS: 1
 
