+++ .travis.yml
@@ -2,6 +2,10 @@ language: r
 
 warnings_are_errors: true
 
+env:
+  global:
+    - R_CHECK_ARGS="--no-tests"
+
 # Test package on multiple platforms
 matrix:
   include:
