+++ DESCRIPTION
@@ -31,7 +31,6 @@ Imports:
   rgdal (>= 1.2-7), 
   rgeos (>= 0.3-23), 
   rnaturalearthdata (>= 0.1.0),
-  rnaturalearthhires (>= 0.1.0),
   rvest (>= 0.3.2), 
   stringr (>= 1.2.0), 
   tibble (>= 1.3.1), 
