### 2017-06-05 df1ffea
Add dependencies for rgdal package
