### 2017-06-05 9683430
Try trusty for rgdal, travis issues
