+++ .travis.yml
@@ -2,6 +2,8 @@ language: r
 
 warnings_are_errors: true
 
+dist: trusty
+
 env:
   global:
     - R_CHECK_ARGS="--no-tests"
