+++ .travis.yml
@@ -1,48 +1,9 @@
 language: r
-
-warnings_are_errors: true
-
+sudo: required
 dist: trusty
-
-env:
-  global:
-    - R_CHECK_ARGS="--no-tests"
 
 addons:
   apt:
     packages:
       - libgdal-dev
       - libproj-dev
-
-# Test package on multiple platforms
-matrix:
-  include:
-    - os: linux
-      dist: precise
-      sudo: false
-      r: 3.3.3
-    - os: linux
-      dist: precise
-      sudo: false
-      r: release
-    - os: linux
-      dist: precise
-      sudo: false
-      r: devel
-    - os: osx
-      r: 3.3.3
-    - os: osx
-      r: release
-    - os: osx
-      r: devel
-
-r_packages:
-  - covr
-
-cache: packages
-
-after_success:
-  - Rscript -e 'library(covr); codecov()'
-
-r_github_packages:
-  - ropenscilabs/rnaturalearthhires
