### 2017-06-05 fa8ebdb
Disable vignettes, manual, tests in appveyor
