+++ DESCRIPTION
@@ -26,6 +26,7 @@ Imports:
   lazyeval (>= 0.2.0), 
   lubridate (>= 1.6.0),
   magrittr (>= 1.5), 
+  maptools (>= 0.9-2), 
   purrr (>= 0.2.2.2), 
   readr (>= 1.1.0), 
   rgdal (>= 1.2-7), 
