### 2017-06-05 a33a03b
Change version to 0.2.0
