+++ README.Rmd
@@ -3,7 +3,7 @@ output: github_document
 ---
 
 [![minimal R version](https://img.shields.io/badge/R%3E%3D-3.3.3-6666ff.svg)](https://cran.r-project.org/) 
-[![packageversion](https://img.shields.io/badge/Package%20version-0.1.1-orange.svg?style=flat-square)](commits/issue-47)
+[![packageversion](https://img.shields.io/badge/Package%20version-0.2.0-orange.svg?style=flat-square)](commits/issue-47)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/rrricanes)](https://cran.r-project.org/package=rrricanes) 
 [![Build Status](https://travis-ci.org/timtrice/rrricanes.svg?branch=issue-47)](https://travis-ci.org/timtrice/rrricanes) 
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/timtrice/rrricanes?branch=issue-47&svg=true)](https://ci.appveyor.com/project/timtrice/rrricanes) 
