### 2017-06-05 441ddea
Change version to 0.2.0
