+++ NEWS.md
@@ -1,4 +1,4 @@
-# rrricanes 0.1.1
+# rrricanes 0.2.0
 
 ## Major Changes
 
@@ -7,6 +7,12 @@ Changed name from `Hurricanes` to `rrricanes`.
 `get_storm_data` can now be chained to other commands and returns a list of dataframes.
 
 `load_storm_data` accesses pre-scraped datasets and returns requested products through the github repo `rrricanesdata`. This was done to make it quicker to get data. It should not be relied on to get the most immediate data for current storms. However, it should be fairly up-to-date. Original functions can be used if for some reason immediate data access is needed.
+
+Added GIS functions to retrieve GIS datasets for historical and current storms. Not all GIS datasets will be available. The functions (`gis_advisory`, `gis_prob_storm_surge`, `gis_windfield`, `gis_wsp`) return URLs to zip files. This is done to allow the user to further filter, if necessary, to obtain the product they want. The return value can be chained to `gis_download` which will then return a list of spatial dataframes depending on the product requested.
+
+`gis_latest` returns a list of URLs to GIS datasets for active storms. If there are no active storms it will return an error.
+
+`gis_outlook` returns link to the latest tropical weather outlook for either basin in GIS format.
 
 ## Minor Changes
 
