### 2017-06-05 357b151
Resolve rebase merge conflicts
