### 2017-06-05 timtrice/rrricanes#66 @timtrice
1998, 2016, 2017 Archives Unavailable
## Users

The archives hosted on the National Hurricane Center are currently down for years 1998, 2016 and 2017. There is no report of what the issue is nor when it will be corrected other than "I hope it will be within the next day or two".

Version 0.2.0 currently exists in the `develop` branch which gives access to function `load_storm_data`. This function gives access to pre-scraped data in the `rrricanesdata` repository. All storms with the exception of EP022017 can be found in this repository.

There are no affects on GIS data that I am aware of.

## Developers

`devtools::document()` fails as R/helper-datasets.R cannot load because of the failure. This file accesses the annual data for subsequent tests. It also happens to (apparently) load when trying to update package documentation. Current workaround is to tempoarily delete contents of R/helper-datasets.R, run `devtools::document()`, then add content back (or checkout the current file from GH).

Tests will fail, thus currently enabling the --no-tests option for Travis and Appveyor.
