### 2017-06-06 2c4da08
Add waffle badge
