### 2017-06-06 f5b72e2
Change format of NEWS
