### 2017-06-06 5efc886
Added tidy functions to replace fun fstadv_split
* `tidy_fstadv` will return current details from a fstadv dataframe
object.

* `tidy_wr` will return tidied current wind radii data from a fstadv
dataframe object.

* `tidy_fcst` will return tidied forecast points from a fstadv
dataframe object.

* `tidy_fcst_wr` will return tidied forecast wind radii data points
from a fstadv dataframe object.
