+++ NAMESPACE
@@ -28,6 +28,10 @@ export(load_storm_data)
 export(mb_to_in)
 export(na.omit)
 export(status_abbr_to_str)
+export(tidy_fcst)
+export(tidy_fcst_wr)
+export(tidy_fstadv)
+export(tidy_wr)
 export(tracking_chart)
 export(twoep)
 export(wndprb)
