+++ man/tidy_fstadv.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/fstadv.R
+\name{tidy_fstadv}
+\alias{tidy_fstadv}
+\title{tidy_fstadv}
+\usage{
+tidy_fstadv(df)
+}
+\description{
+Tidy current details of a fstadv dataframe object.
+}
+\details{
+Use Key, Adv and Date to join with other tidy dataframes.
+}
