### 2017-06-06 a07441c
Add status, name vars to message
