### 2017-06-06 6cd0c9d
Updated vignette getting_started
