+++ vignettes/getting_started.Rmd
@@ -1,0 +1,238 @@
+---
+title: "Getting Started"
+author: "Tim Trice"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Getting Started}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r}
+library(dplyr)
+library(rrricanes)
+```
+
+## General Introduction
+
+`rrricanes` is intended to give easy access to hurricane archives. It is a web-scraping tool that parses the National Hurricane Center's (NHC) archives to get storm data. Data is available for storms dating back to 1998.
+
+There are two basins which data is available: north Atlantic ("AL") and northeastern Pacific ("EP"). The northeastern Pacific basin typically covers from the west coast of North America to -140&deg; longitude (140&deg;W). 
+
+## Get Storms
+
+By default, `get_storms` will return all storms that have developed for the current year in both basins. If no storms have developed, an error will be generated. For this example, we'll use 2016.
+
+```{r get.storms}
+al.2012 <- get_storms(year = 2012, basin = "AL")
+```
+
+## Getting Storm Data
+
+`get_storm_data` can be used to retrieve one or multiple products for one or more cyclones. A list of dataframes is returned.
+
+```{r}
+al182012_fstadv <- al.2012 %>% 
+    filter(Name == "Hurricane Sandy") %>% 
+    .$Link %>% 
+    get_storm_data(products = "fstadv")
+```
+
+We can get the forecast/advisory data and wind speed probabilities at once:
+
+```{r}
+al182012 <- al.2012 %>% 
+    filter(Name == "Hurricane Sandy") %>% 
+    .$Link %>% 
+    get_storm_data(c("fstadv", "wndprb"))
+```
+
+`al182012` now contains two dataframes for Hurricane Sandy; `fstadv` and `wndprb`.
+
+```{r}
+dim(al182012$fstadv)
+```
+
+```{r}
+dim(al182012$wndprb)
+```
+
+## Forecast/Advisory Product (`fstadv`)
+
+The core of a storm's dataset is located in the Forecast/Advisory product, `fstadv`. This product contains current location, forecast position, movement and structural details of the cyclone. 
+
+To access only this product, we can use `get_fstadv`:
+
+```{r}
+al182012_fstadv <- al.2012 %>% 
+    filter(Name == "Hurricane Sandy") %>% 
+    .$Link %>% 
+    get_fstadv()
+```
+
+As you may have noticed above, the dataframe is very wide at 149 variables. There are four groups of variables in this dataset: current details, current wind radii, forecast positions, and forecast wind radii.
+
+### Current Details
+
+Let's look at an example of the current details.
+
+```{r}
+str(al182012_fstadv %>% select(Status:Eye, SeasNE:SeasNW))
+```
+
+The most important variable in this dataset is `Key`. `Key` is a unique identifier for each storm that develops in either basin. It is formatted such as "AABBCCCC" where "AA" is the basin abbreviation (AL or EP), "BB" is the year number of the storm left-padded, and "CC" is the year of the storm.
+
+`Adv` is the second-most important variable here. You'll notice it is in character format. For regularly-scheduled advisories, advisory numbers are always numeric. However, when watches and warnings are in effect, intermediate advisories are issued which are given alpha suffixes; i.e., 1, 2, 3, 3A, 4, 4A, 4B, 5, etc.
+
+Only the Public Advisory (`public`) will be issued more frequently. All other regular products (`discus`, `fstadv`, `prblty`, `wndprb`) are generally issued every six hours.
+
[…]
