+++ README.Rmd
@@ -29,19 +29,18 @@ Data can be obtained for cyclones in the north Atlantic (considered the Atlantic
 Please view the vignette 'Getting Started':
 
 ```r
-vignette("getting-started", package = "rrricanes")
+vignette("getting_started", package = "rrricanes")
+```
+
+To obtain and view GIS data, see vignette:
+
+```r
+vignette("gis_data", package = "rrricanes")
 ```
 
 ### Prerequisites
 
-`rrricanes` does require an active internet connection as data is extracted from online archives. 
-
-The following R packages are also currently used for data processing:
-
-* data.table (>= 1.9.6), 
-* tidyverse (>= 1.0.0)
-
-On future revisions I will work on minimizing package dependencies if it helps increase processing speed.
+`rrricanes` only requires an active internet connection as data is extracted from online archives. 
 
 ### Installing
 
@@ -53,7 +52,7 @@ devtools::install_github("timtrice/rrricanes")
 
 ## Built With
 
-* [R 3.2.3](https://www.r-project.org/) - The R Project for Statistical Computing
+* [R 3.3.3](https://www.r-project.org/) - The R Project for Statistical Computing
 
 ## Contributing
 
