### 2017-06-06 f8d38c7
Update vignette load_storm_data fun
