### 2017-06-06 c17da6e
Update expected error in get_storm_data
