+++ .github/ISSUE_TEMPLATE.md
@@ -1,0 +1,7 @@
+### Error Message
+
+### Reproducible Example
+
+### Traceback
+
+### Session Info
