### 2017-06-06 a6ca2f8
Repo templates
