### 2017-06-06 2ee3e34
Update link to CONTRIBUTING
