### 2017-06-06 e80bfc2
Update link to CONTRIBUTING
