### 2017-06-06 30db88d
Drop first call to html_nodes
