### 2017-06-06 4201b77
Move helper calls back to respective tests
