+++ tests/testthat/test_discus.R
@@ -1,5 +1,8 @@
 context("Storm Discussions (discus)")
 
+## ---- 2008, AL ---------------------------------------------------------------
+al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
+
 ## ---- Base Data --------------------------------------------------------------
 ## ---- * 2008, AL -------------------------------------------------------------
 df.al092008.discus <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_discus()
