+++ tests/testthat/test_prblty.R
@@ -1,5 +1,10 @@
 context("Strike Probabilities (prblty)")
 
+## ---- 1998, AL ---------------------------------------------------------------
+al1998 <- get_storms(year = 1998, basin = "AL") %>% dplyr::select(Link)
+## ---- 2008, AL ---------------------------------------------------------------
+al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
+
 ## ---- Base Data --------------------------------------------------------------
 ## ---- * 1998, AL -------------------------------------------------------------
 df.al011998.prblty <- al1998 %>% dplyr::slice(1) %>% .$Link %>% get_prblty()
