### 2017-06-06 9b463bf
Move helper calls back to respective tests
