+++ tests/testthat/test_wndprb.R
@@ -1,5 +1,8 @@
 context("Wind Speed Probabilities (wndprb)")
 
+## ---- 2008, AL ---------------------------------------------------------------
+al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
+
 ## ---- Base Data --------------------------------------------------------------
 ## ---- * 2008, AL -------------------------------------------------------------
 df.al092008.wndprb <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_wndprb()
