### 2017-06-07 b19e4d8
Add #67 fix to NEWS, CHANGELOG
