### 2017-06-07 bfbc608
Add #67 fix to NEWS, CHANGELOG
