### 2017-06-07 6ef0342
Scrape all contents if xpath pre is empty
