### 2017-06-07 f343ba5
Document fixes in this branch
