### 2017-06-07 a4524cc
Document fixes in this branch
