### 2017-06-07 5b9202b
Extract zip files with str_extract
