### 2017-06-07 2967480
Specific stats namespace for complete.cases()
