### 2017-06-07 1fa2da3
Add test for gis_wsp
