### 2017-06-07 7159f1b
Specify namespace for all stats::na.omit calls
