+++ NAMESPACE
@@ -26,7 +26,6 @@ export(gis_wsp)
 export(knots_to_mph)
 export(load_storm_data)
 export(mb_to_in)
-export(na.omit)
 export(status_abbr_to_str)
 export(tidy_fcst)
 export(tidy_fcst_wr)
@@ -37,4 +36,3 @@ export(twoep)
 export(wndprb)
 importFrom(data.table,rbindlist)
 importFrom(magrittr,"%>%")
-importFrom(stats,na.omit)
