### 2017-06-07 4f1d865
Specify namespace for all stats::na.omit calls
