+++ man/reexports.Rd
@@ -1,16 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/base.R
-\docType{import}
-\name{reexports}
-\alias{reexports}
-\alias{na.omit}
-\title{Objects exported from other packages}
-\keyword{internal}
-\description{
-These objects are imported from other packages. Follow the links
-below to see their documentation.
-
-\describe{
-  \item{stats}{\code{\link[stats]{na.omit}}}
-}}
-
