### 2017-06-07 4141ffa
Remove data.table dependency
