+++ NAMESPACE
@@ -34,5 +34,4 @@ export(tidy_wr)
 export(tracking_chart)
 export(twoep)
 export(wndprb)
-importFrom(data.table,rbindlist)
 importFrom(magrittr,"%>%")
