### 2017-06-07 809e589
Add stats namespace to complete.cases()
