### 2017-06-08 330c0eb
Enable tests, build vignettes in Travis, Appveyor
