+++ .travis.yml
@@ -8,7 +8,7 @@ addons:
       - libgdal-dev
       - libproj-dev
 
-r_check_args: --no-tests --no-build-vignettes --no-manual
+# r_check_args: --no-tests --no-build-vignettes --no-manual
 
 addons:
   apt:
