+++ appveyor.yml
@@ -13,7 +13,7 @@ install:
 # Adapt as necessary starting from here
 
 environment:
-  R_CHECK_ARGS: --no-build-vignettes --no-manual --no-tests # Disable tests due to issues w/ NHC archives.
+#  R_CHECK_ARGS: --no-build-vignettes --no-manual --no-tests # Disable tests due to issues w/ NHC archives.
   global:
     WARNINGS_ARE_ERRORS: 1
 
