### 2017-06-08 2f816a6
Remove expected error message
