### 2017-06-08 5d7d5b9
Disable rnaturalearthhires pkg install
