+++ .travis.yml
@@ -46,5 +46,5 @@ cache: packages
 after_success:
   - Rscript -e 'library(covr); codecov()'
 
-r_github_packages:
-  - ropenscilabs/rnaturalearthhires
+#r_github_packages:
+#  - ropenscilabs/rnaturalearthhires
