### 2017-06-08 #118 @timtrice
@maelle , thank you for taking the time to review my project!

Regarding the code in question, it has been replaced with the following:

```r
    ds <- purrr::map(products, .f = function(x) {
        sprintf("get_%s", x) %>%
            purrr::invoke_map(.x = list(link = link)) %>%
            purrr::flatten_df()})
    names(ds) <- products
    return(ds)
```

This is in branch `develop-0.2.0`. I believe I have modified all of the assign calls; they do not show up in R CMD checks (0 notes).

Regarding the tests; I apologize for them not being enabled. At one point earlier this week I had to [disable them](https://github.com/timtrice/rrricanes/issues/66) due to the NHC archives being down for some annual archive pages. This seems to have been resolved now. Tests have been added to both Travis and Appveyor.

A side note to that last point; I have noticed there are timeouts when accessing the NHC archives. This has been verified on numerous occasions with Travis and Appveyor during tests. I have a feature request pending to add handling for this. Additionally, in 0.2.0, `load_storm_data` has been added that accesses some already-processed datasets through the [GitHub repo](https://github.com/timtrice/rrricanesdata) `rrricanesdata`. So I'm trying to develop some workaround to ensure users have multiple ways to get quick access to data they need.
