### 2017-06-08 #118 @maelle
Now that I could check the package (on the right R version) here is something more formalized.

### Editor checks:

- [ x] **Fit**: The package meets criteria for [fit](policies.md#fit) and [overlap](policies.md#fit)
- [ x] **Automated tests:** Package has a testing suite and is tested via Travis-CI or another CI service.
- [ x] **License:** The package has a CRAN or OSI accepted license
- [ x] **Repository:** The repository link resolves correctly

---

#### Editor comments

* Since some problems mentioned earlier are solved in the development branch, I'd suggest we wait until you merge it with master before looking for reviewers. Does that sound ok @timtrice ?

* Does this NOTE too disappear in the development branch?

```r
* checking R code for possible problems ... NOTE
fstadv_forecast_wind_radius: no visible binding for global variable '.'
fstadv_forecasts: no visible binding for global variable 'd'
fstadv_forecasts: no visible binding for global variable 'mo'
fstadv_forecasts: no visible binding for global variable 'y'
fstadv_forecasts: no visible binding for global variable 'h'
fstadv_forecasts: no visible binding for global variable 'mi'
fstadv_forecasts: no visible binding for global variable 'lath'
fstadv_forecasts: no visible binding for global variable 'lat'
fstadv_forecasts: no visible binding for global variable 'lonh'
fstadv_forecasts: no visible binding for global variable 'lon'
fstadv_forecasts: no visible binding for global variable 'w'
fstadv_forecasts: no visible binding for global variable 'g'
fstadv_forecasts: no visible binding for global variable 'FcstDate'
fstadv_forecasts: no visible binding for global variable 'Lat'
fstadv_forecasts: no visible binding for global variable 'Lon'
fstadv_forecasts: no visible binding for global variable 'Wind'
fstadv_forecasts: no visible binding for global variable 'Gust'
fstadv_forecasts: no visible binding for global variable 'wf'
fstadv_forecasts: no visible binding for global variable '.'
fstadv_seas: no visible binding for global variable 'V2'
fstadv_seas: no visible binding for global variable 'V3'
fstadv_seas: no visible binding for global variable 'V4'
fstadv_seas: no visible binding for global variable 'V5'
fstadv_seas: no visible binding for global variable 'SeasNE'
fstadv_seas: no visible binding for global variable 'SeasSE'
fstadv_seas: no visible binding for global variable 'SeasSW'
fstadv_seas: no visible binding for global variable 'SeasNW'
fstadv_split: no visible global function definition for
  'complete.cases'
fstadv_wind_radius: no visible binding for global variable
  'WindField34'
fstadv_wind_radius: no visible binding for global variable
  'WindField50'
fstadv_wind_radius: no visible binding for global variable
  'WindField64'
prblty: no visible binding for global variable 'A'
prblty: no visible binding for global variable 'B'
prblty: no visible binding for global variable 'C'
prblty: no visible binding for global variable 'D'
prblty: no visible binding for global variable 'E'
  . A B C D E FcstDate Gust Lat Lon SeasNE SeasNW SeasSE SeasSW V2 V3
  V4 V5 Wind WindField34 WindField50 WindField64 complete.cases d g h
  lat lath lon lonh mi mo w wf y
Consider adding
  importFrom("stats", "C", "D", "complete.cases")
to your NAMESPACE file.
```

* It does take a while (580s) to run all tests. 😉 I suppose you skip them on CRAN, at least the longest ones?

* I'll update this with the `goodpractice::gp()` results but it also takes a while to run and I'm afraid I'll soon be away from keyboard.

---

Reviewers:
Due date:
