### 2017-06-08 #118 @timtrice
@maelle, thank you for the compliment! Very much appreciated.

The `master` will be updated with `develop-0.2.0` when I finish the GIS products. They probably could be released now but I'm trying to think over things and make sure there's not a better way to handle them (consolidating functions, for example). Plus one of the products is raster format so I'm looking over ways to view them with ggplot, if even possible, what the alternatives are, etc.

In hindsight I should have incorporated bug fixes into master. It was one of those things that I had already made new changes and started just hammering out bugs - before I knew it I lost track of where I was. This is the first project I'm trying to incorporate R package rules and github best practices along with what I know needs to be done. Some of it (as your example above) is just a sloppy "get it working" attempt before I get back to making the code cleaner.

Regarding `rrricanesdata`, I can certainly add a license and the script to add new datasets and update existing ones as need be. I'll add a README as well today.
