### 2017-06-08 #118 @maelle
Ok so I'll re-do the editor checks once you've updated this thread! :-)
