### 2017-06-08 #118 @maelle
@timtrice I added the results of `goodpractice::gp()` in the comment with other editor checks. Please have a look before the next version, `goodpractice` gives some good suggestions (and some you might prefer to ignore which is fine).
