### 2017-06-08 ffd9134
Remove gis functions
