+++ R/gis.R
@@ -1,386 +1,0 @@
-#' @title gis_advisory
-#' @description Advisory Forecast Track, Cone of Uncertainty, and Watches/Warnings
-#' @param key Key of storm (i.e., AL012008, EP092015)
-#' @param advisory Advisory number. If NULL, all advisories are returned. Intermediate
-#' advisories are acceptable.
-#' @seealso \code{\link{gis_download}}
-#' @export
-gis_advisory <- function(key, advisory = as.character()) {
-
-    if (is.null(key))
-        stop("Please provide storm key")
-
-    key <- stringr::str_to_lower(key)
-
-    if (!grepl("^[[:lower:]]{2}[[:digit:]]{6}$", key))
-        stop("Invalid key")
-
-    key <- stringr::str_match(key, pattern = "([:lower:]{2})([:digit:]{2})([:digit:]{4})")
-    names(key) <- c("original", "basin", "year_num", "year")
-
-    # Get list of GIS forecast zips for storm and download
-    url <- sprintf("http://www.nhc.noaa.gov/gis/archive_forecast_results.php?id=%s%s&year=%s",
-                   key[["basin"]], key[["year_num"]], key[["year"]])
-    contents <- readr::read_lines(url)
-
-    # Match zip files. If advisory is empty then need to pull all zip files for
-    # the storm. Otherwise, pull only selected advisory.
-    if (purrr::is_empty(advisory)) {
-        ptn <- sprintf(".+(forecast/archive/%s.*?\\.zip).+", stringr::str_to_lower(key[["original"]]))
-    } else {
-        advisory <- stringr::str_match(advisory, "([:digit:]{1,3})([:alpha:]*)")
-        names(advisory) <- c("original", "advisory", "int_adv")
-        ptn <- sprintf(".+(forecast/archive/%s.*?%s%s\\.zip).+", stringr::str_to_lower(key["original"]), stringr::str_pad(string = advisory[["advisory"]], width = 3, side = "left", pad = "0"), advisory[["int_adv"]])
-    }
-
-    matches <- contents[stringr::str_detect(contents, pattern = ptn)]
-    # Extract link to zip files. Error gracefully if no matches.
-    tryCatch(links <- stringr::str_match(matches, pattern = ptn)[,2],
-             error = function(c) {
-                 c$message <- "No data avaialable for requested storm/advisory"
-                 stop(c$message, call. = FALSE)
-             })
-    # Append website domain to links
-    links <- paste0("http://www.nhc.noaa.gov/gis/", links)
-    return(links)
-}
-
-#' @title gis_download
-#' @description Get GIS data for storm.
-#' @param url link to GIS dataset to download.
-#' @param destdir Directory to save shapefile data.
-#' @export
-gis_download <- function(url, destdir = tempdir()) {
-    utils::download.file(file.path(url), zip_file <- tempfile())
-    utils::unzip(zip_file, exdir = destdir)
-    shp_files <- list.files(path = destdir, pattern = ".+shp$")
-    ds <- purrr::map2(.x = shp_files, .y = destdir, .f = function(f, d) {
-        shp_file <- stringr::str_match(f, "^(.+)\\.shp$")[,2]
-        sp_object <- rgdal::readOGR(dsn = d, layer = shp_file,
-                                    encoding = "UTF-8",
-                                    stringsAsFactors = FALSE,
-                                    use_iconv = TRUE)
-        return(sp_object)
-    })
-    shp_file_names <- stringr::str_match(shp_files, "^(.+)\\.shp$")[,2] %>%
-        stringr::str_replace_all("[[:punct:][:space:]]", "_")
-    names(ds) <- shp_file_names
-    return(ds)
-}
[…]
