+++ man/gis_wsp.Rd
@@ -1,35 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gis.R
-\name{gis_wsp}
-\alias{gis_wsp}
-\title{gis_wsp}
-\usage{
-gis_wsp(datetime, res = c(5, 0.5, 0.1))
-}
-\arguments{
-\item{datetime}{Datetime in \%Y\%m\%d\%H format. \%m, \%d and \%H are
-optional but will return more datasets.}
-
-\item{res}{Resolution as a numeric vector; 5, 0.5, 0.1.}
-}
-\description{
-Wind Speed Probabilities
-}
-\details{
-Probability winds affecting an area within a forecast period.
-Datasets contain windfields for 34kt, 50kt and 64kt. Resolution is at 5km,
-0.5 degrees and 0.1 degrees. Not all resolutions may be available for all
-storms. Not all windfields will be available for all advisories.
-}
-\examples{
-\dontrun{
-# Return datasets for January 1, 2016 with resolution of 0.5 degrees
-gis_wndprb("20160101", res = 0.5)
-
-# Return wsp of 0.1 and 0.5 degree resolution, July, 2015
-gis_wndprb("201507", res = c(0.5, 0.1))
-}
-}
-\seealso{
-\code{\link{gis_download}}
-}
