+++ man/gis_outlook.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gis.R
-\name{gis_outlook}
-\alias{gis_outlook}
-\title{gis_outlook}
-\usage{
-gis_outlook()
-}
-\description{
-Tropical Weather Outlook
-}
-\seealso{
-\code{\link{gis_download}}
-}
