+++ man/gis_prob_storm_surge.Rd
@@ -1,65 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gis.R
-\name{gis_prob_storm_surge}
-\alias{gis_prob_storm_surge}
-\title{gis_prob_storm_surge}
-\usage{
-gis_prob_storm_surge(key, products, datetime = NULL)
-}
-\arguments{
-\item{key}{Key of storm (i.e., AL012008, EP092015)}
-
-\item{products}{list of products and associated n values; psurge (0:20) or
-esurge (10, 20, 30, 40, 50).}
-
-\item{datetime}{Datetime in \%Y\%m\%d\%H format.}
-}
-\description{
-Probabilistic Storm Surge
-}
-\details{
-Probabilistic Storm Surge Forecasts
-}
-\section{Products}{
-
-\describe{
-    \item{esurge}{The Tropical Cyclone Storm Surge Exceedences (P-Surge 2.5)
-          data shows the probability, in percent, of a specified storm surge,
-          including tides, exceeding the specified height, in feet, during
-          the forecast period indicated. The 10 percent exceedence height,
-          for example, is the storm surge height, including tides, above
-          ground level (AGL) such that there is a 10 percent chance of
-          exceeding it. The product is based upon an ensemble of Sea, Lake,
-          and Overland Surge from Hurricanes (SLOSH) model runs using the
-          National Hurricane Center (NHC) official advisory and accounts for
-          track, size, and intensity errors based on historical errors and
-          astronomical tide. Valid values are 10, 20, 30, 40 or 50.}
-    \item{psurge}{The Tropical Cyclone Storm Surge Probabilities (P-Surge
-          2.5) data shows the probability, in percent, of a specified storm
-          surge occurring during the forecast period indicated. The product
-          is based upon an ensemble of Sea, Lake, and Overland Surge from
-          Hurricanes (SLOSH) model runs using the National Hurricane Center
-          (NHC) official advisory and accounts for track, size, and intensity
-          errors based on historical errors and astronomical tide. Valid
-          values are 0:20.}
-}
-}
-
-\examples{
-\dontrun{
-# Return the last psurge0 product for storm AL092016
-gis_prob_storm_surge("AL092016", products = list("psurge" = 0))
-
-# Return the psurge0 and esurge10 products for storm AL092016
-gis_prob_storm_surge("AL092016", products = list("psurge" = 0, "esurge" = 10))
-
-# Return all psurge0 products for Sep 2, 2016, storm AL092016
-gis_prob_storm_surge("AL092016", products = list("psurge" = 0),
-                     datetime = "20160902")
-}
-}
-\seealso{
-\href{http://www.nhc.noaa.gov/surge/psurge.php}{Tropical Cyclone Storm Surge Probabilities}
-
-\code{\link{gis_download}}
-}
