### 2017-06-08 2aa56c8
Update CHANGELOG, NEWS for 0.1.1
