+++ NEWS.md
@@ -1,4 +1,4 @@
-rrricanes 0.2.0 (yyyy-mm-dd)
+rrricanes 0.1.1 (yyyy-mm-dd)
 ============================
 
 ### NEW FEATURES
@@ -9,23 +9,16 @@ rrricanes 0.2.0 (yyyy-mm-dd)
 
 * `load_storm_data` accesses pre-scraped datasets and returns requested products through the github repo `rrricanesdata`. This was done to make it quicker to get data. It should not be relied on to get the most immediate data for current storms. However, it should be fairly up-to-date. Original functions can be used if for some reason immediate data access is needed.
 
-* Added GIS functions to retrieve GIS datasets for historical and current storms. Not all GIS datasets will be available. The functions (`gis_advisory`, `gis_prob_storm_surge`, `gis_windfield`, `gis_wsp`) return URLs to zip files. This is done to allow the user to further filter, if necessary, to obtain the product they want. The return value can be chained to `gis_download` which will then return a list of spatial dataframes depending on the product requested.
-
-* `gis_latest` returns a list of URLs to GIS datasets for active storms. If there are no active storms it will return an error.
-
-* `gis_outlook` returns link to the latest tropical weather outlook for either basin in GIS format.
-
 ### MINOR IMPROVEMENTS
 
 * Modified numerous regex patterns to ensure data quality.
+* `tidy_fstadv`, `tidy_wr`, `tidy_fcst` and `tidy_fcst_wr` have been added to replaced now-removed `fstadv_split()`.
 
 ### BUG FIXES
 
 * Too many to recall. Apologies. 
-* `tidy_fstadv`, `tidy_wr`, `tidy_fcst` and `tidy_fcst_wr` have been added to replaced now-removed `fstadv_split()`.
 * Call to `get_storms` on some linux distros generated xpath_element error. Corrected. (#67)
 * Modify call to `get_storm_data`. Replaced css parameter in `rvest::html_nodes` calls with xpath parameter. Some products (notably, `get_prblty`) do not have a "pre" tag but are text documents (not HTML). Modified `scrape_contents` to return full contents if "pre" tag doesn't exist. Tested `get_discus` and `get_public`; no errors generated. (#68)
-* Modify `gis_wsp` call to get anchor elements by xpath; use `stringr::str_extract` and `stats::complete.cases` to filter down to zips. (#70)
 
 ### DEPRECATED AND DEFUNCT
 
