### 2017-06-08 41e724c
Remove tests for osx r-devel
