### 2017-06-08 c8d7c91
remove test for r-devel mingw_64
