+++ appveyor.yml
@@ -32,10 +32,6 @@ environment:
   - R_VERSION: devel
     GCC_PATH: mingw_32
 
-  - R_VERSION: devel
-    R_ARCH: x64
-    GCC_PATH: mingw_64
-
 matrix:
   fast_finish: true
 
