### 2017-06-08 0aa8c98
Clean code
