+++ R/fstadv.R
@@ -9,8 +9,7 @@
 get_fstadv <- function(link) {
 
     # Check status of link(s)
-    valid.link <- sapply(link, status)
-    valid.link <- stats::na.omit(valid.link)
+    valid.link <- purrr::map_chr(link, status) %>% stats::na.omit()
 
     # Get all products for the current storm
     products <- purrr::map(valid.link, get_products) %>% purrr::flatten_chr()
@@ -232,11 +231,10 @@ fstadv_forecasts <- function(content, date) {
 
     # Build FcstDate variable
     df <- df %>%
-        dplyr::mutate(FcstDate = lubridate::ymd_hm(paste(paste(Year, Month, Date,
-                                                               sep = '-'),
-                                                         paste(Hour, Minute,
-                                                               sep = ':'),
-                                                         sep = ' ')))
+        dplyr::mutate(FcstDate = lubridate::ymd_hm(
+            paste(paste(Year, Month, Date, sep = '-'),
+                  paste(Hour, Minute, sep = ':'),
+                  sep = ' ')))
 
     # Clean up Latitude, Longitude
     df <- df %>%
@@ -251,11 +249,11 @@ fstadv_forecasts <- function(content, date) {
 
     fcst_periods <- paste0("Hr", c(12, 24, 36, 48, 72, 96, 120))
     # Modify fcst_periods to lengt of df
-    fcst_periods <- fcst_periods[1:nrow(df)]
+    fcst_periods <- fcst_periods[seq_len(nrow(df))]
 
     df <- df %>% split(.$FcstDate)
 
-    df <- purrr::map2(1:length(df),
+    df <- purrr::map2(seq_along(df),
                       fcst_periods,
                       function(a, b) {
                           stats::setNames(df[[a]], paste0(b, names(df[[a]])))
@@ -344,7 +342,8 @@ fstadv_pos_accuracy <- function(contents) {
 #' @return numeric
 #' @keywords internal
 fstadv_pressure <- function(contents) {
-    ptn <- paste0('MINIMUM CENTRAL PRESSURE[:blank:]+([:digit:]{3,4})[:blank:]*MB')
+    ptn <- paste0("MINIMUM CENTRAL PRESSURE[:blank:]+",
+                  "([:digit:]{3,4})[:blank:]*MB")
     pressure <- stringr::str_match(contents, ptn)[,2]
     return(as.numeric(pressure))
 }
@@ -407,8 +406,8 @@ fstadv_lon <- function(contents) {
 }
 
 #' @title fstadv_seas
-#' @description There is only one line of sea data, 12FT seas in each quadrant. So this
-#'   should go easier than the wind fields
+#' @description There is only one line of sea data, 12FT seas in each quadrant.
+#' So this should go easier than the wind fields
 #' @param content text of product
 #' @param wind Wind value of current forecast/advisory product.
 #' @return boolean
@@ -435,11 +434,12 @@ fstadv_seas <- function(content, wind) {
 }
 
 #' @title fstadv_wind_radius
-#' @description Parse wind radius data from product, if exists. This is somewhat tricky as the
-#'   wind fields are 64KT, 50KT and 34KT and are listed in descending order. So,
-#'   the first line will not always be 64KT, 50KT or even 34KT depending on
-#'   strength of storm. What I do here is just extract the entire blob and work
-#'   through it. I'll continue to look for ways to improve it.
+#' @description Parse wind radius data from product, if exists. This is somewhat
+#' tricky as the wind fields are 64KT, 50KT and 34KT and are listed in
+#' descending order. So the first line will not always be 64KT, 50KT or even
+#' 34KT depending on strength of storm. What I do here is just extract the
+#' entire blob and work through it. I'll continue to look for ways to improve
+#' it.
 #'
[…]
