+++ DESCRIPTION
@@ -25,8 +25,6 @@ Imports:
   maptools (>= 0.9-2), 
   purrr (>= 0.2.2.2), 
   readr (>= 1.1.0), 
-  rgdal (>= 1.2-7), 
-  rgeos (>= 0.3-23), 
   rnaturalearthdata (>= 0.1.0),
   rvest (>= 0.3.2), 
   stringr (>= 1.2.0), 
