### 2017-06-09 83f9eba
Add tests, update documentation, additional funs
* Add funs `saffir`, `twoal`, `twoep`

* Finalize package top-level documentation

* Add test cases for above functions and modified existing test cases
for expected results.
