+++ tests/testthat/test_base.R
@@ -34,6 +34,7 @@ test_that("Convert Latitude, Longitude", {
     expect_identical(convert_lat_lon(93.1, "S"), as.numeric(-93.1))
     expect_identical(convert_lat_lon(179.0, "E"), as.numeric(179.0))
     expect_identical(convert_lat_lon(179, "W"), as.numeric(-179))
+    expect_error(convert_lat_lon(179, "X"), "y must be")
 })
 
 ## ---- * Errors ---------------------------------------------------------------
@@ -53,10 +54,29 @@ test_that("Millibars to Inches", {
     expect_equal(mb_to_in(888), 26.222624967048)
 })
 
+## ---- saffir() ---------------------------------------------------------------
+test_that("test saffir()", {
+    expect_identical(saffir(c(32, 45, 70, 90, 110, 125, 140)),
+                     c("TD", "TS", "HU1", "HU2", "HU3", "HU4", "HU5"))
+})
+
 ## ---- status() ---------------------------------------------------------------
 y <- lubridate::year(Sys.Date()) + 1
 test_that("URL Status", {
     expect_warning(
         status(u = sprintf("http://www.nhc.noaa.gov/archive/%d/", y)),
         sprintf("URL unavailable. http://www.nhc.noaa.gov/archive/%d/", y))
+})
+
+## ---- status_abbr_to_str() ---------------------------------------------------
+test_that("test status_abbr_to_str()", {
+    expect_identical(status_abbr_to_str("TD"), "Tropical Depression")
+    expect_identical(status_abbr_to_str("TS"), "Tropical Storm")
+    expect_identical(status_abbr_to_str("HU"), "Hurricane")
+    expect_identical(status_abbr_to_str("EX"), "Extratropical Cyclone")
+    expect_identical(status_abbr_to_str("SD"), "Subtropical Depression")
+    expect_identical(status_abbr_to_str("SS"), "Subtropical Storm")
+    expect_identical(status_abbr_to_str("LO"), "Low")
+    expect_identical(status_abbr_to_str("WV"), "Tropical Wave")
+    expect_identical(status_abbr_to_str("DB"), "Disturbance")
 })
