+++ NAMESPACE
@@ -19,12 +19,14 @@ export(get_wndprb)
 export(knots_to_mph)
 export(load_storm_data)
 export(mb_to_in)
+export(saffir)
 export(status_abbr_to_str)
 export(tidy_fcst)
 export(tidy_fcst_wr)
 export(tidy_fstadv)
 export(tidy_wr)
 export(tracking_chart)
+export(twoal)
 export(twoep)
 export(wndprb)
 importFrom(magrittr,"%>%")

[…]
