+++ NEWS.md
@@ -1,4 +1,4 @@
-rrricanes 0.1.1 (2017-06-08)
+rrricanes 0.1.2 (2017-06-08)
 ============================
 
 ### NEW FEATURES
