### 2017-06-09 timtrice/rrricanes#71 @timtrice
Redevelop tests
The current tests as they stand just take too long. In part, this is because some datasets get hit repeatedly (accessing annual datasets to retrieve individual storm data). But the slowest by far are the forecast/advisory tests. These have been trimmed down but they still take a bit of time.

The forecast/advisory tests should be expanded. A concern is modifiying a regex pattern to fit an issue for a specific advisory but consequentially breaking it for another advisory; there are many inconsistencies in these products. At this point of time there is no real way to know that patching one bug will not open up another. That's an issue.

I'm opening this issue to start noting ideas and methods to redevelop the tests. The time issue isn't going to get better; it just is what it is. But the tests needs to be redeveloped to accomodate the issues noted above and streamline the tests as much as possible.
