### 2017-06-09 772339e
Add section on package options
