### 2017-06-09 2bf2b4a
Added get_url_contents to address #61
* `get_url_contents` now handles all data gathering from the link
parameter. Function makes `rrricanes.http_attempts` with a timeout
value of `rrricanes.http_timeout` to attempt to get contents of page
and return with success. Otherwise, returns error. Hopefully this will
end or at least minimize the timeout issues.

* Added tests for `get_url_contents`

* Options added as noted above

* Removed calls to `status` (and associated function) as no longer
necessary.

* Added options to all tests to hopefully minimize timeout errors
causing tests to fail locall
[…]
