+++ R/wndprb.R
@@ -57,11 +57,8 @@ ep_prblty_stations <- function() {
 #' @export
 get_wndprb <- function(link) {
 
-    # Check status of link(s)
-    valid.link <- purrr::map_chr(link, status) %>% stats::na.omit()
-
     # Get all products for the current storm
-    products <- purrr::map(valid.link, get_products) %>% purrr::flatten_chr()
+    products <- purrr::map(link, get_products) %>% purrr::flatten_chr()
 
     # Filter out wndprb products
     products <- filter_wndprb(products)
