+++ man/status.Rd
@@ -1,22 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/base.R
-\name{status}
-\alias{status}
-\title{status}
-\usage{
-status(u)
-}
-\arguments{
-\item{u}{URL to test}
-}
-\value{
-URL if result is 'OK', otherwise, NA.
-}
-\description{
-Test URL status.
-}
-\details{
-Return URL if status is 'OK'. Otherwise, return NA and print
-failed URL.
-}
-\keyword{internal}

[…]
