### 2017-06-09 #118 @timtrice
I have released 0.1.2 into the master and develop branch.

Local checks and tests pass. I'm having issues passing them on Appveyor. Travis has passed. Appveyor wants to fail 2 or 3 tests in because of timeout issues. I hoped to have resolved this by adding functionality to reattempt data collection on timeout but it apparently isn't enough. At least, for testing. I'm not sure what else I can do to resolve those type of issues. I'll keep re-building the commit if it continues to fail but hopefully it'll catch a good run and finally go through.

Thank you again for taking the time to review my project and I look forward to hearing back from the community about it.

Tim
