### 2017-06-10 #118 @maelle
@timtrice Best loading message ever! 😁 I updated the editor checks.

I'm now looking for reviewers.
