+++ man/get_fstadv.Rd
@@ -15,7 +15,7 @@ Return dataframe of forecast/advisory data.
 \details{
 To disable the progress bar set option dplyr.show_progress to
     FALSE. You can also enable messages to show the storm and advisory
-    currentlby being worked by setting the rrricanes.working_msg to TRUE.
+    currentl by being worked by setting the rrricanes.working_msg to TRUE.
 }
 \seealso{
 \code{\link{get_storms}}, \code{\link{public}}
