### 2017-06-10 4d3ec83
Remove lazyeval import
