### 2017-06-10 #118 @maelle
@timtrice In [this code](https://github.com/ropensci/monkeylearn/blob/43b2ca8b36cf1adcbd34f6c8c34b9383b3dc06c3/R/classify.R#L67) I make the waiting time a bit longer with each attempt. I should have used `purrr::safely` as well. 😉

Have you told the owner of the website you're scraping about your package by the way? It'd also be good to know how ok it is for you to have a copy of their data.
