### 2017-06-10 #118 @timtrice
Thanks @maelle, I'll take a look at it. I'm trying a simple Sys.sleep right now.

Here is the latest example of a fail in Travis: https://travis-ci.org/timtrice/rrricanes/jobs/241505560

```
Quitting from lines 175-179 (getting_started.Rmd)
Error: processing vignette 'getting_started.Rmd' failed with diagnostics:
Timeout was reached
Execution halted
The command "R CMD build  ." failed and exited with 1 during .
Your build has been stopped.
```

Just constant...

The NHC webmaster is aware, yes. Addtionally, the data is considered [public domain](https://www.usa.gov/government-works). The [National Weather Service states](https://www.weather.gov/disclaimer):

```
Use of NOAA/NWS Data and Products

The information on National Weather Service (NWS) Web pages are in the public domain, unless specifically noted otherwise, and may be used without charge for any lawful purpose so long as you do not: 1) claim it is your own (e.g., by claiming copyright for NWS information -- see below), 2) use it in a manner that implies an endorsement or affiliation with NOAA/NWS, or 3) modify its content and then present it as official government material. You also cannot present information of your own in a way that makes it appear to be official government information.
```

I've tried to document thoroughly where the data originates and, to the best of my knowledge, have made no implicit or explicit claims the data is mine or that it is restricted.
