### 2017-06-10 timtrice/rrricanes#72 @mpadge
httr timout
@timtrice in regard to the timeout issues vexing your [ropensci submission](https://github.com/ropensci/onboarding/issues/118#issuecomment-307558695) and discussed with @maelle: While we awaiting the release of [`vcr`](https://github.com/ropensci/vcr), a useful workaround that locally caches `httr` calls is implemented [here](https://github.com/ropensci/osmdata/blob/master/tests/testthat/test-features.R) by simply stubbing out the `httr::GET` call with local data using [`stub.R` code](https://github.com/ropensci/osmdata/blob/master/tests/stub.R) (with appropriate acknowledgement of authors). Seems like that should work easily for all of your `get_url_contents` calls, simply by stubbing out different data for each test.
