### 2017-06-10 a8a9eed
Add sleep option for multiple data requests
