### 2017-06-10 timtrice/rrricanes#72 @timtrice
@mpadge, thank you. I'll take a look at this.

My main concern regarding the tests is if the HTML layout of the archives changes (as the actual text of the product shouldn't).

If, for some instance, the NHC does modify the layout, would this package recognize that? Or would the tests continue to perform under the assumption no changes were made?

If I'm misunderstanding the purpose of the package, I apologize and would appreciate clarity.
