### 2017-06-10 64e5354
Make options integers
