### 2017-06-10 907afef
Notate new option rrricanes.http_sleep
