### 2017-06-10 timtrice/rrricanes#72 @mpadge
yeah, you're right - it'll archive a frozen snapshot of how things looked at the time the call results were cached. Very high on my "things-to-do" list is to write a script to automatically update these cached results at the start of each month along the lines of this daily check for package updates that simply lives in my `.Rprofile`.
```
# only check for new packages once per day
chk_file <- "~/.Rold_pkg_check"
do_check <- TRUE
today <- strsplit (as.character (Sys.time ()), " ") [[1]] [1]
if (file.exists (chk_file)) {
  chk_date <- utils::read.table (chk_file, as.is=TRUE) [1, 1]
  if (chk_date == today)
    do_check = FALSE
}
write (today, file = chk_file)
if (do_check) {
  message ('Old package check for ', today, ' : ', appendLF=FALSE)
  old <- utils::old.packages ()
  if (!is.null (old))
    message ('Updatable packages: ',
             do.call (paste, as.list (rownames (old))), '\n')
  else
    message ('All packages up to date\n')
}
```
Totally easy to adapt that to re-generate your cached files on a daily or monthly basis. As soon as tests start to fail you'll know it's time for a CRAN update.

Having quickly looked through your impressively clean code, I could pretty quickly whip up a PR that stubs out an initial test if you want - just let me know.
