### 2017-06-10 timtrice/rrricanes#72 @timtrice
Mark, anything you want to contribute I'm all for it long as it resolves issues or improves the package. I'd be very much grateful for that!

Let me ask you this; there are maybe three different layouts used in the archives. As I mentioned the text of the products themselves should never change. The tests are meant to ensure little discrepancies here and there are accounted for when modifying regex patterns (and they need to be modified as well, as noted in #71).

So what if those tests are cached through your package and I/we develop a few more tests just to check the overall layout of a page matches what is expected.

What are your thoughts on that?
