### 2017-06-10 timtrice/rrricanes#72 @mpadge
All good info to know - I'll have a go sometime in the coming week and see what I can bash out. And a FYI: my interest in your package is coz I used to be a weather forecaster in cyclone (hurricane) prone Australia, and I'm a strong advocate for open met data. And further: not sure if you're yet on ropensci's slack chat, but @sckott just declared that `vcr ` is approaching finished, so this'll all get much easier quite soon
