### 2017-06-10 timtrice/rrricanes#72 @timtrice
Mark, thank you! Appreciate any way you can help.

I'll take a look at the `vcr` package when I get  chance. I had looked over it initally then got sidetracked. I'll have to find their slack channel as well.

I'm very interested to speak with you more about your experience with forecasting. I myself wanted to be a met and work for the NHC but issues with math in college unfortunately dissuaded me. I've had a fascination for hurricanes ever since I was about ten years old. If you don't mind, shoot me an email at tim.trice@gmail.com as I do want to add western and southern Pacific data to the package but I've never really explored how the other met centers publish their data.

I do have some big plans for the package - probably too big, who knows; recon, forecast models, ship/buoy data, etc. Scott informed me of a package that may help (or resolve) the buoy data. But I'd like to get enough clean data sources to help as much as possible.

If anything, I'll enjoy using it! :)
