### 2017-06-11 f85f185
Correct if statement for http_sleep
