+++ DESCRIPTION
@@ -1,7 +1,6 @@
 Package: rrricanes
 Type: Package
-Title: Web scraper for real-time and archived advisory products for Atlantic and 
-    east Pacific hurricanes and tropical storms
+Title: Web scraper for Atlantic and east Pacific hurricanes and tropical storms
 Description: Get archived data of past and current hurricanes and tropical 
   storms for the Atlantic and eastern Pacific oceans. Data is available for 
   storms since 1998.
