### 2017-06-11 fe276d9
Correct name in acknowledgements
