+++ README.Rmd
@@ -69,4 +69,4 @@ This project is licensed under the MIT License - see the [LICENSE.md](LICENSE.md
 ## Acknowledgments
 
 * [Molyneux, James](https://github.com/jimmylovestea)
-* [Rudis, BobRudis](https://github.com/hrbrmstr)
+* [Rudis, Bob](https://github.com/hrbrmstr)
