### 2017-06-11 16f2664
Update doc for get_prblty
