### 2017-06-11 07b9424
Update docs for fstadv funcs
