### 2017-06-11 826b480
Update docs for fstadv tidy funcs
