+++ man/tidy_wr.Rd
@@ -13,7 +13,16 @@ tidy_wr(df)
 Tidy current wind radius of a fstadv dataframe object.
 }
 \details{
-Takes current wind radius variables in a fstadv dataframe object,
-tidys the data and removes rows where all values are NA. Use Key, Adv and
-Date to join with other tidy dataframes.
+Returns tidy dataframe of current wind radius values for a cyclone.
+Returns only complete.cases (based on quadrants).
+\describe{
+   \item{Key}{Unique identifier of cyclone}
+   \item{Adv}{Advisory number}
+   \item{Date}{Date and time of advisory}
+   \item{Windfield}{Minimum wind speed expected}
+   \item{NE}{Radius of `Windfield` in the northeast quadrant}
+   \item{SE}{Radius of `Windfield` in the southeast quadrant}
+   \item{SW}{Radius of `Windfield` in the southwest quadrant}
+   \item{NW}{Radius of `Windfield` in the northwest quadrant}
+}
 }
