### 2017-06-11 83e9acf
Add pkg rlang to Imports
