+++ DESCRIPTION
@@ -23,6 +23,7 @@ Imports:
   maptools (>= 0.9-2), 
   purrr (>= 0.2.2.2), 
   readr (>= 1.1.0), 
+  rlang (>= 0.1.1),
   rnaturalearthdata (>= 0.1.0),
   rvest (>= 0.3.2), 
   stringr (>= 1.2.0), 
