### 2017-06-11 6d88fa0
Remove duplicate addons call
