+++ NEWS.md
@@ -1,4 +1,4 @@
-rrricanes 0.2.0 (yyyy-mm-dd)
+rrricanes 0.1.3 (yyyy-mm-dd)
 ============================
 
 ### NEW FEATURES
