+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [0.1.3] yyyy-mm-dd
 
 ### Added
   - `rrricanes.http_sleep` to control time to sleep between multiple HTTP requests.
