### 2017-06-11 afd394f
Modify test expectation
