### 2017-06-11 b6e6dd4
Add docs from pkgdown
