### 2017-06-11 426cf72
Hide funcs get_nhc_link, get_products
