+++ man/get_products.Rd
@@ -12,3 +12,4 @@ get_products(link)
 \description{
 Get list of all products from a storm's archive page
 }
+\keyword{internal}
