+++ man/get_nhc_link.Rd
@@ -13,3 +13,4 @@ trailing slash.}
 \description{
 Return root link of NHC archive pages.
 }
+\keyword{internal}
