+++ NAMESPACE
@@ -7,10 +7,8 @@ export(ep_prblty_stations)
 export(ep_tracking_chart)
 export(get_discus)
 export(get_fstadv)
-export(get_nhc_link)
 export(get_posest)
 export(get_prblty)
-export(get_products)
 export(get_public)
 export(get_storm_data)
 export(get_storms)
