### 2017-06-11 #118 @timtrice
Apologies if I keep moving the goal posts, so-to-speak, but I found errors in some of the tidy functions and had to rush a patch. Release 0.1.3 has been pushed to master.

https://github.com/timtrice/rrricanes/releases/tag/v0.1.3
