### 2017-06-13 timtrice/rrricanes#72 @mpadge
I had a go, but it's not going to work in the straightforward way I implement because of `purrr::safely`, which does not want to play nicely with the results of `curl::curl_fetch_memory` for reasons I can't figure out at the moment.
