### 2017-06-13 timtrice/rrricanes#72 @timtrice
The `purrr::safely` can be dropped, if it helps. I've only utilized it in this one function. I copied that piece of code from [Stack Overflow](https://stackoverflow.com/a/37374445/2576935).
