### 2017-06-13 timtrice/rrricanes#72 @mpadge
okay, i'll give it a try that way. Main advantage of `purrr::safely` in these kinds of contexts is in testing, but since you want to avoid actual `GET` requests in tests, it's kinda redundant anyway.
