### 2017-06-13 055a180
add stubbed tests in testthat/test-aaa.R
