+++ README.Rmd
@@ -3,12 +3,11 @@ output: github_document
 ---
 
 [![minimal R version](https://img.shields.io/badge/R%3E%3D-3.3.3-6666ff.svg)](https://cran.r-project.org/) 
-[![packageversion](https://img.shields.io/badge/Package%20version-0.1.3-orange.svg?style=flat-square)](commits/develop-0.2.0)
-[![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/rrricanes)](https://cran.r-project.org/package=rrricanes) 
-[![Build Status](https://travis-ci.org/timtrice/rrricanes.svg?branch=develop-0.2.0)](https://travis-ci.org/timtrice/rrricanes) 
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/timtrice/rrricanes?branch=develop-0.2.0&svg=true)](https://ci.appveyor.com/project/timtrice/rrricanes) 
-[![codecov](https://codecov.io/gh/timtrice/rrricanes/branch/develop-0.2.0/graph/badge.svg)](https://codecov.io/gh/timtrice/rrricanes) 
-
+[![GitHub (pre-)release](https://img.shields.io/github/release/timtrice/rrricanes/all.svg)](https://github.com/timtrice/rrricanes/tags)
+[![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/rrricanes)](https://cran.r-project.org/package=rrricanes)
+[![Build Status](https://img.shields.io/travis/timtrice/rrricanes/feature-47.svg)](https://travis-ci.org/timtrice/rrricanes)
+[![AppVeyor Build Status](https://img.shields.io/appveyor/ci/timtrice/rrricanes/feature-47.svg)](https://ci.appveyor.com/project/timtrice/rrricanes)
+[![codecov](https://codecov.io/gh/timtrice/rrricanes/branch/feature-47/graph/badge.svg)](https://codecov.io/gh/timtrice/rrricanes) 
 
 ```{r, echo = FALSE}
 knitr::opts_chunk$set(
@@ -32,16 +31,26 @@ Please view the vignette 'Getting Started':
 vignette("getting-started", package = "rrricanes")
 ```
 
+[Online documentation](https://timtrice.github.io/rrricanes/) is also available.
+
 ### Prerequisites
 
-`rrricanes` only require an active internet connection as data is extracted from online archives. 
+`rrricanes` requires an active internet connection as data is extracted from online sources.
+
+Additionally, to use high resolution tracking maps you must install the `rnaturalearthhires` package:
+
+```r
+install.packages("rnaturalearthhires",
+                 repos = "http://packages.ropensci.org",
+                 type = "source")
+```
 
 ### Installing
 
 `rrricanes` is currently only available in GitHub. It can be installed using the `devtools` package:
 
 ```r
-devtools::install_github("timtrice/rrricanes")
+devtools::install_github("timtrice/rrricanes@v0.2.0-alpha")
 ```
 
 ## Built With
@@ -69,4 +78,5 @@ This project is licensed under the MIT License - see the [LICENSE.md](LICENSE.md
 ## Acknowledgments
 
 * [Molyneux, James](https://github.com/jimmylovestea)
+* [Padgham, Mark](https://github.com/mpadge)
 * [Rudis, Bob](https://github.com/hrbrmstr)
