### 2017-06-14 fb06532
Update documentation on prblty_station funcs
