### 2017-06-14 a6de85e
Update load_storm_data doc
