### 2017-06-14 9b53a0e
Add GIS functionality per #47
* GIS functions now available. Please note some of these products may not exist for every available cyclone/advisory.
 + `gis_advisory` Typically will include current and past track data, forecast track data, forecast cone (margin of error) and wind radius data.
 + `gis_breakpoints` List of breakpoints typically used for watch/warning areas but is not a requirement.
 + `gis_latest` Retrieves the latest GIS products for all active storms.
 + `gis_outlook` Retrives the latest tropical weather outlook in shapefile format.
 + `gis_prob_storm_surge` Probabilistic storm
[…]
