+++ man/shp_to_df.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/gis.R
+\name{shp_to_df}
+\alias{shp_to_df}
+\title{shp_to_df}
+\usage{
+shp_to_df(obj)
+}
+\arguments{
+\item{obj}{Spatial object to convert. See details.}
+}
+\description{
+Convert shapefile object to dataframe
+}
+\details{
+Takes a SpatialLinesDataFrame object or SpatialPolygonsDataFrame
+object and converts into a dataframe that can be plotted in ggplot2.
+}
