+++ DESCRIPTION
@@ -23,6 +23,7 @@ Imports:
   maptools (>= 0.9-2), 
   purrr (>= 0.2.2.2), 
   readr (>= 1.1.0), 
+  rgdal (>= 1.2-7),
   rlang (>= 0.1.1),
   rnaturalearthdata (>= 0.1.0),
   rvest (>= 0.3.2), 
@@ -32,6 +33,7 @@ Imports:
   xml2 (>= 1.1.1)
 Suggests:
   devtools, 
+  geosphere,
   knitr, 
   rmarkdown,
   testthat,
