+++ vignettes/articles/wind_pressure.Rmd
@@ -1,0 +1,146 @@
+---
+title: "Wind and Pressure"
+author: "Tim Trice"
+date: "June 14, 2017"
+output: html_document
+references:
+- id: sshws
+  title: The Saffir-Simpson Hurricane Wind Scale
+  author:
+  - family: Schott
+    given: Timothy
+  - family: Landsea
+    given: Chris
+  - family: Hafele
+    given: Gene
+  - family: Lorens
+    given: Jeffrey
+  - family: Taylor
+    given: Arthur
+  - family: Thurm
+    given: Harvey
+  - family: Ward
+    given: Bill
+  - family: Willis
+    given: Mark
+  - family: Zaleski
+    given: Walt
+  URL: 'http://www.nhc.noaa.gov/pdf/sshws.pdf'
+  issued:
+    year: 2012
+    month: 2
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE, 
+                      fig.align = "center")
+```
+
+```{r, message = FALSE}
+library(dplyr)
+library(gganimate)
+library(ggplot2)
+library(rrricanes)
+library(tidyr)
+```
+
+```{r}
+# Unique ID of storm
+key <- "AL092008"
+# Advisory holding details we wish to plot
+adv <- 42
+```
+
+## Data
+
+```{r}
+# Loads AL2008_fstadv dataframe
+# Please note load_storm_data is likely to be modified in future releases. 
+load_storm_data(years = 2008, basins = "AL", products = c("fstadv", "wndprb"))
+```
+
+Filter out Hurricane Ike data.
+
+```{r}
+fstadv <- AL2008_fstadv %>% filter(Key == key, Adv <= adv)
+```
+
+## Wind Profile
+
+```{r}
+# Plot wind values
+fstadv %>% ggplot(aes(x = Date, y = Wind)) + 
+    geom_line() + 
+    geom_point(aes(color = Status), size = 3) + 
+    scale_y_continuous(name = "Wind (kts)") + 
+    theme_bw() + 
+    theme(legend.position = "bottom", 
+          legend.box = "vertical") +
+    labs(title = "Wind Profile", 
+         caption = sprintf("rrricanes %s", packageVersion("rrricanes")))
+```
+
+## Pressure Profile
+
+```{r}
+# Plot pressure values
+fstadv %>% ggplot(aes(x = Date, y = Pressure)) + 
+    geom_line() + 
+    geom_point(aes(color = Status), size = 3) + 
+    scale_y_continuous(name = "Pressure (mb)") + 
+    theme_bw() + 
+    theme(legend.position = "bottom", 
+          legend.box = "vertical") +
+    labs(title = "Pressure Profile", 
+         caption = sprintf("rrricanes %s", packageVersion("rrricanes")))
+```
+
+## Wind/Pressure Relational Change
+
+```{r}
+fstadv %>%
+    mutate(WindDist = (Wind - min(Wind))/(max(Wind) - min(Wind)),
+           PressDist = (Pressure - max(Pressure))/(max(Pressure) - min(Pressure))) %>%
+    gather(Var, Val, WindDist, PressDist) %>% 
+    ggplot(aes(x = Date, y = Val, group = Var, color = Var)) + 
+    geom_line(size = 1) + 
+    scale_color_discrete(labels = c("Pressure Change", "Wind Change")) + 
+    theme_bw() + 
+    theme(legend.position = "bottom", 
+          legend.title = element_blank()) + 
+    labs(title = "Wind/Pressure Relational Change", 
+         subtitle = "",
+         caption = sprintf("rrricanes %s", packageVersion("rrricanes")), 
+         y = "")
[…]
