### 2017-06-14 66fadb0
Add rgeos support for gclib
