### 2017-06-14 8576863
Accept dplyr == 0.5.0 and >= 0.5.0 (#74)
* dplyr 0.6.0-rc renames .cols to .vars in mutate_at and other funcs.
Have accomodated this change to remove messages. Support for dplyr 0.5.0
will be removed in future releases.

* All tests and checks pass
