### 2017-06-14 ecc715c
Roll back rlang version dependency
