### 2017-06-15 51c900c
Unstage pkgdown docs
