### 2017-06-15 0c65ddd
Unstage pkgdown articles
