+++ vignettes/articles/plot_latest_advisory_cyclone.Rmd
@@ -1,203 +1,0 @@
----
-title: "Plot Current and Forecast Positions of a Cyclone"
-author: "Tim Trice"
-date: "June 14, 2017"
-output: html_document
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE, 
-                      fig.width = 10, 
-                      fig.asp = 1, 
-                      fig.align = "center")
-```
-
-```{r, message = FALSE}
-library(dplyr)
-library(ggplot2)
-library(rrricanes)
-library(tibble)
-library(tidyr)
-library(sp)
-```
-
-```{r}
-# Unique ID of storm
-key <- "AL092008"
-# Advisory holding details we wish to plot
-adv <- 42
-```
-
-## Forecast/Advisory Data
-
-Load `fstadv` products for Atlantic basin, 2008.
-
-```{r}
-# Please note load_storm_data is likely to be modified in future releases. 
-load_storm_data(years = 2008, basins = "AL", products = "fstadv")
-```
-
-Filter out Hurricane Ike data.
-
-```{r}
-fstadv <- AL2008_fstadv %>% filter(Key == key, Adv <= adv)
-```
-
-Generate a base plot of the Atlantic ocean.
-
-```{r}
-bp <- al_tracking_chart(color = "black", fill = "white", size = 0.1, res = 50) + 
-    scale_x_continuous(expand = c(0, 0)) +
-    scale_y_continuous(expand = c(0, 0)) +
-    labs(x = "Lon",
-         y = "Lat",
-         caption = sprintf("rrricanes %s", packageVersion("rrricanes")))
-```
-
-Tidy advisory points (removing forecast and wind radius details).
-
-```{r}
-df_tidy_adv <- tidy_fstadv(fstadv)
-```
-
-Tidy forecast points.
-
-```{r}
-df_tidy_fcst <- tidy_fcst(fstadv) %>% 
-    filter(Adv == adv)
-```
-
-Plot to-date track and forecast.
-
-```{r}
-bp + 
-    geom_point(data = df_tidy_adv, 
-               aes(x = Lon, y = Lat, color = Status, size = Wind)) + 
-    geom_path(data = df_tidy_fcst, 
-               aes(x = Lon, y = Lat)) + 
-    geom_point(data = df_tidy_fcst, 
-               aes(x = Lon, y = Lat, size = Wind)) + 
-    theme(legend.position = "bottom", 
-          legend.box = "vertical")
-```
-
-## GIS Advisory Forecast Track, Cone of Uncertainty, and Watches/Warnings
-
-```{r}
-gis_adv <- gis_advisory(key = key, advisory = adv) %>% gis_download()
-```
-
-Get bounding box of the forecast polygon.
-
-```{r}
-bbox <- bbox(gis_adv$al092008_042_5day_pgn)
-```
-
-Rebuild a base plot of the Atlantic ocean.
-
-```{r}
-bp <- al_tracking_chart(color = "black", fill = "white", size = 0.1, res = 50)
-```
-
-We can "zoom in" on the chart with `bbox`.
-
-```{r}
-lat_min <- bbox[2,1] - 5
-lat_max <- bbox[2,2] + 5
-lon_min <- bbox[1,1] - 10
-lon_max <- bbox[1,2] + 10
-```
-
-Build a thin tracking map for the inset.
-
-```{r}
-bp_inset <- ggplotGrob(bp +
-    geom_rect(mapping = aes(xmin = lon_min, xmax = lon_max,
-                            ymin = lat_min, ymax = lat_max),
-              color = "red", alpha = 0) +
-    theme_bw() +
-    theme(axis.title = element_blank(),
-          axis.ticks = element_blank(),
[…]
