### 2017-06-15 66062d8
Modify func load_storm_data #76
* load_storm_data returns dataframe requested. fstadv dataset is tidied.
Feeling was that accessing by basin, year was too restrictive. Though
loading complete datasets may be unnecessary for some use, this has
been resolved by the `rrricanesdata` repository being mirrored in
data.world. The data.world option gives users a third choice to view
cyclone data but with much more flexibility to get the data needed.

* Updated documentation

* Opened #77; need a script to scrape advisories for current cyclones
ASAP.
