### 2017-06-15 24999ac
Add pkg dwapi to Suggests
