+++ DESCRIPTION
@@ -34,6 +34,7 @@ Imports:
   xml2 (>= 1.1.1)
 Suggests:
   devtools, 
+  dwapi,
   geosphere,
   knitr, 
   rmarkdown,
