### 2017-06-15 3a0406f
Reintroduce pkgdown docs
