+++ tests/testthat/test_fstadv.R
@@ -14,9 +14,6 @@ al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
 df.al092008.fstadv <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_fstadv()
 load(system.file("extdata", "al092008.fstadv.Rda", package = "rrricanes"))
 
-## ---- * 2016, AL -------------------------------------------------------------
-load_storm_data(years = 2016, basins = "AL", products = "fstadv")
-
 ## ---- Test fstadv() ----------------------------------------------------------
 #' Test return of fstadv()
 test_that("Test fstadv()", {
@@ -25,8 +22,8 @@ test_that("Test fstadv()", {
 
 ## ---- Test tidy_fstadv() -----------------------------------------------------
 test_that("Test tidy_fstadv()", {
-    x <- tidy_fstadv(AL2016_fstadv)
-    expect_identical(dim(x), c(427L, 18L))
+    x <- tidy_fstadv(al092008.fstadv)
+    expect_identical(dim(x), c(53L, 18L))
     expect_identical(names(x),
                      c("Key", "Adv", "Date", "Status", "Name", "Lat", "Lon",
                        "Wind", "Gust", "Pressure", "PosAcc", "FwdDir",
@@ -36,8 +33,8 @@ test_that("Test tidy_fstadv()", {
 
 ## ---- Test tidy_wr() -----------------------------------------------------
 test_that("Test tidy_wr()", {
-    x <- tidy_wr(AL2016_fstadv)
-    expect_identical(dim(x), c(637L, 8L))
+    x <- tidy_wr(al092008.fstadv)
+    expect_identical(dim(x), c(138L, 8L))
     expect_identical(names(x),
                      c("Key", "Adv", "Date", "WindField", "NE", "SE", "SW",
                        "NW"))
@@ -45,8 +42,8 @@ test_that("Test tidy_wr()", {
 
 ## ---- Test tidy_fcst() -----------------------------------------------------
 test_that("Test tidy_fcst()", {
-    x <- tidy_fcst(AL2016_fstadv)
-    expect_identical(dim(x), c(2268L, 8L))
+    x <- tidy_fcst(al092008.fstadv)
+    expect_identical(dim(x), c(298L, 8L))
     expect_identical(names(x),
                      c("Key", "Adv", "Date", "FcstDate", "Lat", "Lon", "Wind",
                        "Gust"))
@@ -54,8 +51,8 @@ test_that("Test tidy_fcst()", {
 
 ## ---- Test tidy_fcst_wr() -----------------------------------------------------
 test_that("Test tidy_fcst_wr()", {
-    x <- tidy_fcst_wr(AL2016_fstadv)
-    expect_identical(dim(x), c(2958L, 9L))
+    x <- tidy_fcst_wr(al092008.fstadv)
+    expect_identical(dim(x), c(587L, 9L))
     expect_identical(names(x),
                      c("Key", "Adv", "Date", "FcstDate", "WindField", "NE",
                        "SE", "SW", "NW"))
