### 2017-06-15 3f6e8db
Remove call to load_storm_data
