### 2017-06-16 8abb063
Update package documentation
