### 2017-06-16 #118 @timtrice
[Release 0.2.0-1](https://github.com/timtrice/rrricanes/releases/tag/v0.2.0-1) is now in the master branch. There are no major changes scheduled anytime soon.
