### 2017-06-16 cf832a1
Set default val for p when func called directly
