+++ R/discus.R
@@ -54,7 +54,7 @@ get_discus <- function(link) {
 #' @return Dataframe
 #' @seealso \code{\link{get_discus}}
 #' @keywords internal
-discus <- function(link, p) {
+discus <- function(link, p = dplyr::progress_estimated(n = 1)) {
 
     p$pause(0.5)$tick()$print()
 
