### 2017-06-16 aab6f5e
Modify exptected storm headers for some products
* Some items in patterns were duplicates (other than the basin
identifier appended to the end); that's been cleaned up

* Some new products did not fit existing header patterns. Modified
accordingly.
