### 2017-06-16 b0668ed
Modify pattern; ignored intermediate advisories
