+++ R/posest.R
@@ -67,7 +67,7 @@ posest <- function(link, p = dplyr::progress_estimated(n = 1)) {
 
     # Make sure this is a public advisory product
     if (!any(stringr::str_count(contents,
-                                c("MIATCEAT", "MIATCEEP", "MEATIEST"))))
+                                c("MIATCE", "MEATIEST", "WTNT"))))
         stop(sprintf("Invalid Position Estimate link. %s", link))
 
     df <- create_df_posest()
