### 2017-06-16 3670b3c
Add fstadv to load_storm_data
