+++ man/load_storm_data.Rd
@@ -4,8 +4,8 @@
 \alias{load_storm_data}
 \title{load_storm_data}
 \usage{
-load_storm_data(dataset = c("adv", "discus", "fcst", "fcst_wr", "posest",
-  "prblty", "public", "storms", "update", "wndprb", "wr"))
+load_storm_data(dataset = c("adv", "discus", "fcst", "fcst_wr", "fstadv",
+  "posest", "prblty", "public", "storms", "update", "wndprb", "wr"))
 }
 \arguments{
 \item{dataset}{A dataset to return}
@@ -24,6 +24,7 @@ storm data and may be modified in future releases.
     \item{fcst}{Forecast positions of tropical cyclones}
     \item{fcst_wr}{Forecast wind radii data for each forecast observation in
     `fcst`}
+    \item{fstadv}{Complete Forecast/Advisory data}
     \item{posest}{Position estimates}
     \item{prblty}{Strike probabilities for a given location. This product was
     deprecated after the 2005 hurricane season.}
