### 2017-06-16 9b414f8
Add WTPA to discus header (EP041999)
