### 2017-06-16 c8e939e
Add MIAWRKAD1 to discus headers
