### 2017-06-17 ac51993
Fix replace NOON with 12 PM (#78)
