### 2017-06-17 0410b15
Return header if param ret is NULL
