### 2017-06-17 2b3acf9
Added fallback for header check
