### 2017-06-17 43840fc
Allow for 1-digit Date values
