### 2017-06-17 28122ed
Account for "MIDNIGHT" in headers
