+++ R/scrapers.R
@@ -87,6 +87,20 @@ scrape_date <- function(header) {
                                                         "[:digit:]{4})\n"),
                                                         "\n12 PM\\2\n")
 
+    # Same thing for "MIDNIGHT"
+    if (stringr::str_count(header,
+                           pattern = paste0("\nMIDNIGHT [:upper:]{3} [:upper:]{3} ",
+                                            "[:upper:]{3} [:digit:]{2} ",
+                                            "[:digit:]{4}\n")))
+        header <- stringr::str_replace(header,
+                                       pattern = paste0("\n(MIDNIGHT)( ",
+                                                        "[:upper:]{3}",
+                                                        " [:upper:]{3} ",
+                                                        "[:upper:]{3} ",
+                                                        "[:digit:]{2} ",
+                                                        "[:digit:]{4})\n"),
+                                       "\n12 AM\\2\n")
+
     ptn <- paste0("(?<=(?:\n|\r))",
                   "([:digit:]{1,2})", # Hour
                   "(?<=[:digit:]{1})([:digit:]{2})?", # Minute
