### 2017-06-17 b51dafe
Account for "12 NOON" in header...
