### 2017-06-18 37b8c18
Update to extract real-time advisories as cron
This script is designed to run on a scheduled cron when cyclones are
active or forecast to become active. Updates the rrricanesdata
repository with new information.
