+++ data-raw/fetch_products.R
@@ -1,156 +1,238 @@
-#' Fetch Products
+#'----------------------------------------------------------------------------'#
+#'                                                                            '#
+#' Fetch Products                                                             '#
+#'                                                                            '#
+#'----------------------------------------------------------------------------'#
 #'
-#' This script is designed to run the latest version of rrricanes and grab
-#' every product available for every storm, each basin. If the product dataframe
-#' is not empty, the dataframe is saved as a CSV to the directory assigned in
-#' `d`. If `s` is TRUE, the dataframe is also returned to the global
-#' environment.
+#' Scrape basin index xml files for active storms. If no storms are present,
+#' exit script.
 #'
-#' Each product output file and dataframe is named to make identification as
-#' easy as possible. It is formatted as AB_C where A is the basin, B is the
-#' year, and C is the product. For example, AL1998_fstadv means all data is
-#' related to the fstadv prodcut for the Atlantic basin, 1998.
+#' If storms do exist, extract wallet ID for each storm.
 #'
-#' If the product being worked is "fstadv", a summary dataframe is built as
-#' well. This summary dataframe contains the Key, Name, StartDate and EndDate
-#' for each storm.
+#' In file pub_dates.csv exists each wallet id for each basin along with the
+#' last published date processed. If each product in each active wallet has a
+#' pubDate greater than the value in pub_dates.csv, parse product, add to
+#' product dataframe and push to GitHub repo.
 #'
-#' The summary dataframe and csv file is named similar to the product files with
-#' only the basin and year. For example, EP2017 is the summary dataframe/file
-#' for all storms in the East Pacific for the year 2017.
+#' This script need not run on a consistent basis but, for the time being, will
+#' have to be micromanaged somewhat. If there is no possibility of a storm
+#' developing within the next 24-48 hours, then no need to run.
 #'
-#' This script can take some time to run; especially for fstadv products. This
-#' time is amplified significantly the more years it is gathered. The biggest
-#' issue is a connection timeout; if internet connection breaks or if the NHC
-#' archive site experiences a brief disruption. Each yearly file is written
-#' on completion. So if you run 1998:2017 and the script is interrupted, review
-#' the directory (or your global environment if `s` is TRUE) for the last
-#' object built or saved. Then modify all variables in section "Variables"
-#' accordingly.
+#' However, if there is, say, >50% chance of development then the script should
+#' probably run every 10-15 minutes.
 #'
-#' For example, if the script stopped executing and the last dataframe (by year)
-#' is AL2014, modify the years variable to 2015:2017.
-#'
-#' Same goes for the products. The products variable is loaded in alphabetical
-#' order. So, if the script is interruped and the last dataframe saved is
-#' AL2014_fstadv, the products variable must be modified with discus and fstadv
-#' removed. Otherwise, you're just being redundant.
+#' Execution:
+#' R CMD BATCH R/fetch_products.R
 
 ## ---- Libraries --------------------------------------------------------------
-library(assertr)
 library(dplyr)
-library(magrittr)
+library(git2r)
 library(purrr)
 library(readr)
 library(rrricanes)
+library(rvest)
+library(sendmailR)
+library(stringr)
 library(tibble)
+library(xml2)
 
-## ---- Variables --------------------------------------------------------------
-# Save dataframe to global env?
-s <- TRUE
-# Where is data being saved? Add trailing slash
-d <- "./datasets/"
-# Basins to retrieve
-basins <- c("AL", "EP")
-# Years to retrieve
[…]
