### 2017-06-19 6af31c2
Add "POTENTIAL TROPICAL CYCLONE" to status
