### 2017-06-19 #118 @maelle
Many thanks  @jsta @robinsones for accepting to review this package! :smile_cat: Your reviews are due on 2017-06-19.

As a reminder here is a link to the [reviewing guide](https://github.com/ropensci/onboarding/blob/master/reviewing_guide.md) and here is one to the [review template](https://github.com/ropensci/onboarding/blob/master/reviewer_template.md).
