### 2017-06-22 30d36d7
shrug emoji breaking devtools::check
