### 2017-06-22 #118 @jsta
@timtrice I am having trouble building after `packrat` seems to have activated itself. I wonder if the issue that is that `packrat` is tracking a lot of packages. Consider the output of `packrat::status()`:

>  The following packages are tracked by packrat, but are no longer available in the local library nor present in your code:
                _
    HURDAT        0.1.0
    PKI           0.1-3
    RgoogleMaps   1.4.1
    assertr       2.0.2.2
    cellranger    1.1.0
    debugme       1.0.2
    forcats       0.2.0
    gganimate     0.1.0.9000
    ggmap         2.6.1
    gh            1.0.0
    haven         1.0.0
    modelr        0.1.0
    pkgdown       0.1.0.9000
    processx      2.0.0
    proto         1.0.0
    rematch       1.0.1
    rsconnect     0.8
    servr         0.6
    usethis       0.0.0.9000
    viridisLite   0.2.0

> You can call packrat::snapshot() to remove these packages from the lockfile, or if you intend to use these packages, use packrat::restore() to restore them to your private library.

> The following packages are used in your code, tracked by packrat, but no longer present in your library:
                              from   to
    RCurl                 1.95-4.8   NA
    bindrcpp                   0.1   NA
    broom                    0.4.2   NA
    caTools                 1.17.1   NA
    desc                     1.1.0   NA
    devtools                1.13.1   NA
    dplyr                    0.7.0   NA
    dwapi                    0.1.1   NA
    evaluate                  0.10   NA
    geosphere                1.5-5   NA
    ggplot2                  2.2.1   NA
    htmltools                0.3.6   NA
    httpuv                   1.3.3   NA
    httr                     1.2.1   NA
    knitr                     1.16   NA
    latticeExtra            0.6-28   NA
    lubridate                1.6.0   NA
    magick                     0.4   NA
    mapproj                  1.2-5   NA
    maptools                 0.9-2   NA
    markdown                   0.8   NA
    memoise                  1.1.0   NA
    munsell                  0.4.3   NA
    pander                   0.6.0   NA
    pkgbuild            0.0.0.9000   NA
    pkgload             0.0.0.9000   NA
    plyr                     1.8.4   NA
    png                      0.1-7   NA
    praise                   1.0.0   NA
    psych                    1.7.5   NA
    purrr                  0.2.2.2   NA
    raster                   2.5-8   NA
    rasterVis                 0.41   NA
    readr                    1.1.1   NA
    readxl                   1.0.0   NA
    reshape2                 1.4.2   NA
    rgdal                    1.2-7   NA
    rgeos                   0.3-23   NA
    rjson                   0.2.15   NA
    rlang               0.1.1.9000   NA
    rmarkdown                  1.5   NA
    rnaturalearthdata        0.1.0   NA
    roxygen2                 6.0.1   NA
    rprojroot                  1.2   NA
    rstudioapi                 0.6   NA
    rvest                    0.3.2   NA
    scales                   0.4.1   NA
    selectr                  0.3-1   NA
    sp                       1.2-4   NA
    stringi                  1.1.5   NA
    stringr                  1.2.0   NA
    testthat                 1.0.2   NA
    tibble                   1.3.3   NA
    tidyr                    0.6.3   NA
    whisker                  0.3-2   NA
    withr                    1.0.2   NA
    xml2                     1.1.1   NA
    yaml                    2.1.14   NA
    zoo                      1.8-0   NA

> Use packrat::restore() to restore these libraries.

Maybe you could consider cleaning these up to a minimal set?
