### 2017-06-22 #118 @timtrice
@jsta, thank you. I'm about getting tired of packrat myself but never noticed outsides with installation. Only use it for testing (such as the new dplyr).

Does it really even need to be in the repo? I have no problems removing it from remote and just keeping it on my setup.
