### 2017-06-22 #118 @jsta
Seems to be most useful for archiving rarely updated packages rather than tracking a lot of dependencies in flux but I'm not very familiar with it myself so I can't say how much it is worth the hassle.
