### 2017-06-22 #118 @timtrice
You have the .Rbuildignore file as well, correct? It shouldn't even pay attention to packrat if I understood correctly.
