### 2017-06-22 #118 @jsta
I do have .Rbuildignore. I ended up running `packrat::disable()` and now everything is OK.

I do think packrat is a little out of it's element in this case and you might consider removing it.
