### 2017-06-22 #118 @jsta
* I am having timeout issues with building the getting_started vignette. I wonder if there is a particular year with very few hurricanes that would solve the timeout problem.

	* I cannot build the data.world vignette. Probably because I don't have an API key set up. You may want to consider setting the code chunks to `eval=FALSE`.

	* You might want to consider numbering your vignettes (https://twitter.com/fmic_/status/875385557819555841) so that the `data.world` vignette is listed last.

	* There are no examples for `knots_to_mph`, `mb_to_in`, `status_abbr_to_str`, `get_discus`, `get_fstadv`, `tidy_fstadv`, `tidy_wr`, `tidy_fcst`. Maybe some can be changed to non-exported?

	* Missing a maintainer field in the DESCRIPTION
	* What does the `Additional_repositories` field signify?

	* I really like the `tidy_` functions. I wonder if it would make it easier on the end-user to have the `get_` functions return tidy results by default with an optional argument to return "messy" results.

	* Wow, this package has a lot of tests! Could you speed up package checking by consolidating some of these? I notice that some of the lines are re-downloading data from earlier lines.

	* Also, you might consider using the `skip_on_cran` function for lines that call an external download as recommended by the rOpenSci packaging guide.

---

This package is VERY comprehensive! I am impressed with the breadth of information accessible from this package.

I get the following note from `devtools::check()` :

```
checking R code for possible problems ... NOTE

fstadv_forecasts: no visible global function definition for

  ‘packageVersion’

prblty: no visible global function definition for ‘packageVersion’

wndprb: no visible global function definition for ‘packageVersion’

  packageVersion

Consider adding

  importFrom("utils", "packageVersion")

to your NAMESPACE file.
```
