### 2017-06-22 #118 @timtrice
@jsta thank you very much for taking the time to review it!

1. Vignettes:
  a. Yes, timeouts have been a [major headache](https://github.com/timtrice/rrricanes/issues/72) that I have to find a way to resolve. This is an issue with the NHC website and occurs on Travis and Appveyor. I added options to control the timeout, number of attempts and a delay between attempts; they just do not seem to have made a difference. But I am looking at alternatives.
  b. The data.world article is not supposed to be a vignette but rather an article through `pkgdown`. I will revisit this and modify it accordingly.
2. Examples: I will add them. `get_discus` and `get_fstadv` along with the tidy functions will be `\dontrun` due to the timeout issues.
3. Community Guidelines
  a. Will update DESCRIPTION to list the maintainer
  b. `Additional repositories` slipped through; originally it was included to add the `rnaturalearthhires` package. However, that was removed as a dependent as I seemed to struggle with getting it through Travis and Appveyor. It should be documented that to use 10nm resolution on the tracking charts that it requires installation of `rnaturalearthhires` which is also mentioned in the README.
4. Functionality: I can certainly add a parameter for tidy or messy. If tidy then `get_fstadv` will have to return a list of dataframes which makes the expected output different from that of the other `get` functions; I didn't personally think that was a good option. If it's acceptable, I'm happy to add it.
5. Automated tests.
  a. Yes, tons of tests. And these bite me in more ways than you can image. Perfect example: the recent development of Bret and Cindy introduced a new header: "Potential Tropical Cyclone". This was a new type of advisory added this year (and the current release does not accurately capture the name and status of the cyclones for these advisories). In this instance, it was easy to modify this. However, these advisory packages are anything but consistent over the last 20 years; anything form formatting changes to typos. I do intend or [rewriting them](https://github.com/timtrice/rrricanes/issues/71); just unsure the best way to do that at this point.
  b. I will review the `skip_on_cran` functionality and add it.
6. Will add `packageVersion` to imports.
