### 2017-06-22 1875108
Add missing examples for functions
