+++ man/tidy_fcst_wr.Rd
@@ -29,3 +29,9 @@ dataframes.
    \item{NW}{Radius in nautical miles for northwest quadrant}
 }
 }
+\examples{
+\dontrun{
+get_fstadv("http://www.nhc.noaa.gov/archive/1998/1998ALEXadv.html") \%>\%
+    tidy_fcst_wr()
+}
+}
