### 2017-06-22 4f70f14
Add Maintainer, remove Additional_repositories
