### 2017-06-23 3c8c1a0
Build vignettes off saved data
Vignette data files are written to data dir using data-raw/vignettes.R.

This should help avoid timeout errors plus help builds faster.
