### 2017-06-23 680e7fc
Modify tests; add skip_on_cran
