+++ tests/testthat/test_fstadv.R
@@ -1,22 +1,14 @@
 context("Forecast/Advisory Products (fstadv)")
 
-# Set timeout options
-opt.timeout <- getOption("rrricanes.http_timeout")
-opt.attempts <- getOption("rrricanes.http_attempts")
-options("rrricanes.http_timeout" = 1)
-options("rrricanes.http_attempts" = 5)
-
-## ---- 2008, AL ---------------------------------------------------------------
-al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
-
-## ---- Base Data --------------------------------------------------------------
-## ---- * 2008, AL -------------------------------------------------------------
-df.al092008.fstadv <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_fstadv()
 load(system.file("extdata", "al092008.fstadv.Rda", package = "rrricanes"))
 
 ## ---- Test fstadv() ----------------------------------------------------------
 #' Test return of fstadv()
 test_that("Test fstadv()", {
+    skip_on_cran()
+    al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
+    df.al092008.fstadv <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_fstadv()
+
     expect_identical(al092008.fstadv, df.al092008.fstadv)
 })
 
@@ -57,7 +49,3 @@ test_that("Test tidy_fcst_wr()", {
                      c("Key", "Adv", "Date", "FcstDate", "WindField", "NE",
                        "SE", "SW", "NW"))
 })
-
-# Reset options
-options("rrricanes.http_timeout" = opt.timeout)
-options("rrricanes.http_attempts" = opt.attempts)
