+++ tests/testthat/test_discus.R
@@ -1,17 +1,5 @@
 context("Storm Discussions (discus)")
 
-# Set timeout options
-opt.timeout <- getOption("rrricanes.http_timeout")
-opt.attempts <- getOption("rrricanes.http_attempts")
-options("rrricanes.http_timeout" = 1)
-options("rrricanes.http_attempts" = 5)
-
-## ---- 2008, AL ---------------------------------------------------------------
-al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
-
-## ---- Base Data --------------------------------------------------------------
-## ---- * 2008, AL -------------------------------------------------------------
-df.al092008.discus <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_discus()
 load(system.file("extdata", "al092008.discus.Rda", package = "rrricanes"))
 
 ## ---- Dataframe Skeleton -----------------------------------------------------
@@ -30,9 +18,10 @@ test_that("Dataframe Skeleton", {
 ## ---- Test get_discus() ------------------------------------------------------
 #' Test return of get_discus()
 test_that("Test get_discus()", {
+    skip_on_cran()
+
+    al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
+    df.al092008.discus <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_discus()
+
     expect_identical(al092008.discus, df.al092008.discus)
 })
-
-# Reset options
-options("rrricanes.http_timeout" = opt.timeout)
-options("rrricanes.http_attempts" = opt.attempts)
