### 2017-06-23 4d8889e
Modify data for vignettes; spell check; manuals
* Corrected data files for vignettes; used devtools::use_data

* Add documentation for vignette datasets

* Spell checks
