### 2017-06-23 3b768d7
bump version; add additional changes
