### 2017-06-23 #118 @timtrice
* Vignettes have been modified to use /data
* data_world pkgdown article moved to vignettes/articles
* Examples added to previously mentioned functions
* Add Maintainer to DESCRIPTION
* Removed `Additional_repositories` field in DESCRIPTION
* Slimmed down tests some and added `skip_on_cran` to all scraping calls.

I have not added the parameter to tidy the return of `get_fstadv` as discussed previously; would prefer to keep it consistent with the others. Is this necessary?

All local checks and tests pass. Travis and Appveyor passed on branch tests; currently running master and release. Any failures will be due to timeout; I'll restart in the morning if that occurs.
