### 2017-06-23 #118 @maelle
Thanks a lot @jsta for your review! :smile: and nice work @timtrice !

Reg. the maintainer field you don't need to have it in DESCRIPTION, e.g. in `ropenaq` this is how my [DESCRIPTION](https://github.com/ropensci/ropenaq/blob/master/DESCRIPTION) looks like but when I build it the maintainer field is automatically added so on CRAN it's [like this](https://cran.r-project.org/web/packages/ropenaq/index.html).
