### 2017-06-23 #118 @jsta
OK, I will wait for the review by @robinsones before I look at the changes.
