### 2017-06-24 339cd89
Reapply PR #79
