+++ data-raw/audit.R
@@ -1,0 +1,253 @@
+#' ########################################################################## '#
+#' Audit
+#'
+#' Examine all datasets for quality issues. This will primarily be beneficial
+#' during cleanup phases but should probably be operated on a weekly basis to
+#' ensure integrity.
+#'
+#' ########################################################################## '#
+
+## ---- Libraries --------------------------------------------------------------
+library(dplyr)
+library(Hmisc)
+library(purrr)
+library(readr)
+library(rrricanes)
+
+## ---- Options ----------------------------------------------------------------
+
+## ---- Variables --------------------------------------------------------------
+valid_options <- c("Hurricane", "Post-Tropical Cyclone",
+                   "Potential Tropical Cyclone", "Remnants of",
+                   "Subtropical Depression", "Tropical Depression",
+                   "Tropical Storm")
+
+invalid_names <- c("Center", "Test")
+
+#' ########################################################################## '#
+## ---- Storm Discussions ------------------------------------------------------
+#' The `discus` product should only parse `status`, `name`, `adv`, `date` and
+#' `contents` of the product.
+discus <- read_csv("./datasets/discus.csv", col_types = cols())
+
+## ---- * Status ---------------------------------------------------------------
+status <- sort(unique(discus$Status))
+if (!(any(status %in% valid_options)))
+    warning(sprintf("Invalid Status in discus:\n%s",
+                    sort(unique(discus$Status[!(discus$Status %in% valid_options)]))))
+
+## ---- * Names ----------------------------------------------------------------
+#' * discus$Name for EP192000 Adv 1 is "Test". This is an err on the NHC's part.
+discus_names <- sort(unique(discus$Name))
+if (any(discus_names %in% invalid_names)) {
+    warning("Invalid Names in discus. See object df_discus_invalid_names.")
+    df_discus_invalid_names <- discus %>% filter(Name %in% invalid_names)
+}
+
+## ---- * Adv ------------------------------------------------------------------
+describe(discus$Adv)
+
+## ---- * Date -----------------------------------------------------------------
+describe(discus$Date)
+
+#' ########################################################################## '#
+## ---- Forecast/Advisory ------------------------------------------------------
+fstadv <- read_csv("./datasets/fstadv.csv", col_types = cols(),
+                   guess_max = 13000)
+
+## ---- * Status ---------------------------------------------------------------
+status <- sort(unique(fstadv$Status))
+if (!(any(status %in% valid_options)))
+    warning(sprintf("Invalid Status in fstadv:\n%s",
+                    sort(unique(fstadv$Status[!(fstadv$Status %in% valid_options)]))))
+
+## ---- * Names ----------------------------------------------------------------
+fstadv_names <- sort(unique(fstadv$Name))
+if (any(fstadv_names %in% invalid_names)) {
+    warning("Invalid Names in fstadv. See object df_fstadv_invalid_names.")
+    df_fstadv_invalid_names <- fstadv %>% filter(Name %in% invalid_names)
+}
+
+## ---- * Date -----------------------------------------------------------------
+describe(fstadv$Date)
+
+## ---- * Key ------------------------------------------------------------------
+if (any(grepl("^[AL|EP][[:digit:]]{6}$", fstadv$Key)))
+    warning("Invalid Keys")
+
+## ---- * Lat ------------------------------------------------------------------
+#' There may be NA values in Lat if a system is no longer tropical or has a
+#' closed center.
+fstadv_lat <- select(fstadv, dplyr::contains("Lat"))
+
[…]
