### 2017-06-24 b597648
Add var Key to `discus` dataset (#80)
* Added variable `key` to `discus` dataframe. `key` was added to storm
discussions beginning 2006 season. For cyclones prior, `key` will be
NA.

* Updated test datasets. All tests pass.

* All checks pass.
