### 2017-06-24 5e20659
Remove `Adv` from `posest` (#81)
* Removed var `Adv` from df `posest`. Position estimates do not have
advisory numbers.

* Updated tests and test dataframes.

* All tests pass.

* All checks pass.
