+++ tests/testthat/test_posest.R
@@ -10,7 +10,6 @@ test_that("Dataframe Skeleton", {
     expect_true(tibble::is_tibble(df))
     expect_identical(class(df$Status), "character")
     expect_identical(class(df$Name), "character")
-    expect_identical(class(df$Adv), "character")
     expect_identical(class(df$Date), c("POSIXct", "POSIXt"))
     expect_identical(class(df$Contents), "character")
 })
