+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Changed
   - Added variable `Key` to `discus` dataframes. (#80)
+  - Removed variable `Adv` from `posest`. Position estimates do not have advisory numbers. (#81)
 
 ### Removed
   - NA
