### 2017-06-24 5794344
Remove geosphere from Suggests
