### 2017-06-24 28e4ea5
Allow for "INTERMEDIATE" in some public advs (#83)
* Fix `scrape_adv_num` to accomodate possible "INTERMEDIATE" text in
Public Advisory headers.

* All tests and checks pass.
