+++ R/scrapers.R
@@ -4,7 +4,8 @@
 #' @seealso \code{\link{scrape_header}}
 #' @keywords internal
 scrape_adv_num <- function(header) {
-    ptn <- paste0("(?:ADVISORY|DISCUSSION|PROBABILITIES)[:blank:]+NUMBER",
+    ptn <- paste0("(?:ADVISORY|DISCUSSION|PROBABILITIES)*[:blank:]+",
+                  "(?:INTERMEDIATE[:blank:])*NUMBER",
                   # Advisory number. Could alphanum; i.e., 1, 1A, 2, 2A, 2B
                   "[:blank:]+([:digit:]{1,3}[:alpha:]*)",
                   "(?:[[:space:][:punct:][:alpha:]]*)+")
