+++ man/get_update.Rd
@@ -15,7 +15,6 @@ Return dataframe of cyclone update data.
   \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
     etc.}
   \item{Name}{Name of storm}
-  \item{Adv}{Advisory Number}
   \item{Date}{Date of advisory issuance}
   \item{Contents}{Text content of product}
 }
