+++ NEWS.md
@@ -9,6 +9,7 @@ rrricanes 0.2.0-4 (yyyy-mm-dd)
 
 * `Key` variable added to `discus` dataframes. `Key` will be NA for all cyclones >= 2005. Should not be <= 2006. (#80)
 * Removed `Adv` variable from `posest` dataframes. Position estimates do not have advisory numbers. (#81)
+* Removed `Adv` variable from `update`. Updates do not have advisory numbers. (#84)
 
 ### BUG FIXES
 
