### 2017-06-24 226e685
Added variable `Key` to `get_public` df (#85)
* Added variable `Key`. Will be NA for cyclones prior to 2006.

* Updated tests.

* All tests and checks pass.
