+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   - Removed variable `Adv` from `posest`. Position estimates do not have advisory numbers. (#81)
   - Fix `scrape_adv_num` to accomodate possible "INTERMEDIATE" text in Public Advisory headers. (#83)
   - Remove variable `Adv` from `update`. Updates do not have advisory numbers. (#84)
+  - Added variable `Key` to `get_public` dataframes. (#85)
 
 ### Removed
   - NA
