### 2017-06-24 4179bf1
Added variable `Key` to `get_update` df (#86)
* Updated tests

* All tests and checks pass
