+++ R/fstadv.R
@@ -1,3 +1,129 @@
+#' @title create_df_fstadv
+#' @description Template for forecast/advisories dataframe
+#' @return empty dataframe
+#' @seealso \code{\link{get_fstadv}}
+#' @keywords internal
+create_df_fstadv <- function() {
+    df <- tibble::data_frame(Status = character(),
+                             Name = character(),
+                             Adv = integer(),
+                             Date = as.POSIXct(character(), tz = "UTC"),
+                             Key = character(),
+                             Lat = numeric(),
+                             Lon = numeric(),
+                             Wind = integer(),
+                             Gust = integer(),
+                             Pressure = integer(),
+                             PosAcc = integer(),
+                             FwdDir = integer(),
+                             FwdSpeed = integer(),
+                             Eye = integer(),
+                             SeasNE = integer(),
+                             SeasSE = integer(),
+                             SeasSW = integer(),
+                             SeasNW = integer(),
+                             NE64 = integer(),
+                             SE64 = integer(),
+                             SW64 = integer(),
+                             NW64 = integer(),
+                             NE50 = integer(),
+                             SE50 = integer(),
+                             SW50 = integer(),
+                             NW50 = integer(),
+                             NE34 = integer(),
+                             SE34 = integer(),
+                             SW34 = integer(),
+                             NW34 = integer(),
+                             Hr12FcstDate = as.POSIXct(character(), tz = "UTC"),
+                             Hr12Lat = numeric(),
+                             Hr12Lon = numeric(),
+                             Hr12Wind = integer(),
+                             Hr12Gust = integer(),
+                             Hr12NE64 = integer(),
+                             Hr12SE64 = integer(),
+                             Hr12SW64 = integer(),
+                             Hr12NW64 = integer(),
+                             Hr12NE50 = integer(),
+                             Hr12SE50 = integer(),
+                             Hr12SW50 = integer(),
+                             Hr12NW50 = integer(),
+                             Hr12NE34 = integer(),
+                             Hr12SE34 = integer(),
+                             Hr12SW34 = integer(),
+                             Hr12NW34 = integer(),
+                             Hr24FcstDate = as.POSIXct(character(), tz = "UTC"),
+                             Hr24Lat = numeric(),
+                             Hr24Lon = numeric(),
+                             Hr24Wind = integer(),
+                             Hr24Gust = integer(),
+                             Hr24NE64 = integer(),
+                             Hr24SE64 = integer(),
+                             Hr24SW64 = integer(),
+                             Hr24NW64 = integer(),
+                             Hr24NE50 = integer(),
+                             Hr24SE50 = integer(),
+                             Hr24SW50 = integer(),
+                             Hr24NW50 = integer(),
+                             Hr24NE34 = integer(),
+                             Hr24SE34 = integer(),
+                             Hr24SW34 = integer(),
+                             Hr24NW34 = integer(),
+                             Hr36FcstDate = as.POSIXct(character(), tz = "UTC"),
+                             Hr36Lat = numeric(),
[…]
