### 2017-06-25 f1e0afd
Update tidy_fcst_wr per issue #89
