+++ R/fstadv.R
@@ -271,11 +271,11 @@ fstadv <- function(link, p = dplyr::progress_estimated(n = 1)) {
 
     df <- df %>%
         tibble::add_row("Status" = status, "Name" = name, "Adv" = adv,
-                             "Date" = date, "Key" = key, "Lat" = lat,
-                             "Lon" = lon, "Wind" = wind, "Gust" = gust,
-                             "Pressure" = pressure, "PosAcc" = posacc,
-                             "FwdDir" = fwd_dir, "FwdSpeed" = fwd_speed,
-                             "Eye" = eye)
+                        "Date" = date, "Key" = key, "Lat" = lat,
+                        "Lon" = lon, "Wind" = wind, "Gust" = gust,
+                        "Pressure" = pressure, "PosAcc" = posacc,
+                        "FwdDir" = fwd_dir, "FwdSpeed" = fwd_speed,
+                        "Eye" = eye)
 
     # Add current wind radius
     wind_radius <- fstadv_wind_radius(contents, wind)
@@ -900,98 +900,49 @@ tidy_fcst_wr <- function(df) {
     if (!is.data.frame(df))
         stop("Expecting a dataframe.")
 
-    df <- df %>%
-        dplyr::select(Key, Adv, Date,
-                      dplyr::matches("Hr\\d{2,3}FcstDate"),
-                      dplyr::matches("Hr\\d{2,3}\\w{2}\\d{2,3}")) %>%
-        dplyr::select(-dplyr::starts_with("Hr96"), -dplyr::starts_with("Hr120"))
+    # Build wind radius dataframe for each forecast position (12:72 hours; 96
+    # and 120 hours are never forecasted). This dataframe will be similar to
+    # fstadv.wr with the exception of FcstDate.
 
-    # This must be simplified. I'm currently in a rush so just trying to get
-    # this to work while matching previous results. Because the WindField values
-    # will not exist for each forecast period (and non-existent for 96, 120
-    # hours) we have to tidy the dataframe a bit differently than the previous
-    # version.
-    hr12 <- dplyr::select(df, Key, Adv, Date, dplyr::starts_with("Hr12")) %>%
-        dplyr::rename(FcstDate = Hr12FcstDate,
-                      NE64 = Hr12NE64,
-                      SE64 = Hr12SE64,
-                      SW64 = Hr12SW64,
-                      NW64 = Hr12NW64,
-                      NE50 = Hr12NE50,
-                      SE50 = Hr12SE50,
-                      SW50 = Hr12SW50,
-                      NW50 = Hr12NW50,
-                      NE34 = Hr12NE34,
-                      SE34 = Hr12SE34,
-                      SW34 = Hr12SW34,
-                      NW34 = Hr12NW34)
+    v <- c("NE", "SE", "SW", "NW")
 
-    hr24 <- dplyr::select(df, Key, Adv, Date, dplyr::starts_with("Hr24")) %>%
-        dplyr::rename(FcstDate = Hr24FcstDate,
-                      NE64 = Hr24NE64,
-                      SE64 = Hr24SE64,
-                      SW64 = Hr24SW64,
-                      NW64 = Hr24NW64,
-                      NE50 = Hr24NE50,
-                      SE50 = Hr24SE50,
-                      SW50 = Hr24SW50,
-                      NW50 = Hr24NW50,
-                      NE34 = Hr24NE34,
-                      SE34 = Hr24SE34,
-                      SW34 = Hr24SW34,
-                      NW34 = Hr24NW34)
+    # What forecast periods are in the current dataset?
+    fcst_periods <- as.list(names(df)) %>%
[…]
