+++ NEWS.md
@@ -1,4 +1,4 @@
-rrricanes 0.2.0-4 (yyyy-mm-dd)
+rrricanes 0.2.0-4 (2017-06-25)
 ==================================
 
 ### NEW FEATURES
