### 2017-06-26 3df6787
Add script to parse all products
This brute-force script is designed to be used to rewrite all product
datasets whenever a major fix to any of the product functions has
occurred. For example, if a bug keeps some storms from having their
proper names saved and that big is resolved, this script should be run.
