### 2017-06-26 133f873
Modified vignettes for func shp_to_df
