+++ vignettes/gis_data.Rmd
@@ -27,13 +27,19 @@ Some datasets require the use of the storm key and an optional advisory. Other p
 
 There are three functions available to help you plot GIS data; `tracking_chart`, `al_tracking_chart` and `ep_tracking_chart`. `al_tracking_chart` and `ep_tracking_chart` are just helpers centered on the Atlantic and northeast Pacific ocean, respectively. 
 
-Building tracking charts requires the `rnaturalearthdata` and, for high resolution charts, `rnaturalearthhires` packages.
-
 ```{r}
 args(rrricanes::tracking_chart)
 ```
 
-The countries and states parameters are TRUE by default. This means a basic call to `tracking_chart` will return a map with country and state borders. The basic call does not show this.
+The `countries` and `states` parameters are TRUE by default. This means a basic call to `tracking_chart` will return a map with country and state borders. The `res` parameter is resolution; one of 110, 50 or 10 nautical miles. Resolutions 110nm and 50nm can be used immediately. To use lower resolution you must install the `rnaturalearthdatahires` package from ropensci:
+
+```{r, eval = FALSE}
+install.packages("rnaturalearthhires",
+                 repos = "http://packages.ropensci.org",
+                 type = "source")
+```
+
+`tracking_chart` will print a basic tracking chart (a map of the planet).
 
 ```{r, fig.width = 7, fig.align = "center"}
 tracking_chart()
@@ -93,56 +99,79 @@ For this particular storm and advisory, included are the base line, point and po
 str(adv$al182012_018_5day_lin)
 ```
 
+As we're dealing with a SpatialLinesDataFrame which needs to be modified, you can use the helper function `shp_to_df` for plotting.
+
 ```{r}
-fcst_line <- as(adv$al182012_018_5day_lin, "SpatialPointsDataFrame") %>%
-    broom::tidy()
+fcst_line <- shp_to_df(adv$al182012_018_5day_lin)
 ```
 
 ```{r, fig.width = 7, fig.align = "center"}
-p + geom_path(data = fcst_line, aes(coords.x1, coords.x2, group = FCSTPRD))
+p + geom_path(data = fcst_line, aes(long, lat, group = FCSTPRD))
+```
+
+There are two groups of data in the set: one for 72-hour forecast period and one for 120-hour. Grouping by `FCSTPRD` will show the forecast track correctly.
+
+There is a pretty wide field of view on the map above. You can use `sp::bbox` to "zoom in" on the map.
+
+```{r}
+(bb <- sp::bbox(adv$al182012_018_5day_lin))
+```
+
+```{r, fig.width = 7, fig.align = "center"}
+(p2 <- p + geom_path(data = fcst_line, aes(long, lat, group = FCSTPRD)) + 
+    coord_equal(xlim = c(bb[1,1] - 5, bb[1,2] + 5), 
+                ylim = c(bb[2,1] - 5, bb[2,2] + 5)))
 ```
 
 ### Point track
 
+Each forecast position is also included in the points dataframe. You can access this in the `adv$al182012_018_5day_pts@data` object; no conversion necessary.
+
 ```{r, fig.width = 7, fig.align = "center"}
-p + geom_point(data = adv$al182012_018_5day_pts@data, aes(LON, LAT))
+(p3 <- p2 + geom_point(data = adv$al182012_018_5day_pts@data, aes(LON, LAT)))
 ```
 
 ## Forecast Cone
 
[…]
