### 2017-06-27 14ca347
Force Adv as numeric
