+++ vignettes/articles/hurricane_ike_adv_32.Rmd
@@ -1,0 +1,354 @@
+---
+title: "Hurricane Ike, Adv #32"
+author: "Tim Trice"
+date: "June 11, 2017"
+output: html_document
+references:
+- id: sshws
+  title: The Saffir-Simpson Hurricane Wind Scale
+  author:
+  - family: Schott
+    given: Timothy
+  - family: Landsea
+    given: Chris
+  - family: Hafele
+    given: Gene
+  - family: Lorens
+    given: Jeffrey
+  - family: Taylor
+    given: Arthur
+  - family: Thurm
+    given: Harvey
+  - family: Ward
+    given: Bill
+  - family: Willis
+    given: Mark
+  - family: Zaleski
+    given: WAlt
+  URL: 'http://www.nhc.noaa.gov/pdf/sshws.pdf'
+  page: 2
+  issued:
+    year: 2012
+    month: 2
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE, 
+                      fig.width = 10, 
+                      fig.asp = 1, 
+                      fig.align = "center")
+```
+
+```{r, message = FALSE}
+library(dplyr)
+library(gganimate)
+library(ggplot2)
+library(HURDAT)
+library(purrr)
+library(rrricanes)
+library(tibble)
+library(tidyr)
+library(sp)
+```
+
+## Obtaining Data
+
+```{r}
+key <- "AL092008"
+adv <- 42
+```
+
+### Forecast/Advisory Data for Hurricane Ike
+
+```{r}
+# No need to do this piece for active storms
+fstadv <- load_storm_data("fstadv") %>% filter(Key == key, Adv <= adv)
+```
+
+### GIS Advisory Forecast Track, Cone of Uncertainty, and Watches/Warnings
+
+```{r}
+gis_adv <- gis_advisory(key = key, advisory = adv) %>% gis_download()
+```
+
+Get bounding box of the forecast polygon.
+
+```{r}
+bbox <- bbox(gis_adv$al092008_042_5day_pgn)
+```
+
+### Probabilistic Storm Surge
+
+Calculate the last date/time of forecast/advisory product and subtract 3 hrs
+
+```{r}
+dt <- (last(fstadv$Date) - (60 * 60 * 3)) %>% 
+    strftime(format = "%Y%m%d%H", tz = "UTC")
+```
+
+Wrap `gis_download` in `safely`. Not all products will exist for every storm or even every advisory. 
+
+```{r}
+dl <- purrr::safely(.f = gis_download)
+gis_surge <- gis_prob_storm_surge(key, products = list("psurge" = c(5)), 
+                                  datetime = dt) %>% dl()
+
+if (!is.null(gis_surge$error))
+    message(gis_surge$error)
+```
+
+## Build a Tracking Chart
+
+Generate a base plot of the Atlantic ocean.
+
+```{r}
+bp <- al_tracking_chart(color = "black", fill = "white", size = 0.1, res = 50)
+```
+
+I like to add a little cushion for the map inset and forecast cone data. 
+
+```{r}
+lat_min <- bbox[2,1] - 5
+lat_max <- bbox[2,2] + 5
+lon_min <- bbox[1,1] - 10
+lon_max <- bbox[1,2] + 10
+```
+
+Build a thin tracking map for the inset.
+
+```{r}
+bp_inset <- ggplotGrob(bp +
+    geom_rect(mapping = aes(xmin = lon_min, xmax = lon_max,
+                            ymin = lat_min, ymax = lat_max),
+              color = "red", alpha = 0) +
+    theme_bw() +
+    theme(axis.title = element_blank(),
+          axis.ticks = element_blank(),
[…]
