### 2017-06-28 5b1679a
Allow readr params to load_storm_data
