+++ vignettes/articles/wind_pressure.Rmd
@@ -1,0 +1,137 @@
+---
+title: "Wind and Pressure"
+author: "Tim Trice"
+date: "June 14, 2017"
+output: html_document
+references:
+- id: sshws
+  title: The Saffir-Simpson Hurricane Wind Scale
+  author:
+  - family: Schott
+    given: Timothy
+  - family: Landsea
+    given: Chris
+  - family: Hafele
+    given: Gene
+  - family: Lorens
+    given: Jeffrey
+  - family: Taylor
+    given: Arthur
+  - family: Thurm
+    given: Harvey
+  - family: Ward
+    given: Bill
+  - family: Willis
+    given: Mark
+  - family: Zaleski
+    given: Walt
+  URL: 'http://www.nhc.noaa.gov/pdf/sshws.pdf'
+  issued:
+    year: 2012
+    month: 2
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE, 
+                      fig.align = "center")
+```
+
+```{r, message = FALSE}
+library(dplyr)
+library(gganimate)
+library(ggplot2)
+library(rrricanes)
+library(tidyr)
+```
+
+```{r}
+key <- "AL092008"
+adv <- 42
+```
+
+### Forecast/Advisory Data for Hurricane Ike Adv #42
+
+```{r}
+# No need to do this piece for active storms
+fstadv <- load_storm_data("fstadv") %>% filter(Key == key, Adv <= adv)
+```
+
+## Wind Profile
+
+```{r}
+# Plot wind values
+fstadv %>% ggplot(aes(x = Date, y = Wind)) + 
+    geom_line() + 
+    geom_point(aes(color = Status), size = 3) + 
+    scale_y_continuous(name = "Wind (kts)") + 
+    theme_bw() + 
+    theme(legend.position = "bottom", 
+          legend.box = "vertical") +
+    labs(title = "Wind Profile", 
+         caption = sprintf("rrricanes %s", packageVersion("rrricanes")))
+```
+
+## Pressure Profile
+
+```{r}
+# Plot pressure values
+fstadv %>% ggplot(aes(x = Date, y = Pressure)) + 
+    geom_line() + 
+    geom_point(aes(color = Status), size = 3) + 
+    scale_y_continuous(name = "Pressure (mb)") + 
+    theme_bw() + 
+    theme(legend.position = "bottom", 
+          legend.box = "vertical") +
+    labs(title = "Pressure Profile", 
+         caption = sprintf("rrricanes %s", packageVersion("rrricanes")))
+```
+
+## Wind/Pressure Relational Change
+
+```{r}
+fstadv %>%
+    mutate(WindDist = (Wind - min(Wind))/(max(Wind) - min(Wind)),
+           PressDist = (Pressure - max(Pressure))/(max(Pressure) - min(Pressure))) %>%
+    gather(Var, Val, WindDist, PressDist) %>% 
+    ggplot(aes(x = Date, y = Val, group = Var, color = Var)) + 
+    geom_line(size = 1) + 
+    scale_color_discrete(labels = c("Pressure Change", "Wind Change")) + 
+    theme_bw() + 
+    theme(legend.position = "bottom", 
+          legend.title = element_blank()) + 
+    labs(title = "Wind/Pressure Relational Change", 
+         subtitle = "",
+         caption = sprintf("rrricanes %s", packageVersion("rrricanes")), 
+         y = "")
+```
+
+## Wind Radius
+
+```{r}
+wr_animate <- tidy_wr(fstadv) %>% 
+    gather(Quadrant, Radius, NE:NW) %>% 
+    ggplot(aes(x = Quadrant, y = Radius, fill = factor(WindField), 
[…]
