### 2017-06-28 6c66e89
Correct version in NEWS
