### 2017-06-28 317bd57
brute force script to update active storms
