### 2017-06-29 b50cbff
Correct vignette param
