+++ README.Rmd
@@ -28,7 +28,7 @@ Data can be obtained for cyclones in the north Atlantic (considered the Atlantic
 Please view the vignette 'Getting Started':
 
 ```r
-vignette("getting-started", package = "rrricanes")
+vignette("getting_started", package = "rrricanes")
 ```
 
 [Online documentation](https://timtrice.github.io/rrricanes/) is also available.
