### 2017-06-29 8db43f6
Add build_vignettes param
