+++ README.Rmd
@@ -50,7 +50,7 @@ install.packages("rnaturalearthhires",
 `rrricanes` is currently only available in GitHub. It can be installed using the `devtools` package:
 
 ```r
-devtools::install_github("timtrice/rrricanes")
+devtools::install_github("timtrice/rrricanes", build_vignettes = TRUE)
 ```
 
 ## Built With
