### 2017-06-30 ce8d45d
Assign col_types if empty
