+++ R/get_storm_data.R
@@ -142,6 +142,24 @@ load_storm_data <- function(dataset = c("adv", "discus", "fcst", "fcst_wr",
     dataset <- match.arg(dataset)
     base_url <- "https://github.com/timtrice/rrricanesdata/blob/master/"
     link <- paste0(base_url, dataset, ".csv?raw=true")
-    df <- readr::read_csv(link, ...)
+    readr_args <- list(...)
+    if (purrr::is_empty(readr_args$col_types)) {
+        if (dataset == "discus") {
+            col_types = "cciTcc"
+        } else if (dataset == "fstadv") {
+            col_types = paste0("cciTcddiidiiiiiiiiiiiiiiiiiiiiTddiiiiiiiiiiiiii",
+                               "TddiiiiiiiiiiiiiiTddiiiiiiiiiiiiiiTddiiiiiiiiii",
+                               "TddiiiiiiiiiiTddiiTddii")
+        } else if (dataset == "public") {
+            col_types = "cccTcc"
+        } else if (dataset == "update") {
+            col_types = "ccTcc"
+        } else if (dataset == "wndprb") {
+            col_types = "ciTciiiiiiiiiiiiii"
+        }
+        df <- readr::read_csv(link, col_types = col_types, ...)
+    } else {
+        df <- readr::read_csv(link, ...)
+    }
     return(df)
 }
