### 2017-07-01 f50f2e1
Remove col class declaration
