### 2017-07-01 4822844
Remove dwapi from Suggests
