+++ DESCRIPTION
@@ -34,9 +34,8 @@ Imports:
   xml2 (>= 1.1.1)
 Suggests:
   devtools, 
-  dwapi,
   knitr, 
   rmarkdown,
-  testthat,
+  testthat
 RoxygenNote: 6.0.1
 VignetteBuilder: knitr
