### 2017-07-01 14441a7
Add col_types fallback
