### 2017-07-01 bf336f8
Update package dependencies
