+++ DESCRIPTION
@@ -15,22 +15,22 @@ License: MIT + file LICENSE
 LazyData: TRUE
 Imports:
   broom (>= 0.4.2), 
-  dplyr (>= 0.5.0), 
+  dplyr (>= 0.7.1), 
   ggplot2 (>= 2.2.1), 
   httr (>= 1.2.1), 
   lubridate (>= 1.6.0),
   magrittr (>= 1.5), 
   maptools (>= 0.9-2), 
   purrr (>= 0.2.2.2), 
-  readr (>= 1.1.0), 
-  rgdal (>= 1.2-7),
+  readr (>= 1.1.1), 
+  rgdal (>= 1.2-8),
   rgeos (>= 0.3-23),
   rlang (>= 0.1.1),
   rnaturalearthdata (>= 0.1.0),
   rvest (>= 0.3.2), 
   stringr (>= 1.2.0), 
-  tibble (>= 1.3.1), 
-  tidyr (>= 0.6.2), 
+  tibble (>= 1.3.3), 
+  tidyr (>= 0.6.3), 
   xml2 (>= 1.1.1)
 Suggests:
   devtools, 
