### 2017-07-01 3fffb8b
Drop rgdal to v1.2-7
