+++ DESCRIPTION
@@ -6,7 +6,6 @@ Description: Get archived data of past and current hurricanes and tropical
   storms since 1998.
 Version: 0.2.0-5
 Authors@R: c(person("Tim", "Trice", email = "tim.trice@gmail.com", role = c("aut", "cre")))
-Maintainer: Tim Trice <tim.trice@gmail.com>
 Depends:
   R (>= 3.3.3)
 URL: https://github.com/timtrice/rrricanes
