+++ DESCRIPTION
@@ -12,6 +12,7 @@ URL: https://github.com/timtrice/rrricanes
 BugReports: https://github.com/timtrice/rrricanes/issues
 License: MIT + file LICENSE
 LazyData: TRUE
+ByteCompile: TRUE
 Imports:
   broom (>= 0.4.2), 
   dplyr (>= 0.7.1), 
