### 2017-07-10 #118 @robinsones
Hi @timtrice, your review is currently in progress! I'm running into one issue though. I see you've updated the README to have `build_vignettes = TRUE`, which helps one problem I was having yesterday, where the vignettes weren't available. However, now when I update the install with that, any function I try to get the help page for (e.g. `?tidy_wr`) results in `Error in fetch(key): lazy-load database./Library/Frameworks/R.framework/Versions/3.4/Resources/library/rrricanes/help/rrricanes.rdb' is corrupt`. This was not happening yesterday, and I also replicated that it didn't happen on another computer when I did `install_github("timtrice/rrricanes"). It only happened when I added the `build_vignettes = TRUE` and reinstalled.
