### 2017-07-10 #118 @timtrice
Hi Emily, thank you for taking a look at it!

When you say you updated the package, did you remove the package then reinstall? I just tested on Debian and Win, R 3.3.3 and 3.4.0 and could not replicate the error. All vignettes appear correctly and I'm unable to find any issues with the help documentation.

If you can post your `sessionInfo` that may help me out.

Thank you!

Tim
