### 2017-07-10 #118 @robinsones
Hi Tim,

What I  did when I say update was `devtools::install_github("timtrice/rrricanes", build_vignettes = TRUE, force = TRUE)`. I also tried removing the package and running `devtools::install_github("timtrice/rrricanes")`, but I continued to get the same error. I can try checking on a colleague's computer tomorrow they also have the same issue.

Even when I uninstall the packages with `remove.packages`, then redownload it without the `build_vignettes` function, I continue to get the same error.
Here is my session info below:

R version 3.4.0 (2017-04-21)
Platform: x86_64-apple-darwin15.6.0 (64-bit)
Running under: OS X El Capitan 10.11.6

Matrix products: default
BLAS: /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/3.4/Resources/lib/libRlapack.dylib

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

attached base packages:

other attached packages:
[1] rrricanes_0.2.0-5 bindrcpp_0.2      ggplot2_2.2.1     dplyr_0.7.1

And from `devtools::sesssion_info()`:

Session info -------------------------------------------------------------------------------------------------
 setting  value
 version  R version 3.4.0 (2017-04-21)
 system   x86_64, darwin15.6.0
 ui       RStudio (1.0.143)
 language (EN)
 collate  en_US.UTF-8
 tz       America/New_York
 date     2017-07-09

 rrricanes         * 0.2.0-5 2017-07-10 Github (timtrice/rrricanes@e99a73e)
