### 2017-07-10 #118 @timtrice
This may be related to a previous issue with devtools (resolved in 1.10.0);

* [Unregister S3 methods explicitly](https://github.com/hadley/devtools/pull/1001)
* [[Rd] unloadNamespace() does not address unevaluated promises in the S3 Methods Table](https://stat.ethz.ch/pipermail/r-devel/2015-December/072150.html)

I'll look into it a bit more and see if there's been anything more recent.
