### 2017-07-10 #118 @jsta
@robinsones I'm not sure if this is your situation, but I recall seeing these errors when I install (and load) a package and then try to build and reload that same package "in place" (without restarting RStudio). In this case, restarting RStudio before (re)installing fixes the issue.

I did my review by git cloning the repo and installing locally from source. I'm glad you are checking `install_github`
