### 2017-07-10 #118 @robinsones
HI @jsta thank you! That seems to be exactly it. Sorry for the alarm @timtrice. When I restart R everything works fine. It also works fine if I remove rrricanes, restart R, then reinstall rrricanes, so no one else should have any problems.
