### 2017-07-10 #118 @timtrice
Glad to hear it! I was just trying again on my Win machine doing what @jsta was mentioning and now my help won't load (but no error message). Which is funny because it's done this to me before while developing but I figured it was Rstudio or packrat. Now we learned something new!
