### 2017-07-10 #118 @robinsones
- I think could you include more information in the README, especially about the source of the data. While I know from the vignette and this review process the data comes from the National Hurricane Center's (NHC) archives, it would be helpful to have that upfront. A link to the archive would also be welcome. This could be paired with a short sentence about how this package provides information previously unavailable to people unfamiliar with web-scraping and also functions that make understanding the data easier.

Estimated hours spent reviewing: 4 hours.

---

Really great work! I am impressed how comprehensive it is.
