### 2017-07-10 e4361b6
Add Linux instructions for supporting pkgs (#95)
