### 2017-07-10 3a5846f
Added additional info to README
