+++ README.Rmd
@@ -19,9 +19,39 @@ knitr::opts_chunk$set(
 
 # rrricanes
 
-`rrricanes` is a R library that extracts information from available archives on past and current tropical cyclones. Currently, archives date back to 1998. 
+`rrricanes` is a R library that extracts information from [available archives](http://www.nhc.noaa.gov/archive/1998/1998archive.shtml) on past and current tropical cyclones. Currently, archives date back to 1998. 
 
-Data can be obtained for cyclones in the north Atlantic (considered the Atlantic Basin) and north-eastern Pacific (the East Pacific Basin from 140&deg;W and eastward, and Central Pacific Basin from 140&deg;W to 180&deg;W). 
+Data can be obtained for cyclones in the north Atlantic (considered the Atlantic Basin) and north-eastern Pacific (the East Pacific Basin from 140&deg;W and eastward. 
+
+Central Pacific data (140&deg;W to 180&deg;W) is included if issued by the [National Hurricane Center](http://www.nhc.noaa.gov/) (generally they're issued by the [Central Pacific Hurricane Center](http://www.prh.noaa.gov/cphc/)).
+
+This library parses the text advisories of all tropical cyclones since 1998. Over the years the formats of the text products have changed and many are loosely formatted. 
+
+I wrote this package with the goal of consolidating messy text data into well-organized formats that can easily be saved to CSV, SQL and other data formats. 
+
+You may explore some features of the package through the [shinycanes](https://timtrice.shinyapps.io/shinycanes/) beta web application (built with R Shiny).
+
+## Advisory Products
+
+Generally speaking, there are five products available for tropical cyclones issued at 03:00, 09:00, 15:00 and 21:00 UTC;  
+
+1. Storm Discussion - These are technical discussions centered on the current structure of the cyclone, satellite presentation, computer forecast model tendencies and more. 
+
+2. Forecast/Adivsory - This data-rich product lists the current location of the cyclone, its wind structure, forecast and forecast wind structure.
+
+3. Public Advisory - These are general text statements issued for the public-at-large. Information in these products is a summary of the Forecast/Advisory product along with any watches and warnings issued, changed, or cancelled. Public Advisory products are the only regularly-scheduled product that may be issued intermittently (every three hours and, occasionally, every two hours) when watches and warnings are in effect.
+
+4. Wind Speed Probabilities - These products list the probability of a minimum sustained wind speed expected in a given forecast window. This product replaces the Strike Probabilities product beginning in 2006 (see below).
+
+5. Updates - Tropical Cyclone Updates may be issued at any time if a storm is an immediate threat to land or if the cyclone undergoes a significant change of strength or structure. The information in this product is general.
+
+__Discontinued Products__
+
+These products are included in the package though they have been discontinued at some point:
+
+1. Strike Probabilities - List the probability of a tropical cyclone passing within 65 nautical miles of a location within a forecast window. Replaced in 2006 by the Wind Speed Probabilities product.
+
+2. Position Estimates - Typically issued as a storm is threatening land but generally rare (see Hurricane Ike 2008, Key AL092008). It is generally just an update of the current location of the cyclone. After the 2011 hurricane season, this product was discontinued; Updates are now issued in their place.
 
 ## Getting Started
 
