### 2017-07-10 #118 @timtrice
Thank you, Emily! I've added some additional info to the README (a bit more than you were asking, probably). I haven't pushed it to master yet; if it fits what you were thinking of let me know and I'll sync it up with master.

https://github.com/timtrice/rrricanes/tree/hotfix/v0.2.0-5.1
