### 2017-07-10 #118 @jsta
Ok, the changes look good :+1: . I ran `devtools::check` one last time and saw:

* a NOTE about not having the `sp` package listed in the DESCRIPTION file under `Suggests`.

* The test titled "_Test get_prblty()_" is timing out for me during download. I see that you have it skipped on CRAN so maybe this is OK.
