### 2017-07-10 fdb21f2
Add pkg sp to Suggests
