+++ DESCRIPTION
@@ -36,6 +36,7 @@ Suggests:
   devtools, 
   knitr, 
   rmarkdown,
+  sp,
   testthat
 RoxygenNote: 6.0.1
 VignetteBuilder: knitr
