### 2017-07-11 7279940
Correct package version requirements
