+++ DESCRIPTION
@@ -15,22 +15,22 @@ LazyData: TRUE
 ByteCompile: TRUE
 Imports:
   broom (>= 0.4.2), 
-  dplyr (>= 0.7.1), 
+  dplyr (>= 0.5.0), 
   ggplot2 (>= 2.2.1), 
   httr (>= 1.2.1), 
   lubridate (>= 1.6.0),
   magrittr (>= 1.5), 
   maptools (>= 0.9-2), 
-  purrr (>= 0.2.2.2), 
-  readr (>= 1.1.1), 
-  rgdal (>= 1.2-7),
+  purrr (>= 0.2.2), 
+  readr (>= 1.1.0), 
+  rgdal (>= 1.2-6),
   rgeos (>= 0.3-23),
   rlang (>= 0.1.1),
   rnaturalearthdata (>= 0.1.0),
   rvest (>= 0.3.2), 
   stringr (>= 1.2.0), 
-  tibble (>= 1.3.3), 
-  tidyr (>= 0.6.3), 
+  tibble (>= 1.3.0), 
+  tidyr (>= 0.6.1), 
   xml2 (>= 1.1.1)
 Suggests:
   devtools, 
