### 2017-07-11 ec0b391
Add test crul functions
