+++ R/crul.R
@@ -1,0 +1,330 @@
+#' @title crul_get_url_contents
+#' @description Get contents from URL
+#' @param link URL to download
+#' @export
+crul_get_url_contents <- function(links, p) {
+  p$pause(0.5)$tick()$print()
+  links <- crul::Async$new(urls = links)
+  res <- links$get()
+  # Check status codes
+  #purrr::map(res, ~.$status_code) %>% purrr::flatten_dbl() %>% unique()
+  return(res)
+}
+
+#' @title crul_get_storms
+#' @description Returns storms and product link.
+#' @param year numeric or vector, four digits (\%Y format)
+#' @param basin One or both of c("AL", "EP")
+#' @export
+crul_get_storms <- function(year = format(Sys.Date(), "%Y"),
+                       basin = c("AL", "EP")) {
+
+  year <- validate_year(year)
+
+  # No archives earlier than 1998 for now
+  if (any(year < 1998))
+    stop('Archives currently only available for 1998 to current year.')
+
+  if (!all(basin %in% c("AL", "EP")))
+    stop("Basin must 'AL' and/or 'EP'")
+
+  links <- purrr::map(year, .f = year_archives_link) %>%
+    purrr::flatten_chr()
+
+  # 1998 is only year with slightly different URL. Modify accordingly
+  links[grep("1998", links)] <- paste0(links[grep("1998", links)],
+                                       "1998archive.shtml")
+
+  # There is an 80-hit/10-second limit to the NHC pages (note Issue #94), or 8
+  # requests/second. The below request will process 4 links every 0.5 seconds.
+  links <- split(links, ceiling(seq_along(links)/4))
+  p <- dplyr::progress_estimated(n = length(links))
+  contents <- purrr::map(links, .f = function(x) {crul_get_url_contents(x, p)}) %>%
+    purrr::flatten()
+
+  storm_df <- purrr::map_df(basin, extract_storms, contents) %>%
+    dplyr::group_by(Year, Basin) %>%
+    dplyr::arrange(Year, Basin) %>%
+    dplyr::ungroup()
+
+  return(storm_df)
+}
+
+#' @title crul_get_products
+#' @description Retrieve a list of one or multiple products from a cyclones
+#' archive page
+#' @param links links to storm archive page
+#' @export
+crul_get_products <- function(links) {
+
+  # Get year
+  year <- extract_year_archive_link(link)
+
+  nhc_url <- get_nhc_link(withTrailingSlash = FALSE)
+
+  # Depending on the year these URL's are formatted various ways.
+  if (year == 1998) {
+    products <- stringr::str_c(nhc_url, '/archive/', year, '/', products)
+  } else {
+    products <- stringr::str_c(nhc_url, products)
+  }
+
+  return(products)
+}
+
+#' @title crul_get_storm_data
+#' @description Retrieve storm data for specified products
+#' @param links Links to a cyclones archive pages
+#' @param products c("discus", "fstadv", "posest", "public", "prblty", "update",
+#' "wndprb")
+#' @export
+crul_get_storm_data <- function(links,
+                                products = c("discus", "fstadv", "posest",
+                                             "public", "prblty", "update",
+                                             "wndprb")) {
+
+  products <- match.arg(products, several.ok = TRUE)
+
+  # There is an 80-hit/10-second limit to the NHC pages (note Issue #94), or 8
+  # requests/second. The below request will process 4 links every 0.5 seconds.
+  links <- split(links, ceiling(seq_along(links)/4))
[…]
