+++ NAMESPACE
@@ -3,6 +3,17 @@
 export(al_prblty_stations)
 export(al_tracking_chart)
 export(cp_prblty_stations)
+export(crul_get_discus)
+export(crul_get_fstadv)
+export(crul_get_posest)
+export(crul_get_prblty)
+export(crul_get_products)
+export(crul_get_public)
+export(crul_get_storm_data)
+export(crul_get_storms)
+export(crul_get_update)
+export(crul_get_url_contents)
+export(crul_get_wndprb)
 export(ep_prblty_stations)
 export(ep_tracking_chart)
 export(get_discus)
