### 2017-07-11 #118 @timtrice
Joseph,

`sp` is added to Suggests (apologies for the oversight).

And all tests are passing in Travis and Appveyor so I would assume what you saw was one of the known hiccups from the NHC.
