+++ DESCRIPTION
@@ -15,6 +15,7 @@ LazyData: TRUE
 ByteCompile: TRUE
 Imports:
   broom (>= 0.4.2), 
+  crul (>= 0.3.4),
   dplyr (>= 0.5.0), 
   ggplot2 (>= 2.2.1), 
   httr (>= 1.2.1), 
