### 2017-07-11 bd9b55c
Add crul funcs documentation
