+++ man/crul_get_products.Rd
@@ -1,0 +1,15 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_get_products}
+\alias{crul_get_products}
+\title{crul_get_products}
+\usage{
+crul_get_products(links)
+}
+\arguments{
+\item{links}{links to storm archive page}
+}
+\description{
+Retrieve a list of one or multiple products from a cyclones
+archive page
+}
