+++ man/crul_get_fstadv.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_get_fstadv}
+\alias{crul_get_fstadv}
+\title{crul_get_fstadv}
+\usage{
+crul_get_fstadv(links)
+}
+\arguments{
+\item{links}{URLs to cyclone archive pages}
+}
