+++ man/crul_fstadv.Rd
@@ -1,0 +1,15 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_fstadv}
+\alias{crul_fstadv}
+\title{crul_fstadv}
+\usage{
+crul_fstadv(contents)
+}
+\arguments{
+\item{contents}{text of product}
+}
+\description{
+Extrapolate data from FORECAST/ADVISORY products.
+}
+\keyword{internal}
