+++ man/crul_get_storm_data.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_get_storm_data}
+\alias{crul_get_storm_data}
+\title{crul_get_storm_data}
+\usage{
+crul_get_storm_data(links, products = c("discus", "fstadv", "posest",
+  "public", "prblty", "update", "wndprb"))
+}
+\arguments{
+\item{links}{Links to a cyclones archive pages}
+
+\item{products}{c("discus", "fstadv", "posest", "public", "prblty", "update",
+"wndprb")}
+}
+\description{
+Retrieve storm data for specified products
+}
