+++ man/crul_get_posest.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_get_posest}
+\alias{crul_get_posest}
+\title{crul_get_posest}
+\usage{
+crul_get_posest(links)
+}
+\arguments{
+\item{links}{URLs to cyclone archive pages}
+}
