+++ man/crul_get_public.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_get_public}
+\alias{crul_get_public}
+\title{crul_get_public}
+\usage{
+crul_get_public(links)
+}
+\arguments{
+\item{links}{URLs to cyclone archive pages}
+}
