### 2017-07-11 bb33a49
Add crul funcs documentation
