+++ man/crul_get_update.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_get_update}
+\alias{crul_get_update}
+\title{crul_get_update}
+\usage{
+crul_get_update(links)
+}
+\arguments{
+\item{links}{URLs to cyclone archive pages}
+}
