+++ man/crul_get_storms.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_get_storms}
+\alias{crul_get_storms}
+\title{crul_get_storms}
+\usage{
+crul_get_storms(year = format(Sys.Date(), "\%Y"), basin = c("AL", "EP"))
+}
+\arguments{
+\item{year}{numeric or vector, four digits (\%Y format)}
+
+\item{basin}{One or both of c("AL", "EP")}
+}
+\description{
+Returns storms and product link.
+}
