+++ man/crul_get_discus.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_get_discus}
+\alias{crul_get_discus}
+\title{crul_get_discus}
+\usage{
+crul_get_discus(links)
+}
+\arguments{
+\item{links}{URLs to cyclone archive pages}
+}
