+++ man/crul_get_prblty.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_get_prblty}
+\alias{crul_get_prblty}
+\title{crul_get_prblty}
+\usage{
+crul_get_prblty(links)
+}
+\arguments{
+\item{links}{URLs to cyclone archive pages}
+}
