+++ man/crul_get_wndprb.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_get_wndprb}
+\alias{crul_get_wndprb}
+\title{crul_get_wndprb}
+\usage{
+crul_get_wndprb(links)
+}
+\arguments{
+\item{links}{URLs to cyclone archive pages}
+}
