+++ man/crul_get_url_contents.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_get_url_contents}
+\alias{crul_get_url_contents}
+\title{crul_get_url_contents}
+\usage{
+crul_get_url_contents(links, p)
+}
+\arguments{
+\item{link}{URL to download}
+}
+\description{
+Get contents from URL
+}
