### 2017-07-11 #118 @maelle
Thanks a lot for your review @robinsones ! 😁

Can I assume you're both happy with the changes @robinsones @jsta ?

@timtrice do you still have the repo with data? might be worth to take a look at other approaches like the idea of having a data package on drat (cf our discussion on Slack), and to check license?
