### 2017-07-11 #118 @timtrice
Yup, so right now I'm implementing Scott's `crul` package on getting the data to begin with. I'm hoping to have that finished tonight. Then I'll be making `rrricanesdata` a package.

I did read over Brooke's document and look over her repo's so I think I have a pretty good idea of the drat concept. And of course if I have any questions I'll post them to slack or twitter.

Regarding the license, the link below seems to be the one related to the hurricane data. The NHC itself does not list any data restrictions nor does the National Oceanic and Atmospheric Administration (NHC's parent body).

https://www.weather.gov/disclaimer

It is stated,

> The information on National Weather Service (NWS) Web pages are in the public domain, unless specifically noted otherwise, and may be used without charge for any lawful purpose so long as you do not: 1) claim it is your own (e.g., by claiming copyright for NWS information -- see below), 2) use it in a manner that implies an endorsement or affiliation with NOAA/NWS, or 3) modify its content and then present it as official government material. You also cannot present information of your own in a way that makes it appear to be official government information.

And, additionally,

> As required by 17 U.S.C. § 403, third parties producing copyrighted works consisting predominantly of the material appearing in NWS Web pages must provide notice with such work(s) identifying the NWS material incorporated and stating that such material is not subject to copyright protection.

So, I'm not sure if there is a specific license for that other than the [CC0 v0.1 Universal](https://github.com/timtrice/rrricanesdata/blob/master/LICENSE) which I had gone with. I honestly have no idea!
