### 2017-07-11 #118 @maelle
I am no license expert at all, can you ask this in #general of the Slack? :-)

Cool about the drat package project!
