### 2017-07-11 #118 @maelle
Thanks a lot @robinsones @jsta !

@timtrice great work and two things:

* I see you have an Acknowledgements part in your README. For acknowledging the reviewers' contribution you _can_ add them as "rev" in DESCRIPTION, see [this](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#authorship) It's only a suggestion!

* Will the `rrricanesdata` package be part of your submission? It makes sense I guess since it empowers `rrricanes`, but it didn't exist yet at the moment of the reviews so maybe it should be in a separate submission. 🤔  I'll ping @ropensci/editors here for their opinion. As a reminder/explanation for them `rrricanesdata` will be a package hosted on drat with data that `rrricanes` can output for the user (instead of downloading them from the frail website every time). Currently there is such a system in `rrricanes` but the data comes from a simple Github repository of @timtrice's, not a package.
