### 2017-07-11 71ce5fa
Corrected to use crul func
