### 2017-07-11 4a7c3ea
crul_extract_storms func
