+++ R/crul.R
@@ -1,53 +1,195 @@
-#' @title crul_get_url_contents
-#' @description Get contents from URL
-#' @param link URL to download
+#' @title crul_discus
+#' @description Parse storm Discussion products
+#' @param contents text of Storm Discussion product
+#' @keywords internal
+crul_discus <- function(contents) {
+
+  # Replace all carriage returns with empty string.
+  contents <- stringr::str_replace_all(contents, "\r", "")
+
+  # Make sure this is a discussion product
+  if (!any(stringr::str_count(contents,
+                              c("MIATCD", "MIATCM", "TCD", "WTPA", "WTPZ",
+                                "MIAWRKAD1")))) {
+    # Check if the term "DISCUSSION" appears in header
+    if (!stringr::str_detect(scrape_header(contents), "DISCUSSION"))
+      stop(sprintf("Invalid Discussion link. %s", link))
+  }
+
+  df <- create_df_discus()
+
+  status <- scrape_header(contents, ret = "status")
+  name <- scrape_header(contents, ret = "name")
+  adv <- scrape_header(contents, ret = "adv") %>% as.numeric()
+  date <- scrape_header(contents, ret = "date")
+
+  # Keys were added to discus products beginning 2006. Prior, it doesn't
+  # exist. safely run scrape_header for key. If error, then use NA. Otherwise,
+  # add it.
+  safely_scrape_header <- purrr::safely(scrape_header)
+  key <- safely_scrape_header(contents, ret = "key")
+  if (is.null(key$error)) {
+    key <- key$result
+  } else {
+    key <- NA
+  }
+
+  if (getOption("rrricanes.working_msg"))
+    message(sprintf("Working %s %s Storm Discussion #%s (%s)",
+                    status, name, adv, date))
+
+  df <- df %>%
+    tibble::add_row("Status" = status,
+                    "Name" = name,
+                    "Adv" = adv,
+                    "Date" = date,
+                    "Key" = key,
+                    "Contents" = contents)
+
+  return(df)
+}
+
+#' @title crul_extract_storms
+#' @description Extract storms for the given basin
+#' @param basin AL or EP
+#' @param contents Contents of archive pages
 #' @export
-crul_get_url_contents <- function(links, p) {
-  p$pause(0.5)$tick()$print()
-  links <- crul::Async$new(urls = links)
-  res <- links$get()
-  # Check status codes
-  #purrr::map(res, ~.$status_code) %>% purrr::flatten_dbl() %>% unique()
+crul_extract_storms <- function(basin, contents) {
+
+  if (basin == "AL") {
+    link_xpath <- "//td[(((count(preceding-sibling::*) + 1) = 1) and parent::*)]//a"
+  } else if (basin == "EP") {
+    link_xpath <- "//td[(((count(preceding-sibling::*) + 1) = 2) and parent::*)]//a"
+  } else {
+    stop("No basin")
+  }
+
+  contents <- purrr::map(contents, ~.$parse("UTF-8")) %>%
+    purrr::map(xml2::read_html)
+
+  years <- purrr::map(contents, rvest::html_nodes, xpath = "//title") %>%
+    purrr::map(rvest::html_text) %>%
+    stringr::str_sub(0L, 4L) %>%
+    as.numeric()
+
+  storms <- purrr::map(contents, rvest::html_nodes, xpath = link_xpath)
+  names <- purrr::map(storms, rvest::html_text) %>%
+    purrr::map(stringr::str_to_title)
+  links <- purrr::map(storms, rvest::html_attr, name = "href") %>%
+    purrr::map2(years, ~paste0(year_archives_link(.y), .x))
+  basins <- purrr::map(names, purrr::rep_along, basin)
[…]
