### 2017-07-11 c0afe8e
Modify func to filter multiple links
