### 2017-07-11 efcaa1e
Parse only non-empty datasets
