### 2017-07-11 0630d95
Modify return of crul_get_fstadv
