### 2017-07-11 4810ec3
Modify return of crul_get_posest
