### 2017-07-11 b4c8707
Modify return of crul_get_prblty
