### 2017-07-11 8dc93e0
Modify return of crul_get_public
