### 2017-07-11 9ed1a33
Modify return of crul_get_update
