### 2017-07-11 fb7c9a3
Modify return of crul_get_wndprb
