+++ DESCRIPTION
@@ -41,3 +41,4 @@ Suggests:
   testthat
 RoxygenNote: 6.0.1
 VignetteBuilder: knitr
+Additional_repositories: https://geanders.github.io/drat
