+++ R/scrapers.R
@@ -238,7 +238,7 @@ scrape_header <- function(contents, ret = NULL) {
     # Use day, month, date and year which seems to be consistent across all
     # products.
     # (timtrice): Added backtick for AL162005 public #18
-    ptn_header <- paste0("^[[:alnum:][:blank:][:punct:]`\n\r]*?",
+    ptn_header <- paste0("[[:alnum:][:blank:][:punct:]`\n\r]*?",
                          "[:alpha:]{3}", # Day of week
                          "[:blank:]*",
                          "[:alpha:]{3}", # Month, abbreviated
