+++ R/crul.R
@@ -280,19 +280,56 @@ crul_get_storm_data <- function(links,
   product_links <- split(product_links, ceiling(seq_along(product_links)/4))
   # set progress
   p <- dplyr::progress_estimated(length(product_links))
-  res <- purrr::map(product_links, crul_get_url_contents, p) %>% purrr::flatten()
-  res_parsed <- purrr::map(res, ~.$parse("UTF-8"))
+  res <- purrr::map(product_links, crul_get_url_contents, p) %>%
+    purrr::flatten()
+  #res_parsed <- purrr::map(res, ~.$parse("UTF-8"))
+  # res_parsed <- purrr::map(res, ~xml2::read_html(.$content)) %>%
+  #   purrr::map(rvest::html_nodes, xpath = "//pre") %>%
+  #   purrr::map(rvest::html_text)
+  res_parsed <- purrr::map(res, ~xml2::read_html(.$content)) %>%
+    purrr::map(.f = function(x) {
+      txt <- rvest::html_node(x, xpath = "//pre") %>% rvest::html_text()
+      if (is.na(txt)) txt <- rvest::html_text(x)
+    })
 
   list_products <- list(
-    "discus" = purrr::map(res, ~.$url) %>% filter_discus() %>% purrr::map(~(!purrr::is_empty(.))) %>% purrr::flatten_lgl() %>% res_parsed[.],
-    "fstadv" = purrr::map(res, ~.$url) %>% filter_fstadv() %>% purrr::map(~(!purrr::is_empty(.))) %>% purrr::flatten_lgl() %>% res_parsed[.],
-    "posest" = purrr::map(res, ~.$url) %>% filter_posest() %>% purrr::map(~(!purrr::is_empty(.))) %>% purrr::flatten_lgl() %>% res_parsed[.],
-    "prblty" = purrr::map(res, ~.$url) %>% filter_prblty() %>% purrr::map(~(!purrr::is_empty(.))) %>% purrr::flatten_lgl() %>% res_parsed[.],
-    "public" = purrr::map(res, ~.$url) %>% filter_public() %>% purrr::map(~(!purrr::is_empty(.))) %>% purrr::flatten_lgl() %>% res_parsed[.],
-    "update" = purrr::map(res, ~.$url) %>% filter_update() %>% purrr::map(~(!purrr::is_empty(.))) %>% purrr::flatten_lgl() %>% res_parsed[.],
-    "wndprb" = purrr::map(res, ~.$url) %>% filter_wndprb() %>% purrr::map(~(!purrr::is_empty(.))) %>% purrr::flatten_lgl() %>% res_parsed[.])
+    "discus" = purrr::map(res, ~.$url) %>%
+      filter_discus() %>%
+      purrr::map(~(!purrr::is_empty(.))) %>%
+      purrr::flatten_lgl() %>% res_parsed[.],
+    "fstadv" = purrr::map(res, ~.$url) %>%
+      filter_fstadv() %>%
+      purrr::map(~(!purrr::is_empty(.))) %>%
+      purrr::flatten_lgl() %>%
+      res_parsed[.],
+    "posest" = purrr::map(res, ~.$url) %>%
+      filter_posest() %>%
+      purrr::map(~(!purrr::is_empty(.))) %>%
+      purrr::flatten_lgl() %>%
+      res_parsed[.],
+    "prblty" = purrr::map(res, ~.$url) %>%
+      filter_prblty() %>%
+      purrr::map(~(!purrr::is_empty(.))) %>%
+      purrr::flatten_lgl() %>%
+      res_parsed[.],
+    "public" = purrr::map(res, ~.$url) %>%
+      filter_public() %>%
[…]
