+++ man/crul_prblty.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_prblty}
+\alias{crul_prblty}
+\title{crul_prblty}
+\usage{
+crul_prblty(contents)
+}
+\arguments{
+\item{contents}{Link to a storm's specific strike probability advisory product.}
+}
+\description{
+Parse strike probability products
+}
