+++ man/crul_public.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_public}
+\alias{crul_public}
+\title{crul_public}
+\usage{
+crul_public(contents)
+}
+\arguments{
+\item{contents}{Link to a storm's specific public advisory product.}
+}
+\description{
+Parse Public Advisory products
+}
