+++ man/crul_update.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/crul.R
+\name{crul_update}
+\alias{crul_update}
+\title{crul_update}
+\usage{
+crul_update(contents)
+}
+\arguments{
+\item{contents}{Link to a storm's specific update advisory product.}
+}
+\description{
+Parse cyclone update products
+}
