+++ NAMESPACE
@@ -15,6 +15,11 @@ export(crul_get_storms)
 export(crul_get_update)
 export(crul_get_url_contents)
 export(crul_get_wndprb)
+export(crul_posest)
+export(crul_prblty)
+export(crul_public)
+export(crul_update)
+export(crul_wndprb)
 export(ep_prblty_stations)
 export(ep_tracking_chart)
 export(get_discus)
