### 2017-07-13 b6a48c4
Return txt object for parsing
