### 2017-07-13 f50621c
Add caret back to pattern
