### 2017-07-13 4ffb089
crul functions implemented replacing old functions
Pkg now has near-complete use of `crul` package; all products and
funcs accessing storm data (other than GIS) are implementing crul.

Some functions that are no longer in use have been removed. Associated
tests have also been removed.

Only one test for the `posest` dataframe was modified from
expect_identical to expect_equal. Links may not be accessed in order
which means observations are not likely to be in order. Where Key, Adv
exists this is easy to sort out. For other products such as posest,
data can only be sorted by date.

Additional t
[…]
