+++ R/public.R
@@ -27,74 +27,57 @@ create_df_public <- function() {
 #'   \item{Key}{Unique ID of the cyclone}
 #'   \item{Contents}{Text content of product}
 #' }
-#' @param link URL to storm's archive page.
+#' @param links URL to storm's archive page.
 #' @seealso \code{\link{get_storms}}, \code{\link{public}}
 #' @export
-get_public <- function(link) {
-
-    products <- purrr::map(link, get_products) %>% purrr::flatten_chr()
-
-    products <- filter_public(products)
-
-    # Set progress bar
-    p <- dplyr::progress_estimated(n = length(products))
-
-    products.public <- purrr::map(products, public, p)
-
-    public <- purrr::map_df(products.public, dplyr::bind_rows)
-
-    return(public)
+get_public <- function(links) {
+  df <- get_storm_data(links, products = "public")
+  return(df$public)
 }
 
 #' @title public
 #' @description Parse Public Advisory products
 #' @details Given a direct link to a public advisory product, parse and return
 #' dataframe of values.
-#' @param link Link to a storm's specific public advisory product.
-#' @param p dplyr::progress_estimate.
+#' @param contents Link to a storm's specific public advisory product.
 #' @return Dataframe
 #' @seealso \code{\link{get_public}}
 #' @keywords internal
-public <- function(link, p = dplyr::progress_estimated(n = 1)) {
+public <- function(contents) {
+  # Replace all carriage returns with empty string.
+  contents <- stringr::str_replace_all(contents, "\r", "")
 
-    p$pause(0.5)$tick()$print()
+  # Make sure this is a public advisory product
+  if (!any(stringr::str_count(contents, c("MIATCP", "TCP", "WTPA",
+                                          "MIAWRKAP"))))
+    stop(sprintf("Invalid Public Advisory link. %s", link))
 
-    contents <- scrape_contents(link)
+  df <- create_df_public()
 
-    # Replace all carriage returns with empty string.
-    contents <- stringr::str_replace_all(contents, "\r", "")
+  status <- scrape_header(contents, ret = "status")
+  name <- scrape_header(contents, ret = "name")
+  adv <- scrape_header(contents, ret = "adv")
+  date <- scrape_header(contents, ret = "date")
 
-    # Make sure this is a public advisory product
-    if (!any(stringr::str_count(contents, c("MIATCP", "TCP", "WTPA",
-                                            "MIAWRKAP"))))
-        stop(sprintf("Invalid Public Advisory link. %s", link))
+  safely_scrape_header <- purrr::safely(scrape_header)
+  key <- safely_scrape_header(contents, ret = "key")
+  if (is.null(key$error)) {
+    key <- key$result
+  } else {
+    key <- NA
+  }
 
-    df <- create_df_public()
+  if (getOption("rrricanes.working_msg"))
+    message(sprintf("Working %s %s Public Advisory #%s (%s)",
+                    status, name, adv, date))
 
-    status <- scrape_header(contents, ret = "status")
-    name <- scrape_header(contents, ret = "name")
-    adv <- scrape_header(contents, ret = "adv")
-    date <- scrape_header(contents, ret = "date")
+  df <- df %>%
+    tibble::add_row("Status" = status,
+                    "Name" = name,
+                    "Adv" = adv,
+                    "Date" = date,
+                    "Key" = key,
+                    "Contents" = contents)
 
-    safely_scrape_header <- purrr::safely(scrape_header)
-    key <- safely_scrape_header(contents, ret = "key")
-    if (is.null(key$error)) {
-        key <- key$result
-    } else {
-        key <- NA
-    }
-
-    if (getOption("rrricanes.working_msg"))
-        message(sprintf("Working %s %s Public Advisory #%s (%s)",
[…]
