+++ tests/testthat/test_get_storm_data.R
@@ -18,5 +18,5 @@ test_that("get_storm_data()", {
                      .$Link %>%
                      get_storm_data(products = "test"))
     expect_error(get_storm_data(),
-                 "argument \"link\" is missing, with no default")
+                 "argument \"links\" is missing, with no default")
 })
