+++ DESCRIPTION
@@ -3,7 +3,9 @@ Type: Package
 Title: Web scraper for Atlantic and east Pacific hurricanes and tropical storms
 Description: Get archived data of past and current hurricanes and tropical 
   storms for the Atlantic and eastern Pacific oceans. Data is available for 
-  storms since 1998.
+  storms since 1998. Datasets are updated via the rrricanesdata package. 
+  Currently, this package is about 6MB of datasets. See the README or view 
+  `vignette("drat")` for more information.
 Version: 0.2.0-6
 Authors@R: c(person("Tim", "Trice", email = "tim.trice@gmail.com", role = c("aut", "cre")))
 Depends:
@@ -37,8 +39,9 @@ Suggests:
   devtools, 
   knitr, 
   rmarkdown,
+  rrricanesdata (>= 0.0.1),
   sp,
   testthat
 RoxygenNote: 6.0.1
 VignetteBuilder: knitr
-Additional_repositories: https://geanders.github.io/drat
+Additional_repositories: https://timtrice.github.io/drat
