### 2017-07-14 #118 @timtrice
@maelle ,

Wanted to post an update; I've implemented drat for `rrricanesdata`. By and large all seems well except my Travis and Appveyor tests now fail in `rrricanes` as it's looking for pkg `rrricanesdata` (listed under Suggests as suggested) but, of course, it's not there. So, I'll take a look at that later and try to correct it.

I can submit `rrricanesdata` as it's own package if need be; it doesn't matter to me one way or another. My only concern or thinking is on updating. My initial thought is to update on the 1st of every month any new storms and then bump the version as a patch. Is there any recommendation on this?  As it stands right now it's updated as of today (however, not yet pushed to the master branch until I get `rrricanes` ready).

While I'm at it, thanks to you for pointing out `crul` and @sckott for writing it. It has made getting data significantly faster and I'm not seeing any timeout issues at. all. It's been remarkable. I need to clean the code that's getting the data as I was kind of hacking through it but it's definitely been a huge benefit!

Tim Trice
