### 2017-07-14 f786ab8
Remove duplicate onAttach funcs
