### 2017-07-14 dea1060
Remove failing tests
