+++ tests/testthat/test_wndprb.R
@@ -20,11 +20,3 @@ test_that("Test cp_prblty_stations()", {
 test_that("Test ep_prblty_stations()", {
     expect_identical(ep_prblty_stations(), FALSE)
 })
-
-## ---- Test get_wndprb() ------------------------------------------------------
-test_that("Test get_wndprb()", {
-    skip_on_cran()
-    al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
-    df.al092008.wndprb <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_wndprb()
-    expect_identical(al092008.wndprb, df.al092008.wndprb)
-})
