+++ tests/testthat/test_discus.R
@@ -15,14 +15,3 @@ test_that("Dataframe Skeleton", {
     expect_identical(class(df$Key), "character")
     expect_identical(class(df$Contents), "character")
 })
-
-## ---- Test get_discus() ------------------------------------------------------
-#' Test return of get_discus()
-test_that("Test get_discus()", {
-    skip_on_cran()
-
-    al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
-    df.al092008.discus <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_discus()
-
-    expect_identical(al092008.discus, df.al092008.discus)
-})
