### 2017-07-14 timtrice/rrricanes#71 @timtrice
Many tests were removed during development of hotfix/v0.2.0-6. Some are ok as they tested functions that no longer exist.

However, tests removed in commit dea1060] need to be re-added. Unsure at the moment why they were failling. When ran tests check they failed. But if ran from script or console they passed.
