### 2017-07-14 c3a86d5
Attempt build of rrricanesdata in Appveyor
