### 2017-07-14 28b08ff
Do not check Suggests on osx
