### 2017-07-14 4712bcf
Attempt r_github_packages for drat
