### 2017-07-14 4fc1659
Add rrricanesdata info to pkg doc
