### 2017-07-14 85aa888
Add rrricanesdata info to README
