### 2017-07-14 4016c09
Add rrricanesdata info to pkg doc
