+++ vignettes/installing_rrricanesdata.Rmd
@@ -1,0 +1,220 @@
+---
+title: "Installing rrricanesdata"
+author: "Tim Trice"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Installing rrricanesrdata}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r echo = FALSE}
+hasData <- requireNamespace("rrricanesdata", quietly=TRUE)
+if (!hasData) {
+    knitr::opts_chunk$set(eval = FALSE)
+    msg <- paste("Note: Examples in this vignette require that the", 
+                 "`rrricanesdata` package be installed. The system",
+                 "currently running this vignette does not have that package",
+                 "installed, so code examples will not be evaluated.")
+    msg <- paste(strwrap(msg), collapse="\n")
+    message(msg)
+}
+```
+
+```{r echo = FALSE, message = FALSE, warning = FALSE}
+# If you need to, install the `hurricaneexposuredata` package from the drat repo:
+# library(drat)
+# addRepo("geanders")
+# install.packages("hurricaneexposuredata")
+library(rrricanesdata)
+library(rrricanes)
+```
+
+With `rrricanes` you can access current and archived advisories as-issued from the National Hurricane Center archives. `rrricanesdata` is a complimentary data package meant to make it faster to get most data avialable from these storms. 
+
+Installing `rrricanesdata` is simple:
+
+```{r, eval = FALSE}
+install.packages("rrricanesdata", 
+                 repos = "https://timtrice.github.io/drat/", 
+                 type = "source")
+```
+
+Data in `rrricanesdata` will be updated on the first of every month with a cutoff date of midnight on the last day of the month. So, advisories issued at any time during the current month will not be available; you will need to use any of `rrricanes` `get_*` functions. 
+
+## Datasets
+
+### `adv`
+  * Key: Unique identifier of cyclone
+  * Adv: Advisory number
+  * Date: Date and time of advisory
+  * Status: Classification of cyclone
+  * Name: Name of cyclone
+  * Lat: Latitude of cyclone center
+  * Lon: Longitude of cyclone center
+  * Wind: Maximum sustained one-minute winds in knots
+  * Gust: Maximum sustained one-minute gusts in knots
+  * Pressure: Minimum central pressure in millibars
+  * PosAcc: Position accuracy of cyclone in nautical miles
+  * FwdDir: Compass angle of forward motion
+  * FwdSpeed: Forward speed in miles per hour
+  * Eye: Size of eye in nautical miles
+  * SeasNE: Radius of 12ft seas in northeast quadrant
+  * SeasSE: Radius of 12ft seas in southeast quadrant
+  * SeasSW: Radius of 12ft seas in southwest quadrant
+  * SeasNW: Radius of 12ft seas in northwest quadrant
+
+### `discus`
+  * Status: Classification of storm, e.g., Tropical Storm, Hurricane,
+    etc.
+  * Name: Name of storm
+  * Adv: Advisory Number
+  * Date: Date of advisory issuance
+  * Key: ID of cyclone
+  * Contents: Text content of product
+
+### `fcst`
+  * Key: Unique identifier of cyclone
+  * Adv: Advisory number
+  * Date: Date and time of advisory
+  * FcstDate: Forecast date and time in UTC
+  * Lat: Forecast latitude
+  * Lon: Forecast Longitude
+  * Wind: Forecast wind in knots
[…]
