### 2017-07-14 60ee511
Remove func load_storm_data
