+++ R/get_storm_data.R
@@ -170,75 +170,3 @@ get_storm_data <- function(links, products = c("discus", "fstadv", "posest",
 
   return(ds)
 }
-
-#' @title load_storm_data
-#' @description Load storm and year data from data repository.
-#' @param dataset A dataset to return
-#' @param ... additional parameters for readr
-#' @details This function is designed to give quicker access to post-scraped
-#' storm data and may be modified in future releases.
-#' \describe{
-#'     \item{adv}{This data file contains base information for every storm
-#'     advisory issued by the National Hurricane Center for the Atlantic and
-#'     northeast Pacific oceans.}
-#'     \item{discus}{Storm discussion text}
-#'     \item{fcst}{Forecast positions of tropical cyclones}
-#'     \item{fcst_wr}{Forecast wind radii data for each forecast observation in
-#'     `fcst`}
-#'     \item{fstadv}{Complete Forecast/Advisory data}
-#'     \item{posest}{Position estimates}
-#'     \item{prblty}{Strike probabilities for a given location. This product was
-#'     deprecated after the 2005 hurricane season.}
-#'     \item{public}{Public advisory text}
-#'     \item{storms}{Summary for all storms in this dataset.}
-#'     \item{update}{Update text}
-#'     \item{wndprb}{Wind Speed Probabilities. Probability of a location
-#'     experiencing minimum wind-speed values within a given forecast period.
-#'     This product replaces Strike Probabilities after the 2005 hurricane
-#'     season. May not exist for all cyclones.}
-#'     \item{wr}{Current wind radius, if available, for a cyclone.}
-#' }
-#'
-#' Datasets 'discus', 'posest', 'public' and 'update' are not included as of
-#' this writing but will be added as soon as possible.
-#'
-#' 'adv', 'fcst', 'fcst_wr', and 'wr' are tidied data of
-#' \code{\link{get_fstadv}} using functions \code{\link{tidy_fcst}},
-#' \code{\link{tidy_fcst_wr}}, \code{\link{tidy_fstadv}} and
-#' \code{\link{tidy_wr}}.
-#'
-#' See \url{https://timtrice.github.io/rrricanes/articles/articles/data_world.html}
-#' for a third alternative to access datasets with the ability to filter data
-#' without loading entire datasets into your environment.
-#' @seealso \url{https://timtrice.github.io/rrricanes/articles/articles/data_world.html}
-#' @export
-load_storm_data <- function(dataset = c("adv", "discus", "fcst", "fcst_wr",
-                                        "fstadv", "posest", "prblty", "public",
-                                        "storms", "update", "wndprb", "wr"),
-                            ...) {
-    dataset <- match.arg(dataset)
-    base_url <- "https://github.com/timtrice/rrricanesdata/blob/master/"
-    link <- paste0(base_url, dataset, ".csv?raw=true")
-    readr_args <- list(...)
-    if (purrr::is_empty(readr_args$col_types)) {
-        if (dataset == "discus") {
-            col_types = "cciTcc"
-        } else if (dataset == "fstadv") {
-            col_types = paste0("cciTcddiidiiiiiiiiiiiiiiiiiiiiTddiiiiiiiiiiiiii",
-                               "TddiiiiiiiiiiiiiiTddiiiiiiiiiiiiiiTddiiiiiiiiii",
-                               "TddiiiiiiiiiiTddiiTddii")
-        } else if (dataset == "public") {
-            col_types = "cccTcc"
-        } else if (dataset == "update") {
-            col_types = "ccTcc"
-        } else if (dataset == "wndprb") {
-            col_types = "ciTciiiiiiiiiiiiii"
-        } else {
-            col_types = readr::cols()
-        }
[…]
