+++ man/load_storm_data.Rd
@@ -1,57 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get_storm_data.R
-\name{load_storm_data}
-\alias{load_storm_data}
-\title{load_storm_data}
-\usage{
-load_storm_data(dataset = c("adv", "discus", "fcst", "fcst_wr", "fstadv",
-  "posest", "prblty", "public", "storms", "update", "wndprb", "wr"), ...)
-}
-\arguments{
-\item{dataset}{A dataset to return}
-
-\item{...}{additional parameters for readr}
-}
-\description{
-Load storm and year data from data repository.
-}
-\details{
-This function is designed to give quicker access to post-scraped
-storm data and may be modified in future releases.
-\describe{
-    \item{adv}{This data file contains base information for every storm
-    advisory issued by the National Hurricane Center for the Atlantic and
-    northeast Pacific oceans.}
-    \item{discus}{Storm discussion text}
-    \item{fcst}{Forecast positions of tropical cyclones}
-    \item{fcst_wr}{Forecast wind radii data for each forecast observation in
-    `fcst`}
-    \item{fstadv}{Complete Forecast/Advisory data}
-    \item{posest}{Position estimates}
-    \item{prblty}{Strike probabilities for a given location. This product was
-    deprecated after the 2005 hurricane season.}
-    \item{public}{Public advisory text}
-    \item{storms}{Summary for all storms in this dataset.}
-    \item{update}{Update text}
-    \item{wndprb}{Wind Speed Probabilities. Probability of a location
-    experiencing minimum wind-speed values within a given forecast period.
-    This product replaces Strike Probabilities after the 2005 hurricane
-    season. May not exist for all cyclones.}
-    \item{wr}{Current wind radius, if available, for a cyclone.}
-}
-
-Datasets 'discus', 'posest', 'public' and 'update' are not included as of
-this writing but will be added as soon as possible.
-
-'adv', 'fcst', 'fcst_wr', and 'wr' are tidied data of
-\code{\link{get_fstadv}} using functions \code{\link{tidy_fcst}},
-\code{\link{tidy_fcst_wr}}, \code{\link{tidy_fstadv}} and
-\code{\link{tidy_wr}}.
-
-See \url{https://timtrice.github.io/rrricanes/articles/articles/data_world.html}
-for a third alternative to access datasets with the ability to filter data
-without loading entire datasets into your environment.
-}
-\seealso{
-\url{https://timtrice.github.io/rrricanes/articles/articles/data_world.html}
-}
