+++ NAMESPACE
@@ -23,7 +23,6 @@ export(gis_prob_storm_surge)
 export(gis_windfield)
 export(gis_wsp)
 export(knots_to_mph)
-export(load_storm_data)
 export(mb_to_in)
 export(saffir)
 export(shp_to_df)
