### 2017-07-14 7a3d2f2
Add Basin to globalVariables from get_storms
