### 2017-07-14 b404cce
Remove product validation; done in get_storm-data
