+++ R/public.R
@@ -47,11 +47,6 @@ public <- function(contents) {
   # Replace all carriage returns with empty string.
   contents <- stringr::str_replace_all(contents, "\r", "")
 
-  # Make sure this is a public advisory product
-  if (!any(stringr::str_count(contents, c("MIATCP", "TCP", "WTPA",
-                                          "MIAWRKAP"))))
-    stop(sprintf("Invalid Public Advisory link. %s", link))
-
   df <- create_df_public()
 
   status <- scrape_header(contents, ret = "status")
