### 2017-07-14 510a306
Re-add rrricanes.working_msg opt
