### 2017-07-14 6149938
Add more product descriptions to pkg doc
