+++ R/data.R
@@ -1,35 +1,35 @@
-#' GIS advisory dataset for Hurricane Sandy Adv 18
-#' @source \url{http://www.nhc.noaa.gov/gis/archive_forecast_results.php?id=al18&year=2012&name=Hurricane\%20SANDY}
-"adv"
-
 #' Atlantic cyclones for 2012
 #' @source \url{http://www.nhc.noaa.gov/archive/2012/}
-"al.2012"
+"df.al_2012"
 
 #' Strike probabilities for Hurricane Katrina (AL122005)
 #' @source \url{http://www.nhc.noaa.gov/archive/2005/KATRINA.shtml?}
-"al122005_prblty"
+"df.al_12_2005_prblty"
 
 #' Wind speed probabilities for Hurricane Sandy (AL182012)
 #' @source \url{http://www.nhc.noaa.gov/archive/2012/SANDY.shtml?}
-"al182012_wndprb"
+"df.al_18_2012_wndprb"
 
 #' Forecast/Advisory and Wind Speed Probabilities for Hurricane Sandy (AL182012)
 #' @source \url{http://www.nhc.noaa.gov/archive/2012/SANDY.shtml?}
-"al182012"
+"df.al_18_2012"
 
 #' Forecast/Advisory for Hurricane Sandy (AL182012)
 #' @source \url{http://www.nhc.noaa.gov/archive/2012/SANDY.shtml?}
-"al182012_fstadv"
+"df.al_18_2012_fstadv"
+
+#' GIS advisory dataset for Hurricane Sandy Adv 18
+#' @source \url{http://www.nhc.noaa.gov/gis/archive_forecast_results.php?id=al18&year=2012&name=Hurricane\%20SANDY}
+"df.gis_adv"
 
 #' GIS storm surge shapefile dataset for Hurricane Sandy (AL182012)
 #' @source \url{http://www.nhc.noaa.gov/gis/archive_psurge_results.php?id=al18&year=2012&name=Hurricane\%20SANDY}
-"ss"
+"df.gis_storm_surge"
 
 #' GIS windfield and forecast wind radii for Hurricane Sandy (AL182012)
 #' @source \url{http://www.nhc.noaa.gov/gis/archive_forecast_info_results.php?id=al18&year=2012&name=Hurricane\%20SANDY}
-"wf"
+"df.gis_wind_radii"
 
 #' GIS wind speed probabilities for Hurricane Sandy (AL182012)
 #' @source \url{http://www.nhc.noaa.gov/gis/archive_wsp.php}
-"wsp"
+"df.gis_wsp"
