+++ man/df.gis_storm_surge.Rd
@@ -1,15 +1,15 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/data.R
 \docType{data}
-\name{ss}
-\alias{ss}
+\name{df.gis_storm_surge}
+\alias{df.gis_storm_surge}
 \title{GIS storm surge shapefile dataset for Hurricane Sandy (AL182012)}
 \format{An object of class \code{list} of length 1.}
 \source{
 \url{http://www.nhc.noaa.gov/gis/archive_psurge_results.php?id=al18&year=2012&name=Hurricane\%20SANDY}
 }
 \usage{
-ss
+df.gis_storm_surge
 }
 \description{
 GIS storm surge shapefile dataset for Hurricane Sandy (AL182012)
