+++ man/df.gis_wsp.Rd
@@ -1,15 +1,15 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/data.R
 \docType{data}
-\name{wsp}
-\alias{wsp}
+\name{df.gis_wsp}
+\alias{df.gis_wsp}
 \title{GIS wind speed probabilities for Hurricane Sandy (AL182012)}
 \format{An object of class \code{list} of length 3.}
 \source{
 \url{http://www.nhc.noaa.gov/gis/archive_wsp.php}
 }
 \usage{
-wsp
+df.gis_wsp
 }
 \description{
 GIS wind speed probabilities for Hurricane Sandy (AL182012)
