+++ man/df.al_18_2012.Rd
@@ -1,15 +1,15 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/data.R
 \docType{data}
-\name{al182012}
-\alias{al182012}
+\name{df.al_18_2012}
+\alias{df.al_18_2012}
 \title{Forecast/Advisory and Wind Speed Probabilities for Hurricane Sandy (AL182012)}
 \format{An object of class \code{list} of length 2.}
 \source{
 \url{http://www.nhc.noaa.gov/archive/2012/SANDY.shtml?}
 }
 \usage{
-al182012
+df.al_18_2012
 }
 \description{
 Forecast/Advisory and Wind Speed Probabilities for Hurricane Sandy (AL182012)
