+++ man/df.al_18_2012_wndprb.Rd
@@ -1,15 +1,15 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/data.R
 \docType{data}
-\name{al182012_wndprb}
-\alias{al182012_wndprb}
+\name{df.al_18_2012_wndprb}
+\alias{df.al_18_2012_wndprb}
 \title{Wind speed probabilities for Hurricane Sandy (AL182012)}
 \format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2227 rows and 18 columns.}
 \source{
 \url{http://www.nhc.noaa.gov/archive/2012/SANDY.shtml?}
 }
 \usage{
-al182012_wndprb
+df.al_18_2012_wndprb
 }
 \description{
 Wind speed probabilities for Hurricane Sandy (AL182012)
