+++ man/df.gis_wind_radii.Rd
@@ -1,15 +1,15 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/data.R
 \docType{data}
-\name{wf}
-\alias{wf}
+\name{df.gis_wind_radii}
+\alias{df.gis_wind_radii}
 \title{GIS windfield and forecast wind radii for Hurricane Sandy (AL182012)}
 \format{An object of class \code{list} of length 2.}
 \source{
 \url{http://www.nhc.noaa.gov/gis/archive_forecast_info_results.php?id=al18&year=2012&name=Hurricane\%20SANDY}
 }
 \usage{
-wf
+df.gis_wind_radii
 }
 \description{
 GIS windfield and forecast wind radii for Hurricane Sandy (AL182012)
