+++ man/al_prblty_stations.Rd
@@ -1,16 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/wndprb.R
-\name{al_prblty_stations}
-\alias{al_prblty_stations}
-\title{al_prblty_stations}
-\usage{
-al_prblty_stations()
-}
-\description{
-Retrieve list of probability stations based in the Atlantic
-basin from the NHC. To be used in tandem with `wndprb` products.
-}
-\details{
-This function may be deprecated soon as the data sources may be
-removed from the National Hurricane Center.
-}

[…]
