### 2017-07-14 cd32500
Re-add funcs al*; I ran a dumb command earlier"
