### 2017-07-15 65157e7
Clean func get_storm_data, progress bar
Progress bar now captures entire process of data gathering and
scraping. Cleaned up process as much of it was duplicated and
confusing.

All tests and checks pass. 0 ERRORS, 0 WARNINGS, 0 NOTES
