+++ R/get_storm_data.R
@@ -1,3 +1,67 @@
+#' @title extract_product_contents
+#' @description Get and parse product contents for each links
+#' @param links URLs to storm products
+#' @param product specific product to parse
+#' @param p dplyr::progress_estimated object
+#' @keywords internal
+extract_product_contents <- function(links, product, p) {
+
+  p$pause(0.5)$tick()$print()
+
+  res <- purrr::map(links, get_url_contents) %>% purrr::flatten()
+
+  res_parsed <- purrr::map(res, ~xml2::read_html(.$content)) %>%
+    purrr::map(.f = function(x) {
+      if (is.na(txt <- rvest::html_node(x, xpath = "//pre") %>%
+                rvest::html_text()))
+        txt <- rvest::html_text(x)
+      return(txt)
+    })
+
+  df <- purrr::invoke_map_df(product, .x = res_parsed)
+
+  return(df)
+
+}
+
+#' @title extract_storm_links
+#' @description Extract product links from a storm's archive page
+#' @param links URLs to a storm's archive page
+#' @param p dplyr::progress_estimated object
+#' @keywords internal
+extract_storm_links <- function(links, p) {
+
+  if (p$i > 0) p$pause(0.5)
+  p$tick()$print()
+
+  res <- purrr::map(links, .f = function(x) {get_url_contents(x)}) %>%
+    purrr::flatten()
+
+  # Get contents of all storms
+  storm_contents <- purrr::map(res, ~.$parse("UTF-8")) %>%
+    purrr::map(xml2::read_html)
+
+  years <- purrr::map(res, ~.$url) %>%
+    purrr::flatten_chr() %>%
+    stringr::str_extract("[[:digit:]]{4}") %>%
+    as.numeric()
+
+  # Extract all links
+  product_links <- purrr::map(storm_contents,
+                              ~rvest::html_nodes(x = .x, xpath = "//td//a")) %>%
+    purrr::map(~rvest::html_attr(x = .x, name = "href"))
+
+  # 1998 product links are relative and prefixed with "/archive/1998/" whereas
+  # other years, product_links are absolute. If product_links exist for 1998
+  # they must be modified. After, all product_links must be prefixed with
+  # nhc_domain
+  nhc_domain <- get_nhc_link(withTrailingSlash = FALSE)
+  product_links[years == 1998] <- purrr::map(product_links[years == 1998],
+                                             ~sprintf("/archive/1998/%s", .))
+  product_links <- purrr::map(product_links, ~sprintf("%s%s", nhc_domain, .))
+  return(product_links)
+}
+
 #' @title get_storm_content
 #' @description Extract content from a storm's archive page
 #' @param link of archive page
@@ -61,112 +125,41 @@ get_storm_data <- function(links, products = c("discus", "fstadv", "posest",
 
   products <- match.arg(products, several.ok = TRUE)
 
-  # There is an 80-hit/10-second limit to the NHC pages (note Issue #94), or 8
+  # There is an 80-hit/10-second limit to the NHC pages (#94), or 8
   # requests/second. The below request will process 4 links every 0.5 seconds.
   links <- split(links, ceiling(seq_along(links)/4))
+
   # Set progress bar
   p <- dplyr::progress_estimated(n = length(links))
+
   if (getOption("rrricanes.working_msg"))
     message("Gathering storm product links.")
-  res <- purrr::map(links, .f = function(x) {get_url_contents(x, p)}) %>%
-    purrr::flatten()
 
-  # Get contents of all storms
[…]
