+++ man/extract_storm_links.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_storm_data.R
+\name{extract_storm_links}
+\alias{extract_storm_links}
+\title{extract_storm_links}
+\usage{
+extract_storm_links(links, p)
+}
+\arguments{
+\item{links}{URLs to a storm's archive page}
+
+\item{p}{dplyr::progress_estimated object}
+}
+\description{
+Extract product links from a storm's archive page
+}
+\keyword{internal}
