### 2017-07-15 81e57db
Rewrote get_storms, added func get_storms_link
With the removal of progress bar in get_url_contents, get_storms
had to be rewritten to properly utitlize dplyr's progress_estimated
object. Requests now go through get_storms_link which keeps track of
progress and prints the bar.
