### 2017-07-15 04c1b95
Update CHANGELOG, NEWS with previous commit info
