### 2017-07-16 2ca70f5
Cosmetic changes for goodpractice::gp
