+++ R/update.R
@@ -4,20 +4,20 @@
 #' @seealso \code{\link{get_update}}
 #' @keywords internal
 create_df_update <- function() {
-    df <- tibble::data_frame("Status" = character(),
-                             "Name" = character(),
-                             "Date" = as.POSIXct(character(), tz = "UTC"),
-                             "Key" = character(),
-                             "Contents" = character())
+  df <- tibble::data_frame("Status" = character(),
+               "Name" = character(),
+               "Date" = as.POSIXct(character(), tz = "UTC"),
+               "Key" = character(),
+               "Contents" = character())
 
-    return(df)
+  return(df)
 }
 
 #' @title get_update
 #' @description Return dataframe of cyclone update data.
 #' \describe{
 #'   \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
-#'     etc.}
+#'   etc.}
 #'   \item{Name}{Name of storm}
 #'   \item{Date}{Date of advisory issuance}
 #'   \item{Key}{Unique ID of cyclone}
@@ -53,21 +53,21 @@ update <- function(contents) {
   safely_scrape_header <- purrr::safely(scrape_header)
   key <- safely_scrape_header(contents, ret = "key")
   if (is.null(key$error)) {
-    key <- key$result
+  key <- key$result
   } else {
-    key <- NA
+  key <- NA
   }
 
   if (getOption("rrricanes.working_msg"))
-    message(sprintf("Working %s %s Update #%s (%s)",
-                    status, name, date))
+  message(sprintf("Working %s %s Update #%s (%s)",
+          status, name, date))
 
   df <- df %>%
-    tibble::add_row("Status" = status,
-                    "Name" = name,
-                    "Date" = date,
-                    "Key" = key,
-                    "Contents" = contents)
+  tibble::add_row("Status" = status,
+          "Name" = name,
+          "Date" = date,
+          "Key" = key,
+          "Contents" = contents)
 
   return(df)
 }
