+++ R/public.R
@@ -4,23 +4,23 @@
 #' @seealso \code{\link{get_public}}
 #' @keywords internal
 create_df_public <- function() {
-    df <- tibble::data_frame("Status" = character(),
-                             "Name" = character(),
-                             # Allow for intermediate advisories,
-                             # i.e., "1A", "2", "2A"...
-                             "Adv" = character(),
-                             "Date" = as.POSIXct(character(), tz = "UTC"),
-                             "Key" = character(),
-                             "Contents" = character())
+  df <- tibble::data_frame("Status" = character(),
+               "Name" = character(),
+               # Allow for intermediate advisories,
+               # i.e., "1A", "2", "2A"...
+               "Adv" = character(),
+               "Date" = as.POSIXct(character(), tz = "UTC"),
+               "Key" = character(),
+               "Contents" = character())
 
-    return(df)
+  return(df)
 }
 
 #' @title get_public
 #' @description Return dataframe of public advisory data.
 #' \describe{
 #'   \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
-#'     etc.}
+#'   etc.}
 #'   \item{Name}{Name of storm}
 #'   \item{Adv}{Advisory Number}
 #'   \item{Date}{Date of advisory issuance}
@@ -57,22 +57,22 @@ public <- function(contents) {
   safely_scrape_header <- purrr::safely(scrape_header)
   key <- safely_scrape_header(contents, ret = "key")
   if (is.null(key$error)) {
-    key <- key$result
+  key <- key$result
   } else {
-    key <- NA
+  key <- NA
   }
 
   if (getOption("rrricanes.working_msg"))
-    message(sprintf("Working %s %s Public Advisory #%s (%s)",
-                    status, name, adv, date))
+  message(sprintf("Working %s %s Public Advisory #%s (%s)",
+          status, name, adv, date))
 
   df <- df %>%
-    tibble::add_row("Status" = status,
-                    "Name" = name,
-                    "Adv" = adv,
-                    "Date" = date,
-                    "Key" = key,
-                    "Contents" = contents)
+  tibble::add_row("Status" = status,
+          "Name" = name,
+          "Adv" = adv,
+          "Date" = date,
+          "Key" = key,
+          "Contents" = contents)
 
   return(df)
 }
