### 2017-07-16 488373d
Rewrite tests; move get_ func tests to one file
Tests have been corrected to reflect new data gathering processes.
All get func tests have been moved to one file to simplify the data
collection process. This should make it a bit quicker to run.
