+++ tests/testthat/test_posest.R
@@ -1,15 +1,0 @@
-context("Position Estimates (posest)")
-
-load(system.file("extdata", "al092008.posest.Rda", package = "rrricanes"))
-
-## ---- Dataframe Skeleton -----------------------------------------------------
-#' Test structure of dataframe skeleton
-test_that("Dataframe Skeleton", {
-    df <- create_df_posest()
-    expect_true(is.data.frame(df))
-    expect_true(tibble::is_tibble(df))
-    expect_identical(class(df$Status), "character")
-    expect_identical(class(df$Name), "character")
-    expect_identical(class(df$Date), c("POSIXct", "POSIXt"))
-    expect_identical(class(df$Contents), "character")
-})
