+++ tests/testthat/test_discus.R
@@ -1,17 +1,0 @@
-context("Storm Discussions (discus)")
-
-load(system.file("extdata", "al092008.discus.Rda", package = "rrricanes"))
-
-## ---- Dataframe Skeleton -----------------------------------------------------
-#' Test structure of dataframe skeleton
-test_that("Dataframe Skeleton", {
-    df <- create_df_discus()
-    expect_true(is.data.frame(df))
-    expect_true(tibble::is_tibble(df))
-    expect_identical(class(df$Status), "character")
-    expect_identical(class(df$Name), "character")
-    expect_identical(class(df$Adv), "integer")
-    expect_identical(class(df$Date), c("POSIXct", "POSIXt"))
-    expect_identical(class(df$Key), "character")
-    expect_identical(class(df$Contents), "character")
-})
