+++ data-raw/build_test_dataframes.R
@@ -1,68 +1,39 @@
 library(dplyr)
-library(purrr)
 library(rrricanes)
 
-## ---- Base Data --------------------------------------------------------------
-## ---- * 1998, AL -------------------------------------------------------------
-al1998 <- get_storms(year = 1998, basin = "AL") %>% dplyr::select(Link)
+al_1998 <- get_storms(years = 1998, basins = "AL")
+al_2008 <- get_storms(years = 2008, basins = "AL")
+al_2017 <- get_storms(years = 2017, basins = "AL")
 
-## ---- * 2008, AL -------------------------------------------------------------
-al2008 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
-
-## ---- * 1998, AL -------------------------------------------------------------
-al1998 <- get_storms(year = 2008, basin = "AL") %>% dplyr::select(Link)
-
-## ---- get_url_contents -------------------------------------------------------
-df.get_url_contents <- rrricanes:::get_url_contents("http://httpbin.org/delay/0")
-save(df.get_url_contents, file = "./inst/extdata/df.get_url_contents.Rda",
+al_01_2017_products <- get_storm_data(al_2017[[1,4]],
+                                      products = c("discus", "fstadv"))
+save(al_01_2017_products, file = "./inst/extdata/al_01_2017_products.Rdata",
      compression_level = 9)
 
-## ---- get_storm_data ---------------------------------------------------------
-al_2017_storm_data <- get_storms(year = 2017, basin = "AL") %>%
-    dplyr::slice(1) %>%
-    .$Link %>%
-    get_storm_data(products = c("discus", "fstadv"))
-save(al_2017_storm_data, file = "./inst/extdata/al_2017_storm_data.Rda",
+al_09_2008_discus <- get_discus(al_2008[[9,4]])
+save(al_09_2008_discus, file = "./inst/extdata/al_09_2008_discus.Rdata",
      compression_level = 9)
 
-## ---- discus -----------------------------------------------------------------
-## ---- * 2008, AL, 09 ---------------------------------------------------------
-al092008.discus <- al2008 %>% slice(9) %>% .$Link %>% get_discus()
-save(al092008.discus, file = "./inst/extdata/al092008.discus.Rda",
+al_09_2008_fstadv <- get_fstadv(al_2008[[9,4]])
+save(al_09_2008_fstadv, file = "./inst/extdata/al_09_2008_fstadv.Rdata",
      compression_level = 9)
 
-## ---- fstadv -----------------------------------------------------------------
-## ---- * 2008, AL, 09 ---------------------------------------------------------
-al092008.fstadv <- al2008 %>% slice(9) %>% .$Link %>% get_fstadv()
-save(al092008.fstadv, file = "./inst/extdata/al092008.fstadv.Rda",
+al_09_2008_posest <- get_posest(al_2008[[9,4]])
+save(al_09_2008_posest, file = "./inst/extdata/al_09_2008_posest.Rdata",
      compression_level = 9)
 
-## ---- posest -----------------------------------------------------------------
-## ---- * 2008, AL, 09 ---------------------------------------------------------
-al092008.posest <- al2008 %>% slice(9) %>% .$Link %>% get_posest()
-save(al092008.posest, file = "./inst/extdata/al092008.posest.Rda",
+al_01_1998_prblty <- get_prblty(al_1998[[1,4]])
+save(al_01_1998_prblty, file = "./inst/extdata/al_01_1998_prblty.Rdata",
      compression_level = 9)
 
-## ---- prblty -----------------------------------------------------------------
-## ---- * 1998, AL, 01 ---------------------------------------------------------
[…]
