### 2017-07-16 d3d4524
Add base tests for most gis funcs
