+++ vignettes/articles/forecast_advisory.Rmd
@@ -16,6 +16,7 @@ knitr::opts_chunk$set(echo = TRUE,
 library(dplyr)
 library(ggplot2)
 library(rrricanes)
+library(rrricanesdata)
 library(sp)
 ```
 
@@ -25,7 +26,7 @@ adv <- 42
 ```
 
 ```{r}
-fstadv <- load_storm_data("fstadv") %>% filter(Key == key, Adv <= adv)
+fstadv <- fstadv %>% filter(Key == key, Adv <= adv)
 ```
 
 ### GIS Advisory Forecast Track, Cone of Uncertainty, and Watches/Warnings
