+++ vignettes/articles/wind_speed_probabilities.Rmd
@@ -16,6 +16,7 @@ knitr::opts_chunk$set(echo = TRUE,
 library(dplyr)
 library(ggplot2)
 library(rrricanes)
+library(rrricanesdata)
 library(sp)
 ```
 
@@ -25,7 +26,7 @@ adv <- 42
 ```
 
 ```{r}
-wndprb <- load_storm_data("wndprb") %>% filter(Key == key, Adv <= adv)
+wndprb <- wndprb %>% filter(Key == key, Adv <= adv)
 ```
 
 ### GIS Advisory Forecast Track, Cone of Uncertainty, and Watches/Warnings
