+++ vignettes/articles/wind_pressure.Rmd
@@ -41,6 +41,7 @@ library(dplyr)
 library(gganimate)
 library(ggplot2)
 library(rrricanes)
+library(rrricanesdata)
 library(tidyr)
 ```
 
@@ -52,8 +53,7 @@ adv <- 42
 ### Forecast/Advisory Data for Hurricane Ike Adv #42
 
 ```{r}
-# No need to do this piece for active storms
-fstadv <- load_storm_data("fstadv") %>% filter(Key == key, Adv <= adv)
+fstadv <- fstadv %>% filter(Key == key, Adv <= adv)
 ```
 
 ## Wind Profile
