### 2017-07-16 59b2b01
Close hotfix/v0.2.0-6
