### 2017-07-16 d099d01
Add ropensci reviewers
