+++ DESCRIPTION
@@ -7,7 +7,10 @@ Description: Get archived data of past and current hurricanes and tropical
   Currently, this package is about 6MB of datasets. See the README or view 
   `vignette("drat")` for more information.
 Version: 0.2.0-6
-Authors@R: c(person("Tim", "Trice", email = "tim.trice@gmail.com", role = c("aut", "cre")))
+Authors@R: c(person("Tim", "Trice", email = "tim.trice@gmail.com", role = c("aut", "cre")), 
+  person("Maelle", "Salmon", role = c("rev"), comment = "Maëlle Salmon reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/118"), 
+  person("Joseph", "Stachelek", role = c("rev"), comment = "Joseph Stachelek reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/118"), 
+  person("Emily", "Robinson", role = c("rev"), comment = "Emily Robinson reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/118"))
 Depends:
   R (>= 3.3.3)
 URL: https://github.com/timtrice/rrricanes
