+++ .travis.yml
@@ -26,7 +26,7 @@ matrix:
 r_packages:
   - covr
 r_github_packages:
-  - timtrice/rrricanesdata@feature/drat
+  - timtrice/rrricanesdata
 cache: packages
 after_success:
   - Rscript -e 'library(covr); codecov()'
