### 2017-07-16 #118 @timtrice
v0.2.0-6 is now released which completely implements `rrricanesdata` using the drat system. Both packages are passing in Travis and Appveyor.

Is there any word on if I need to submit `rrricanesdata` as a separate package?
