### 2017-07-18 #118 @maelle
@timtrice `rrricanesdata` will be handled in this same thread. :-)

@jsta @robinsones sorry to bother you again, but would you have time to have a quick look at `rrricanesdata`? It's the package providing already downloaded data for `rrricanes`. https://github.com/timtrice/rrricanesdata
