+++ README.Rmd
@@ -119,6 +119,7 @@ This project is licensed under the MIT License - see the [LICENSE.md](LICENSE.md
 
 * [Molyneux, James](https://github.com/jimmylovestea)
 * [Padgham, Mark](https://github.com/mpadge)
+* [Robinson, Emily](https://github.com/robinsones)
 * [Rudis, Bob](https://github.com/hrbrmstr)
 * [Salmon, Maëlle](https://github.com/maelle)
 * [Stachelek, Joseph](https://github.com/jsta)
