### 2017-07-18 #118 @jsta
I took a look. Seems good to me. I did see, however, that the man pages need to be rebuilt from the inline roxygen.
