### 2017-07-18 #118 @timtrice
@jsta, thank you. I corrected the package title and updated the documentation.
