### 2017-07-23 #118 @maelle
Looking at both packages one last time before approving, sorry for the delay.

* In `rrricanesdata` README "complimentary dataset" I think you mean "complementary" 😉

* In that README too, I think the "Getting started" part comes a bit late, I'd suggest moving it before "Datasets".

* When running R CMD Check on `rrricanesdata` I get `checking DESCRIPTION meta-information ... WARNING
Non-standard license specification:
  CC0-1.0 + file LICENSE
Standardizable: FALSE` but if I remember correctly this license was the best one that was advised to you, right?

* It it the first time ever I get this message when running `goodpractice::gp()` on a package (`rrricanesdata` again): `♥ Mmhm! Majestic package! Keep up the mind-blowing work!` 👏

* Does `rrricanesdata` really "require an active internet connection as data is extracted from online sources."?

* Reg. the "rev" role although I had successfully submitted a package to CRAN with this role in the past now CRAN maintainers are having an internal discussion to decide whether they'll accept such roles in the future so please hold on a bit with the CRAN submission, sorry about that. Oh and I see you added me, I didn't review the `rrricanes` package, just did the editor checks, which was less work. 😉

Despite these few remarks the 2 packages are approved! 🎉 Thanks a lot to the reviewers @jsta and @robinsones and to you @timtrice !

To-dos

- [ ] Transfer the repos to the rOpenSci organization under "Settings" in your repo.  I have invited you to 2 teams that should allow you to do so.
- [ ] I'll give you a link soon to add a badge for peer-review in the READMEs.
- [ ] Fix any links in DESCRIPTION, and in badges for CI and coverage to point to the ropensci URL. (we'll turn on the services on our end.)
