+++ DESCRIPTION
@@ -8,7 +8,6 @@ Description: Get archived data of past and current hurricanes and tropical
   `vignette("drat")` for more information.
 Version: 0.2.0-6
 Authors@R: c(person("Tim", "Trice", email = "tim.trice@gmail.com", role = c("aut", "cre")), 
-  person("Maelle", "Salmon", role = c("rev"), comment = "Maëlle Salmon reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/118"), 
   person("Joseph", "Stachelek", role = c("rev"), comment = "Joseph Stachelek reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/118"), 
   person("Emily", "Robinson", role = c("rev"), comment = "Emily Robinson reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/118"))
 Depends:
