### 2017-07-23 #118 @timtrice
@maelle,

Very glad to hear the news~

I have removed you as a reviewer to `rrricanes`.

I have also made some modifications to `rrricanesdata` including removing**`Getting Started**, rewriting **Installation**, and relocating **Data**.

I will wait to push that to master until you provide the README links.

Once I push to master I'll transfer both repos.

Regarding the license, I never really got a consensus on this. The discussion, if I recall correctly, centered more around if it mattered as it was not a CRAN package. But, there wasn't a definitive, "This is/is not the correct license for this type of package".

Regarding `rrricanes` to CRAN; does it have to be submitted at any point soon? I feel the GIS documentation can be improved and it is next on my list before I push a minor release. I need to document this procedure better (one of my weaknesses) and I'm not sure if the way I'm handling shapefiles is the "proper" way.
