### 2017-07-23 #118 @maelle
@timtrice nice work!

* I hope to provide the badge link in the next days. There was some issues with the existing badges (I experienced them for a package of mine actually) so there's work going on.

* No you don't need to submit to CRAN any point soon. You can get some feedback about this documentation/code via rOpenSci forum/Slack, where some people are experienced with shapefiles.
