### 2017-07-23 #118 @maelle
Oh good question! @sckott I have no drat experience, could you please answer the question above (and @timtrice can provide more info about the package he distributes via drat)? Thanks 😸
