### 2017-07-25 6a6c08f
Add rOpenSci badge
