### 2017-07-25 f83b778
Change repo links to ropensci
