### 2017-07-25 a4c3598
Updated badges to https:
