### 2017-07-25 #118 @timtrice
@maelle ,

For the most part I'm done. I need to update the pkgdown docs for `rrricanes` but I have to do that on another machine; will do so tonight.

But, the URL in the description for each repo, I can't change. I guess I'm not an admin for my repo anymore? It still points to the old repo address for the pkgdown documentation which I need to update.
