### 2017-07-26 0405fdf
Correct Appveyor badge link
