### 2017-07-26 #118 @sckott
Carl and I discussed about drat - we have our own as you may know, and so once these two pkgs are in ropensci org, they will be included in our drat - so i'd say not move the drat repo
