### 2017-07-27 #118 @maelle
@timtrice I forgot to ask whether you'd like to write a a blog post about your package for rOpenSci blog, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). Let me know if you are interested.
