### 2017-08-03 #118 @maelle
@timtrice if you submit the package to CRAN before R3.5, please build it on R-devel in order not to get any   note because of the "rev" role. :-)
