### 2017-08-19 #118 @timtrice
@maelle ,

I've added [introducing_rrricanes](https://gist.github.com/timtrice/ae78fe3a5b9cc5805c3dc004e0a14fe9) gist for review per your previous comment. If I'm on the wrong track with the purpose of this, please let me know. Or if you have additional suggestions, of course, I'm all for it.

Thank you for your help!
