### 2017-08-20 #118 @maelle
Nice to see this text being written! Some comments:

* I think it'd be nice to start with why you created this package, to give it a personal touch before the technical stuff? Well if that's not intended as technical note of course.

* streamling -> streaming, it's own "get" function -> its own "get" function

* Maybe add even more examples of visualizations in the post? And maybe comment on the existing one, what does it show and mean? Could you include the code used for generating it?

* In "Optional products are Strike Probabilities (since deprecated and replaced by Wind Speed Probabilities), Position Estimates (since rolled into Tropical Updates)." what does "since" refer to?

* When saying contributions are welcome, maybe give more details about what sort of contributions you expect? E.g. bug reports, feature requests, or PRs. When I wrote about my own package `ropenaq` Stefanie encouraged me to open issues labelled "Beginner" and "Intermediate", see [here](https://github.com/ropensci/ropenaq/issues) (yep no one volunteered but don't be disheartened by this!)
