### 2017-08-23 #118 @stefaniebutland
@timtrice Great to hear you're interested in contributing a blog post about `rrricanes` and thanks for starting out with the [introducing_rrricanes](https://gist.github.com/timtrice/ae78fe3a5b9cc5805c3dc004e0a14fe9) gist. Here are some practical details. I'll give some feedback on the draft in a separate comment below.

As Maëlle mentioned, you could do either a short post for the [Developer Blog](https://ropensci.org/tech-notes/) that has a technical focus, or do a post for the [main blog](https://ropensci.org/blog/) whose audience is broader.

Main blog post would include more narrative about your motivation for creating the package, unmet need, how-to use, good to end with a thank you to package reviewers with links to their GitHub or Twitter, point readers to issues and what you think is next to improve the package and invite people to open or address an issue etc.

Deadlines:
- Developer blog post can come any time.
- For main blog we post about once a week. Next open slots are late September. We ask for a draft submitted as a pull request a week in advance.

Practical instructions:
- https://github.com/ropensci/roweb/blob/master/.github/CONTRIBUTING.md
- suggested tags: R, community, software, review, onboarding, package, package_name, topic labels used in onboarding review

Which type of post are you thinking of?
Did I miss anything?
