### 2017-08-23 #118 @stefaniebutland
Some comments on [introducing_rrricanes](https://gist.github.com/timtrice/ae78fe3a5b9cc5805c3dc004e0a14fe9) to add to Maëlle's:

- "archived" in "...current and archive tropical cyclone data..."
- If for the main blog, definitely good to add a short narrative intro about your motivation etc as noted in my comment above.
- maybe create another heading for the section on installation so it doesn't fall under "What is rrricanes". Here's a [good main blog post example](https://ropensci.org/blog/blog/2017/04/18/tabulizer) where sections are well laid out with headings
- echoing @maelle's comment "Maybe add even more examples of visualizations in the post? And maybe comment on the existing one, what does it show and mean? Could you include the code used for generating it?" This would add a lot of value and help with the narrative so people can see for themselves what they have to gain by trying out the package.

I hope those comments help. Thank you again for doing this. Please tag me here or ping me on slack with any questions or to look at next version.
