### 2017-09-02 #118 @timtrice
@maelle , @stefaniebutland ,

I've added additional visuals to the next version and moved the document to it's own repo:

https://github.com/timtrice/introducing_rrricanes

This may be a bit longer than what was intended but I really wanted to cover the big ideas behind the package and try to explain them under the assumption that anyone using it would be unfamiliar with what the different data meant. I hope that I've accomplished this (even if a bit too thoroughly).

Look forward to any input you both may have to offer.
