### 2017-09-06 #118 @stefaniebutland
Hi @timtrice
This looks ready to set a date as an rOpenSci blog post. Please send a pull request according to instructions here: https://github.com/ropensci/roweb/blob/master/.github/CONTRIBUTING.md and we will do final review on the pull request.

- I propose posting on Tues Sep 19, so please submit pull request by Tues Sep 12.
- For tags (see CONTRIBUTING.md above for format), I suggest: R, onboarding, community (since this pkg is a community contribution not written by rOpenSci staff), package, software, review, package_name; package-topics (copied from onboarding review topic labels)
- just say this in intro and people will understand, and be prepared
> This may be a bit longer than what was intended but I really wanted to cover the big ideas behind the package and try to explain them under the assumption that anyone using it would be unfamiliar with what the different data meant.
- Perhaps give a brief outline of what people will find in the post or outline examples you will show so that people are motivated to read through
- define NHC archives first time it's noted (I think you did in second mention)
- Here `gis_advisory(key = key)` could you just list a few of the datasets to save space/reader attention since full list is no more informative? Same for `gis <- gis_latest()` and maybe you can identify others that could be shortened.
- love this: "You do not need to submit code in order to be listed as a contributor. ..."
