### 2017-09-11 #118 @timtrice
Hello @stefaniebutland , @maelle ,

I wanted to give you guys an update on this. I finished the post last week and was ready to submit to roweb. However, currently the NHC website is down for GIS products making it impossible to render the Rmd file.

They are aware of the issue and, I've been told are trying to correct it. But I have no idea when that will be. Basically, I'm on hold for now.
