### 2017-09-11 #118 @stefaniebutland
Thanks @timtrice for the update! While your post is scheduled to go up Sep 19, I can easily move that date. Just ping me when you have submitted to roweb.
