### 2017-09-14 #118 @timtrice
@stefaniebutland @maelle

I was finally able to render the Rmd file and submit the [pull request](https://github.com/ropensci/roweb/pull/348) as referenced above.
